TST: add montgomery schema tests (#13)
