+++ tests/fixtures/montgomery_test_fixtures.py
@@ -1,3 +1,58 @@
+parse_event_item_expected = {'Event Start Date': '2019-01-30', 
+'Event End Date': '2019-01-30',
+'Event Start Time': '10:30:00', 'Event End Time': '11:30:00', 'Event Website': 'https://www.montgomeryparks.org/events/nature-rx-forest-therapy-walks-1-hour-/', 'Event Name': 'Nature Rx: Forest Therapy Walks (1 hour)', 'Event Venue Name': 'Brookside Nature Center', 'Event Cost': '6', 'Event Description': 'Experience the healing and wellness promoting effects of Shinrin-Yoku, the practice of bathing the senses in the atmosphere of the forest. Take a slow and mindful walk with a Forest Therapy guide on a trail at Brookside Nature Center to awaken your senses and reconnect with nature.', 'Event Category': 'category', 'Timezone': 'America/New_York', 'Event Organizers': 'Brookside Nature Center', 'Event Currency Symbol': '$',
+'All Day Event':False}
+
+category_id_map_expected = {'Archaeology': '1751',
+                    'Butterfly Show': '2230',
+                    'Camp': '2701',
+                    'Clean Up': '2235',
+                    'Community Events': '1738',
+                    'Community Gardens': '1739',
+                    'Community Meeting': '2577',
+                    'Earth Month': '3105',
+                    'Education': '2239',
+                    'Event Center': '3331',
+                    'Events': '1740',
+                    'Family': '1741',
+                    'Featured': '2328',
+                    'Festival': '2232',
+                    'Free': '1746',
+                    'Gardens': '2243',
+                    'Golf': '2242',
+                    'Hikes': '1733',
+                    'Historical/Cultural': '1743',
+                    'History': '2240',
+                    'History in the Parks': '3215',
+                    'History Program': '2234',
+                    'Hockey': '1727',
+                    'Holiday': '2241',
+                    'Ice Skating': '1728',
+                    'Kids': '1729',
+                    'MLK Day of Service': '2705',
+                    'Music': '1744',
+                    'Nature': '82',
+                    'open house': '2901',
+                    'Openings/Dedications': '1753',
+                    'Other': '2233',
+                    'Park Planning': '2902',
+                    'Park Projects': '2916',
+                    'Planning Board Meeting': '2751',
+                    'Planning Board Worksession': '2756',
+                    'pop up': '2569',
+                    'Public Hearing': '2754',
+                    'Public Meetings': '1750',
+                    'Run': '2231',
+                    'Special Events': '1730',
+                    'Sports': '1731',
+                    'Tennis': '1734',
+                    'Trail Planning': '2903',
+                    'Trail Work': '2706',
+                    'Trails': '1735',
+                    'Trips': '1736',
+                    'Volunteering': '1747',
+                    'Weed Warrior': '1748'}
+
 calendar_page_content = b'<!DOCTYPE html>\r\n<html lang="en-US" xmlns:fb="https://www.facebook.com/2008/fbml" xmlns:addthis="https://www.addthis.com/help/api-spec"  prefix="og: http://ogp.me/ns#" class="no-js">\r\n\t<head>\r\n\t\t<meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">\r\n\t\t<script src="https://use.typekit.net/bxi2fgb.js"></script>\r\n\t\t<script>try{Typekit.load({ async: true });}catch(e){}</script>\r\n\t\t<meta charset="UTF-8">\r\n\t\t<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=12.0, minimum-scale=.25, user-scalable=yes"/>\r\n\t\t<!-- CHECK META CONTENT -->\r\n\t\t<title>Calendar - Montgomery Parks</title>\r\n\r\n        <script>\r\n          // helpful for understanding dojoConfig.packages vs. dojoConfig.paths:\r\n          // http://www.sitepen.com/blog/2013/06/20/dojo-faq-what-is-the-difference-packages-vs-paths-vs-aliases/\r\n          var dojoConfig = {\r\n            paths: {\r\n              extras: "/wp-content/themes/mdparks/dist/js/extras"\r\n            }\r\n          };\r\n        </script>\r\n\r\n\t\t<link rel="pingback" href="/xmlrpc.php">\r\n\t\t\t\t<style>#skipNav{display: block;background-color:#fff;padding:1em;position: absolute;top: 0;left: -9999px;}#skipNav:focus{left:0;}</style>\r\n\r\n\t\t\n<!-- This site is optimized with the Yoast SEO plugin v9.5 - https://yoast.com/wordpress/plugins/seo/ -->\n<link rel="canonical" href="/calendar/" />\n<meta property="og:locale" content="en_US" />\n<meta property="og:type" content="article" />\n<meta property="og:title" content="Calendar - Montgomery Parks" />\n<meta property="og:url" content="/calendar/" />\n<meta property="og:site_name" content="Montgomery Parks" />\n<meta name="twitter:card" content="summary" />\n<meta name="twitter:title" content="Calendar - Montgomery Parks" />\n<!-- / Yoast SEO plugin. -->\n\n<link rel=\'dns-prefetch\' href=\'//js.arcgis.com\' />\n<link rel=\'dns-prefetch\' href=\'//s.w.org\' />\n<link rel=\'dns-prefetch\' href=\'//s3.amazonaws.com\' />\n<link rel="alternate" type="application/rss+xml" title="Montgomery Parks &raquo; Feed" href="/feed/" />\n<link rel="alternate" type="application/rss+xml" title="Montgomery Parks &raquo; Comments Feed" href="/comments/feed/" />\n<!-- This site uses the Google Analytics by MonsterInsights plugin v7.4.1 - Using Analytics tracking - https://www.monsterinsights.com/ -->\r\n<script type="text/javascript" data-cfasync="false">\r\n\tvar mi_version         = \'7.4.1\';\r\n\tvar mi_track_user      = true;\r\n\tvar mi_no_track_reason = \'\';\r\n\t\r\n\tvar disableStr = \'ga-disable-UA-493554-2\';\r\n\r\n\t/* Function to detect opted out users */\r\n\tfunction __gaTrackerIsOptedOut() {\r\n\t\treturn document.cookie.indexOf(disableStr + \'=true\') > -1;\r\n\t}\r\n\r\n\t/* Disable tracking if the opt-out cookie exists. */\r\n\tif ( __gaTrackerIsOptedOut() ) {\r\n\t\twindow[disableStr] = true;\r\n\t}\r\n\r\n\t/* Opt-out function */\r\n\tfunction __gaTrackerOptout() {\r\n\t  document.cookie = disableStr + \'=true; expires=Thu, 31 Dec 2099 23:59:59 UTC; path=/\';\r\n\t  window[disableStr] = true;\r\n\t}\r\n\t\r\n\tif ( mi_track_user ) {\r\n\t\t(function(i,s,o,g,r,a,m){i[\'GoogleAnalyticsObject\']=r;i[r]=i[r]||function(){\r\n\t\t\t(i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),\r\n\t\t\tm=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)\r\n\t\t})(window,document,\'script\',\'//www.google-analytics.com/analytics.js\',\'__gaTracker\');\r\n\r\n\t\t__gaTracker(\'create\', \'UA-493554-2\', \'auto\');\n\t\t__gaTracker(\'set\', \'forceSSL\', true);\n\t\t__gaTracker(\'set\', \'anonymizeIp\', true);\n\t\t__gaTracker(\'require\', \'linkid\', \'linkid.js\');\n\t\t__gaTracker(\'send\',\'pageview\');\n\t} else {\r\n\t\tconsole.log( "" );\r\n\t\t(function() {\r\n\t\t\t/* https://developers.google.com/analytics/devguides/collection/analyticsjs/ */\r\n\t\t\tvar noopfn = function() {\r\n\t\t\t\treturn null;\r\n\t\t\t};\r\n\t\t\tvar noopnullfn = function() {\r\n\t\t\t\treturn null;\r\n\t\t\t};\r\n\t\t\tvar Tracker = function() {\r\n\t\t\t\treturn null;\r\n\t\t\t};\r\n\t\t\tvar p = Tracker.prototype;\r\n\t\t\tp.get = noopfn;\r\n\t\t\tp.set = noopfn;\r\n\t\t\tp.send = noopfn;\r\n\t\t\tvar __gaTracker = function() {\r\n\t\t\t\tvar len = arguments.length;\r\n\t\t\t\tif ( len === 0 ) {\r\n\t\t\t\t\treturn;\r\n\t\t\t\t}\r\n\t\t\t\tvar f = arguments[len-1];\r\n\t\t\t\tif ( typeof f !== \'object\' || f === null || typeof f.hitCallback !== \'function\' ) {\r\n\t\t\t\t\tconsole.log( \'Not running function __gaTracker(\' + arguments[0] + " ....) because you are not being tracked. " + mi_no_track_reason );\r\n\t\t\t\t\treturn;\r\n\t\t\t\t}\r\n\t\t\t\ttry {\r\n\t\t\t\t\tf.hitCallback();\r\n\t\t\t\t} catch (ex) {\r\n\r\n\t\t\t\t}\r\n\t\t\t};\r\n\t\t\t__gaTracker.create = function() {\r\n\t\t\t\treturn new Tracker();\r\n\t\t\t};\r\n\t\t\t__gaTracker.getByName = noopnullfn;\r\n\t\t\t__gaTracker.getAll = function() {\r\n\t\t\t\treturn [];\r\n\t\t\t};\r\n\t\t\t__gaTracker.remove = noopfn;\r\n\t\t\twindow[\'__gaTracker\'] = __gaTracker;\r\n\t\t\t\t\t})();\r\n\t\t}\r\n</script>\r\n<!-- / Google Analytics by MonsterInsights -->\r\n<link rel=\'stylesheet\' id=\'wp-block-library-group-css\' href=\'/wp-content/plugins/bwp-minify/min/?f=wp-includes/css/dist/block-library/style.min.css,wp-content/themes/mdparks/dist/css/main.min.css,wp-content/plugins/addthis/frontend/build/addthis_wordpress_public.min.css\' type=\'text/css\' media=\'all\' />\r\n<link rel=\'stylesheet\' id=\'brink_gdpr_css-group-css\' href=\'/wp-content/plugins/bwp-minify/min/?f=wp-content/plugins/brink-gdpr/css/brink-gdpr.css\' type=\'text/css\' media=\'screen\' />\r\n<link rel=\'stylesheet\' id=\'esri-css-css\'  href=\'https://js.arcgis.com/3.18/esri/css/esri.css?ver=5.0.3\' type=\'text/css\' media=\'all\' />\n<script type=\'text/javascript\'>\n/* <![CDATA[ */\nvar monsterinsights_frontend = {"js_events_tracking":"true","download_extensions":"doc,exe,js,pdf,ppt,tgz,zip,xls","inbound_paths":"[]","home_url":"http:\\/\\/staging.montgomeryparks.org","hash_tracking":"false"};\n/* ]]> */\n</script>\n<script type=\'text/javascript\' src=\'/wp-content/plugins/google-analytics-for-wordpress/assets/js/frontend.min.js?ver=7.4.1\'></script>\n<script type=\'text/javascript\' src=\'/wp-content/plugins/brink-gdpr/js/brink_gdpr.js?ver=5.0.3\'></script>\n<link rel=\'https://api.w.org/\' href=\'/wp-json/\' />\n<link rel="EditURI" type="application/rsd+xml" title="RSD" href="/xmlrpc.php?rsd" />\n<link rel="wlwmanifest" type="application/wlwmanifest+xml" href="/wp-includes/wlwmanifest.xml" /> \n<meta name="generator" content="WordPress 5.0.3" />\n<link rel=\'shortlink\' href=\'/?p=32\' />\n<link rel="alternate" type="application/json+oembed" href="/wp-json/oembed/1.0/embed?url=http%3A%2F%2Fstaging.montgomeryparks.org%2Fcalendar%2F" />\n<link rel="alternate" type="text/xml+oembed" href="/wp-json/oembed/1.0/embed?url=http%3A%2F%2Fstaging.montgomeryparks.org%2Fcalendar%2F&#038;format=xml" />\n<script type="text/javascript">\nvar ajaxurl = \'/wp-admin/admin-ajax.php/\';\n</script>\n\t\t<style type="text/css">.recentcomments a{display:inline !important;padding:0 !important;margin:0 !important;}</style>\n\t\t<link rel="icon" href="/uploads/2016/07/cropped-favicon-490x490.png" sizes="32x32" />\n<link rel="icon" href="/uploads/2016/07/cropped-favicon-490x490.png" sizes="192x192" />\n<link rel="apple-touch-icon-precomposed" href="/uploads/2016/07/cropped-favicon-490x490.png" />\n<meta name="msapplication-TileImage" content="/uploads/2016/07/cropped-favicon-490x490.png" />\n\r\n\t\t\t</head>\r\n\r\n\t<body class="page-template page-template-page-calendar page-template-page-calendar-php page page-id-32 group-blog">\r\n\r\n\r\n\r\n\t\t<div id="page" class="page">\r\n\t\t\t<div class="main-header parks-header">\r\n\t\t    <div class="mncppc">\r\n\t\t        <a href="http://www.mncppc.org" title="Maryland-National Capital Parks and Commission" class="mncppc-link" target="_blank">Maryland-National Capital Park and Planning Commission</a>\r\n\t\t        <ul class="sites-links">\r\n\t\t\t\t\t<li class="parks current"><a href="/parks-home">Parks</a></li>\r\n\t\t            <li class="planning"><a href="http://montgomeryplanning.org" target="_blank">Planning</a></li>\r\n\t\t            <li class="board"><a href="http://montgomeryplanningboard.org" target="_blank">Planning Board</a></li>\r\n\t\t        </ul>\r\n\t\t    </div>\r\n\t\t    <div class="branding">\r\n\t\t    \t<div class="col col-tab-9 col-sm-desk-6 col-lg-5 left-wrap p-x">\r\n\t\t    \t\t<div class="mid-v">\r\n\t\t\t\t        <div class="site-icon">\r\n\t\t\t\t        \t<a href="/" class="parks-logo t-indent" title="Back to Home">Montgomery Parks</a>\r\n\t\t\t\t        </div>\r\n\t\t\t\t        <div class="site-title">\r\n\t\t\t\t        \t<a href="/" class="parks-logo" title="Back to Home">\r\n\t\t\t\t        \t<div class="title-wrap">\r\n\t\t\t\t\t            <p class="montco">Montgomery County, MD</p>\r\n\t\t\t\t\t            <h2 class="name">Montgomery Parks</h2>\r\n\t\t\t\t            </div>\r\n\t\t\t\t            </a>\r\n\t\t\t\t        </div>\r\n\t\t\t        </div>\r\n\t\t        </div>\r\n\t\t\t\t<div class="col col-tab-3 col-desk-2 viewing-options">\r\n\t\t\t\t<!-- \t<div class="head-col-wrap"> -->\r\n\t\t\t\t\t\t<div class="options-wrap">\r\n\t\t\t\t\t\t\t<div id="google_translate_element" tabindex="-1"></div>\r\n\t\t\t\t\t\t\t<script type="text/javascript">\r\n                            function fireEvent(element,event){\r\n                                if (document.createEventObject){\r\n                                    // dispatch for IE\r\n                                    var evt = document.createEventObject();\r\n                                    return element.fireEvent(\'on\'+event,evt)\r\n                                }\r\n                                else{\r\n                                    // dispatch for firefox + others\r\n                                    var evt = document.createEvent("HTMLEvents");\r\n                                    evt.initEvent(event, true, true ); // event type,bubbling,cancelable\r\n                                    return !element.dispatchEvent(evt);\r\n                                }\r\n                             }\r\n\r\n                            function googleTranslateElementInit() {\r\n                                new google.translate.TranslateElement({pageLanguage: \'en\', includedLanguages: \'af,ach,ak,am,ar,az,be,bem,bg,bh,bn,br,bs,ca,chr,ckb,co,crs,cs,cy,da,de,ee,el,en,eo,es,es-419,et,eu,fa,fi,fo,fr,fy,ga,gaa,gd,gl,gn,gu,ha,haw,hi,hr,ht,hu,hy,ia,id,ig,is,it,iw,ja,jw,ka,kg,kk,km,kn,ko,kri,ku,ky,la,lg,ln,lo,loz,lt,lua,lv,mfe,mg,mi,mk,ml,mn,mo,mr,ms,mt,ne,nl,nn,no,nso,ny,nyn,oc,om,or,pa,pcm,pl,ps,pt-BR,pt-PT,qu,rm,rn,ro,ru,rw,sd,sh,si,sk,sl,sn,so,sq,sr,sr-ME,st,su,sv,sw,ta,te,tg,th,ti,tk,tl,tn,to,tr,tt,tum,tw,ug,uk,ur,uz,vi,wo,xh,xx-bork,xx-elmer,xx-hacker,xx-klingon,xx-pirate,yi,yo,zh-CN,zh-TW,zu\', layout: google.translate.TranslateElement.InlineLayout}, \'google_translate_element\');\r\n                                setTimeout(function () {\r\n                                    $(\'select.goog-te-combo\').selectize({\r\n                                        onChange: function (val) {\r\n                                            $(\'select.goog-te-combo\').val(val);\r\n                                            fireEvent($(\'select.goog-te-combo\').get(0), \'change\');\r\n                                        }\r\n                                    });\r\n                                    // $(\'#google_translate_element .selectize-input > input\').attr(\'id\', \'google_translate_element\');\r\n                                    // $(\'#google_translate_element .selectize-input > input\').wrap("<label for=\'google_translate_element\'></label>");\r\n                                }, 1000);\r\n                            }\r\n\t\t\t\t\t\t\t</script>\r\n\t\t\t\t\t\t\t<script type="text/javascript" src="//translate.google.com/translate_a/element.js?cb=googleTranslateElementInit"></script>\r\n\t\t\t\t\t\t\t<span class=\'divide-line\'>|</span> <div class="open-wrap"><a class="modal-open" href="#">Text Size</a></div>\r\n\t\t\t\t\t\t</div>\r\n<!-- \t\t\t\t\t</div>\r\n -->\t\t\t\t</div>\r\n\t\t\t\t<div class="mob-hide col-desk-2 col-lg-3 rel">\r\n\t\t\t\t\t<div class="neat-search">\r\n                        <form action="/">\r\n                        \t<label for="search_2">Search the Site</label>\r\n    \t\t\t\t\t\t<input id="search_2" class="closed" type="text" name="s" placeholder="Search" />\r\n    \t\t\t\t\t\t<button type="submit" class="search-link t-indent">Search</button>\r\n                        </form>\r\n\t\t\t\t\t</div>\r\n\t\t\t\t</div>\r\n\t\t\t\t<div class="social col col-sm-desk-3 col-desk-2 rel mob-hide">\r\n\t\t\t\t\t<div class="share">\n\t<a href="https://www.facebook.com/MontgomeryParks" target="_blank">facebook<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1fb{fill:#547bbc}.cls-2{fill:#fff}</style></defs><title>social-fb</title><path class="cls-1fb" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M11.4 12.5H13V11a4 4 0 0 1 .5-2.3 2.8 2.8 0 0 1 2.4-1.1 9.7 9.7 0 0 1 2.8.3l-.4 2.3-1.3-.3a.9.9 0 0 0-1.1.8v1.8h2.5l-.2 2.2h-2.4v7.8h-2.9v-7.8h-1.5v-2.2z"/></svg></a>\n\t<a href="https://twitter.com/MontgomeryParks" target="_blank">twitter<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1tw{fill:#5ea8dd}.cls-2{fill:#fff}</style></defs><title>social-tw</title><path class="cls-1tw" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M22.5 11.1l-1.8.5a3.1 3.1 0 0 0 1.4-1.7 6.2 6.2 0 0 1-2 .7 3.1 3.1 0 0 0-5.3 2.8 8.8 8.8 0 0 1-6.4-3.2 3.1 3.1 0 0 0 1 4.1 3.1 3.1 0 0 1-1.4-.4 3.1 3.1 0 0 0 2.5 3H9.2a3.1 3.1 0 0 0 2.9 2.1 6.2 6.2 0 0 1-3.8 1.3h-.8A8.8 8.8 0 0 0 21 13.1v-.4a6.2 6.2 0 0 0 1.5-1.6"/></svg></a>\n\t<a href="https://www.instagram.com/montgomeryparks/" target="_blank">instagram<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 291.319 291.319"><path d="M145.659 0c80.44 0 145.66 65.219 145.66 145.66S226.1 291.319 145.66 291.319 0 226.1 0 145.66 65.21 0 145.659 0z" fill="#9b6954"/><path d="M195.93 63.708H95.38c-17.47 0-31.672 14.211-31.672 31.672v100.56c0 17.47 14.211 31.672 31.672 31.672h100.56c17.47 0 31.672-14.211 31.672-31.672V95.38c-.001-17.461-14.212-31.672-31.682-31.672zm9.978 18.326l3.587-.009v27.202l-27.402.091-.091-27.202 23.906-.082zm-60.248 36.205c22.732 0 27.42 21.339 27.42 27.429 0 15.103-12.308 27.411-27.42 27.411-15.121 0-27.42-12.308-27.42-27.411-.01-6.09 4.688-27.429 27.42-27.429zm63.99 75.716c0 8.658-7.037 15.704-15.713 15.704H97.073c-8.667 0-15.713-7.037-15.713-15.704v-66.539h22.759c-2.112 5.198-3.305 12.299-3.305 18.253 0 24.708 20.101 44.818 44.818 44.818s44.808-20.11 44.808-44.818c0-5.954-1.193-13.055-3.296-18.253h22.486v66.539h.02z" fill="#fff"/></svg></a>\n\t<a href="https://www.youtube.com/user/MontgomeryParks" target="_blank">youtube<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1yt{fill:#e9644a}.cls-2{fill:#fff}</style></defs><title>social-yt</title><path class="cls-1yt" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M17.9 15L13 17.9v-5.8zm4.6 2.7v-5.4a2.5 2.5 0 0 0-2.6-2.6h-9.8a2.5 2.5 0 0 0-2.6 2.6v5.4a2.5 2.5 0 0 0 2.6 2.6h9.8a2.5 2.5 0 0 0 2.6-2.6"/></svg></a>\n\t<a href="http://www.montgomeryplanningboard.org/rss/rss.shtm" target="_blank">rss<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1rss{fill:#f29100}.cls-2{fill:#fff}</style></defs><title>social-rss</title><path class="cls-1rss" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M10.9 19.1a2 2 0 0 1 0 2.8 2 2 0 1 1-2.8-2.8 2 2 0 0 1 2.8 0zm-3.4-6.5v2.9a7 7 0 0 1 7 7h2.9a9.9 9.9 0 0 0-9.9-9.9zm0-5.1v2.9a12.2 12.2 0 0 1 12.1 12.1h2.9a15 15 0 0 0-15-15z"/></svg></a>\n</div>\t\t\t\t</div>\r\n\r\n\t\t    </div>\r\n\t\t    <div class="nav-primary count-7">\r\n\t\t    \t<div class="icon-wrap rel">\r\n\t\t\t        <div class="ico-link">\r\n\t\t\t        \t<a href="/" class="ico ico-home" title="Back to Home">Back to Home<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 31.9 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-home</title><path class="cls-1" d="M26.7 14.7a.9.9 0 0 0-.9.9v12.6h-6.5v-5.1a.9.9 0 0 0-.9-.9h-4.7a.9.9 0 0 0-.9.9v5.1H6.2V16a.9.9 0 0 0-1.8 0v13.1a.9.9 0 0 0 .9.9h8.4a.9.9 0 0 0 .9-.9V24h2.9v5.1a.9.9 0 0 0 .9.9h8.3a.9.9 0 0 0 .9-.9V15.6a.9.9 0 0 0-.9-.9zm0 0l5 .7-7.3-7.3v-5a.9.9 0 0 0-.9-.9.9.9 0 0 0-.9.9v3.2l-6-6A.9.9 0 0 0 16 0a.9.9 0 0 0-.7.3l-15 15a.9.9 0 0 0 0 1.3.9.9 0 0 0 1.3 0L16 2.2l14.4 14.5a.9.9 0 0 0 .7.2l.7-.2a1 1 0 0 0 0-1.3zm0 0"/></svg></a>\r\n\t\t\t        </div>\r\n\t\t\t        <div class="ico-link">\r\n\t\t\t        \t<a href="/calendar/?v=1" class="ico ico-calendar" title="Calendar of Events">Calendar of Events<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-calendar</title><path class="cls-1" d="M27.5 2.8h-4.1V.9a.9.9 0 0 0-1.9 0v1.9h-5.6V.9a.9.9 0 0 0-1.9 0v1.9H8.4V.9a.9.9 0 1 0-1.8 0v1.9H2.5A2.5 2.5 0 0 0 0 5.3v22.2A2.5 2.5 0 0 0 2.5 30h25a2.5 2.5 0 0 0 2.5-2.5V5.3a2.5 2.5 0 0 0-2.5-2.5zm.6 24.7a.6.6 0 0 1-.6.6h-25a.6.6 0 0 1-.6-.6V5.3a.6.6 0 0 1 .6-.6h4.1v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h4.1a.6.6 0 0 1 .6.6v22.2zm0 0"/><path class="cls-1" d="M6.6 11.3h3.7v2.81H6.6zm0 4.6h3.7v2.81H6.6zm0 4.7h3.7v2.81H6.6zm6.5 0h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7zm6.6 9.3h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7z"/></svg></a>\r\n\t\t\t        </div>\r\n\t\t\t        <div class="ico-link">\r\n\t\t\t        \t<a href="/news/" class="ico ico-news" title="Latest News">Latest News<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30.7 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-news</title><path class="cls-1" d="M30 2.5a.8.8 0 0 0-.8.8v21.1a4 4 0 0 1-4 4H5.7a4 4 0 0 1-4-4V1.6h23.2v20.5a.8.8 0 1 0 1.6 0V.8a.8.8 0 0 0-.8-.8H.8a.8.8 0 0 0-.8.8v23.6A5.6 5.6 0 0 0 5.7 30h19.4a5.6 5.6 0 0 0 5.6-5.6V3.3a.8.8 0 0 0-.8-.8zm0 0"/><path class="cls-1" d="M21.9 6.2a.8.8 0 0 0 0-1.6H4.5a.8.8 0 1 0 0 1.6h17.4zm0 14.6H4.5a.8.8 0 1 0 0 1.6h17.4a.8.8 0 0 0 0-1.6zm0 4.1H7.3a.8.8 0 0 0 0 1.6h14.6a.8.8 0 0 0 0-1.6zM3.5 9v8.6a.8.8 0 0 0 .8.8h8.6a.8.8 0 0 0 .8-.8V9a.8.8 0 0 0-.8-.8H4.3a.8.8 0 0 0-.8.8zm1.6.8h7v7h-7v-7zM22.3 9h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 0-1.6zm0 3.6h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 1 0 0-1.6zm.8 4.6a.8.8 0 0 0-.8-.8h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 .8-.8zm0 0"/></svg></a>\r\n\t\t\t        </div>\r\n\t\t        </div>\r\n\t\t        <div class="menu-primary-container"><ul id="menu-primary" class="menu"><li id="menu-item-2177" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-2177"><a href="/activities/">Things To Do</a>\n<ul class="sub-menu">\n\t<li id="menu-item-664" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-664"><a href="/activities/">Activities</a></li>\n\t<li id="menu-item-8304" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-8304"><a href="/classes-camps/">Classes &#038; Camps</a></li>\n\t<li id="menu-item-950" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-950"><a href="/special-events/">Special Events &#038; Festivals</a></li>\n</ul>\n</li>\n<li id="menu-item-853" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-853"><a href="/parks-directory/">Parks &#038; Trails Directory</a>\n<ul class="sub-menu">\n\t<li id="menu-item-1089" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1089"><a href="/parks-trails/parks/">Parks</a></li>\n\t<li id="menu-item-39673" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-39673"><a href="https://www.montgomeryparks.org/activities/park-trails/">Trails</a></li>\n\t<li id="menu-item-1092" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1092"><a href="/parks-trails/gardens/">Gardens</a></li>\n\t<li id="menu-item-2941" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2941"><a href="/parks-trails/facilities/">Facilities</a></li>\n\t<li id="menu-item-1090" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1090"><a href="/parks-trails/nature-centers/">Nature Centers</a></li>\n\t<li id="menu-item-2360" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2360"><a href="/parks-trails/event-centers/">Event Centers</a></li>\n</ul>\n</li>\n<li id="menu-item-837" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-837"><a href="/services/">Permits, Rentals &#038; Requests</a>\n<ul class="sub-menu">\n\t<li id="menu-item-2850" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2850"><a href="/services/permits-rentals/">Permits &#038; Rentals</a></li>\n\t<li id="menu-item-838" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-838"><a href="/services/report-a-problem/">Parks Information &#038; Customer Service</a></li>\n</ul>\n</li>\n<li id="menu-item-839" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-839"><a href="/caring-for-our-parks/">Caring for Our Parks</a>\n<ul class="sub-menu">\n\t<li id="menu-item-840" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-840"><a href="/caring-for-our-parks/culture-heritage/">Culture &#038; Heritage</a></li>\n\t<li id="menu-item-842" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-842"><a href="/caring-for-our-parks/natural-spaces/">Natural Spaces</a></li>\n\t<li id="menu-item-959" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-959"><a href="/caring-for-our-parks/parks-and-trails/">Park and Trail Planning</a></li>\n\t<li id="menu-item-31505" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-31505"><a href="/caring-for-our-parks/sustainability/">Sustainability</a></li>\n\t<li id="menu-item-844" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-844"><a href="/caring-for-our-parks/wildlife/">Living with Wildlife</a></li>\n</ul>\n</li>\n<li id="menu-item-122" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-122"><a href="/projects/">Parks &#038; Projects</a>\n<ul class="sub-menu">\n\t<li id="menu-item-26054" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-26054"><a href="/projects/">Projects</a></li>\n\t<li id="menu-item-2503" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2503"><a href="/projects/capital-improvements-program/">Funding</a></li>\n\t<li id="menu-item-16631" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-16631"><a href="/projects/public-input/">Open Town Hall</a></li>\n\t<li id="menu-item-912" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-912"><a href="/projects/bids-procurement/">Bids / Procurement</a></li>\n\t<li id="menu-item-897" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-897"><a href="/projects/resources/">Resources</a></li>\n</ul>\n</li>\n<li id="menu-item-123" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-123"><a href="/support/">Support the Parks</a>\n<ul class="sub-menu">\n\t<li id="menu-item-19531" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-19531"><a href="/support/advocacy/">Advocacy</a></li>\n\t<li id="menu-item-907" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-907"><a href="/support/donate/">Donate</a></li>\n\t<li id="menu-item-15740" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-15740"><a href="/support/partnerships/">Partnerships</a></li>\n\t<li id="menu-item-900" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-900"><a href="/support/sponsorships/">Sponsorships</a></li>\n\t<li id="menu-item-901" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-901"><a href="/support/volunteer/">Volunteer</a></li>\n</ul>\n</li>\n<li id="menu-item-1047" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-1047"><a href="/about/parks/">About</a>\n<ul class="sub-menu">\n\t<li id="menu-item-660" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-660"><a href="/about/parks/">About the Parks</a></li>\n\t<li id="menu-item-659" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-659"><a href="/about/divisions/">Divisions &#038; Offices</a></li>\n\t<li id="menu-item-30188" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-30188"><a href="/about/media-center/">Media Center</a></li>\n\t<li id="menu-item-2592" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2592"><a href="/about/park-police/">Police</a></li>\n\t<li id="menu-item-655" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-655"><a href="/about/employment/">Employment</a></li>\n\t<li id="menu-item-932" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-932"><a href="/about/contact-us/">Contact Us</a></li>\n</ul>\n</li>\n</ul></div>\t\t    </div>\r\n\t\t</div>\r\n\t\t<div id="modal">\r\n\t    \t<div class="close">\xc3\x97</div>\r\n\t\t\t<div class="modal-wrapper">\r\n\t\t\t\t<div class="close">\xc3\x97</div>\r\n\t\t\t\tTo change the text size you can use your web browser\'s settings. Most browsers include functionality to let you increase or decrease the text on a web page. For example, to increase text size using:<br><br>\r\n\r\n\t\t\t\t<strong>Chrome</strong><br>\r\n\r\n\t\t\t\tIn the menu to the right of the address bar, select and set Zoom level. Menu > Zoom > +<br><br>\r\n\r\n\t\t\t\t<strong>Firefox</strong><br>\r\n\r\n\t\t\t\tIn the View menu, select Zoom. View > Zoom > Zoom In<br><br>\r\n\r\n\t\t\t\t<strong>Internet Explorer</strong><br>\r\n\r\n\t\t\t\tIn the View menu, select Text Size. View > Text Size > Largest<br><br>\r\n\r\n\t\t\t\t<strong>Safari</strong><br>\r\n\r\n\t\t\t\tIn the View menu, select Zoom In. View > Zoom In\r\n\t\t\t\tMacintosh Shortcut: Command+\r\n\t\t\t</div>\r\n\t\t</div>\r\n\t\t<div class="mobile-nav count-7">\r\n\t\t\t<div class="just-click"></div>\r\n\t\t\t<div class="upper-half">\r\n\t\t\t\t<div class="menu-primary-container"><ul id="menu-primary-1" class="menu"><li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-2177"><a href="/activities/">Things To Do</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-664"><a href="/activities/">Activities</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-8304"><a href="/classes-camps/">Classes &#038; Camps</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-950"><a href="/special-events/">Special Events &#038; Festivals</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-853"><a href="/parks-directory/">Parks &#038; Trails Directory</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1089"><a href="/parks-trails/parks/">Parks</a></li>\n\t<li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-39673"><a href="https://www.montgomeryparks.org/activities/park-trails/">Trails</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1092"><a href="/parks-trails/gardens/">Gardens</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2941"><a href="/parks-trails/facilities/">Facilities</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1090"><a href="/parks-trails/nature-centers/">Nature Centers</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2360"><a href="/parks-trails/event-centers/">Event Centers</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-837"><a href="/services/">Permits, Rentals &#038; Requests</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2850"><a href="/services/permits-rentals/">Permits &#038; Rentals</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-838"><a href="/services/report-a-problem/">Parks Information &#038; Customer Service</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-839"><a href="/caring-for-our-parks/">Caring for Our Parks</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-840"><a href="/caring-for-our-parks/culture-heritage/">Culture &#038; Heritage</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-842"><a href="/caring-for-our-parks/natural-spaces/">Natural Spaces</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-959"><a href="/caring-for-our-parks/parks-and-trails/">Park and Trail Planning</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-31505"><a href="/caring-for-our-parks/sustainability/">Sustainability</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-844"><a href="/caring-for-our-parks/wildlife/">Living with Wildlife</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-122"><a href="/projects/">Parks &#038; Projects</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-26054"><a href="/projects/">Projects</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2503"><a href="/projects/capital-improvements-program/">Funding</a></li>\n\t<li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-16631"><a href="/projects/public-input/">Open Town Hall</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-912"><a href="/projects/bids-procurement/">Bids / Procurement</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-897"><a href="/projects/resources/">Resources</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-123"><a href="/support/">Support the Parks</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-19531"><a href="/support/advocacy/">Advocacy</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-907"><a href="/support/donate/">Donate</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-15740"><a href="/support/partnerships/">Partnerships</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-900"><a href="/support/sponsorships/">Sponsorships</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-901"><a href="/support/volunteer/">Volunteer</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-1047"><a href="/about/parks/">About</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-660"><a href="/about/parks/">About the Parks</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-659"><a href="/about/divisions/">Divisions &#038; Offices</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-30188"><a href="/about/media-center/">Media Center</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2592"><a href="/about/park-police/">Police</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-655"><a href="/about/employment/">Employment</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-932"><a href="/about/contact-us/">Contact Us</a></li>\n</ul>\n</li>\n</ul></div>\t\t\t</div>\r\n\t\t\t<div class="lower-half rel">\r\n\t\t\t\t<div class="neat-search">\r\n\t\t\t\t\t<label for="m_search" />\r\n\t\t\t\t\t<input id="m_search" class="closed" type="text" placeholder="Search" />\r\n\t\t\t\t\t<a class="search-link t-indent" href="#m_search">Search</a>\r\n\t\t\t\t</div>\r\n\t\t\t\t<div class="page-nav">\r\n\t\t\t\t\t<a href="http://www.montgomeryparks.org/" target="_blank">Parks</a> &#8226;\r\n\t\t\t\t\t<a href="http://montgomeryplanning.org/" target="_blank">Planning</a> &#8226;\r\n\t\t\t\t\t<a href="http://montgomeryplanningboard.org/" target="_blank">Planning Board</a>\r\n\t\t\t\t</div>\r\n\t\t\t\t\t<div class="share">\n\t<a href="https://www.facebook.com/MontgomeryParks" target="_blank">facebook<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1fb{fill:#547bbc}.cls-2{fill:#fff}</style></defs><title>social-fb</title><path class="cls-1fb" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M11.4 12.5H13V11a4 4 0 0 1 .5-2.3 2.8 2.8 0 0 1 2.4-1.1 9.7 9.7 0 0 1 2.8.3l-.4 2.3-1.3-.3a.9.9 0 0 0-1.1.8v1.8h2.5l-.2 2.2h-2.4v7.8h-2.9v-7.8h-1.5v-2.2z"/></svg></a>\n\t<a href="https://twitter.com/MontgomeryParks" target="_blank">twitter<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1tw{fill:#5ea8dd}.cls-2{fill:#fff}</style></defs><title>social-tw</title><path class="cls-1tw" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M22.5 11.1l-1.8.5a3.1 3.1 0 0 0 1.4-1.7 6.2 6.2 0 0 1-2 .7 3.1 3.1 0 0 0-5.3 2.8 8.8 8.8 0 0 1-6.4-3.2 3.1 3.1 0 0 0 1 4.1 3.1 3.1 0 0 1-1.4-.4 3.1 3.1 0 0 0 2.5 3H9.2a3.1 3.1 0 0 0 2.9 2.1 6.2 6.2 0 0 1-3.8 1.3h-.8A8.8 8.8 0 0 0 21 13.1v-.4a6.2 6.2 0 0 0 1.5-1.6"/></svg></a>\n\t<a href="https://www.instagram.com/montgomeryparks/" target="_blank">instagram<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 291.319 291.319"><path d="M145.659 0c80.44 0 145.66 65.219 145.66 145.66S226.1 291.319 145.66 291.319 0 226.1 0 145.66 65.21 0 145.659 0z" fill="#9b6954"/><path d="M195.93 63.708H95.38c-17.47 0-31.672 14.211-31.672 31.672v100.56c0 17.47 14.211 31.672 31.672 31.672h100.56c17.47 0 31.672-14.211 31.672-31.672V95.38c-.001-17.461-14.212-31.672-31.682-31.672zm9.978 18.326l3.587-.009v27.202l-27.402.091-.091-27.202 23.906-.082zm-60.248 36.205c22.732 0 27.42 21.339 27.42 27.429 0 15.103-12.308 27.411-27.42 27.411-15.121 0-27.42-12.308-27.42-27.411-.01-6.09 4.688-27.429 27.42-27.429zm63.99 75.716c0 8.658-7.037 15.704-15.713 15.704H97.073c-8.667 0-15.713-7.037-15.713-15.704v-66.539h22.759c-2.112 5.198-3.305 12.299-3.305 18.253 0 24.708 20.101 44.818 44.818 44.818s44.808-20.11 44.808-44.818c0-5.954-1.193-13.055-3.296-18.253h22.486v66.539h.02z" fill="#fff"/></svg></a>\n\t<a href="https://www.youtube.com/user/MontgomeryParks" target="_blank">youtube<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1yt{fill:#e9644a}.cls-2{fill:#fff}</style></defs><title>social-yt</title><path class="cls-1yt" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M17.9 15L13 17.9v-5.8zm4.6 2.7v-5.4a2.5 2.5 0 0 0-2.6-2.6h-9.8a2.5 2.5 0 0 0-2.6 2.6v5.4a2.5 2.5 0 0 0 2.6 2.6h9.8a2.5 2.5 0 0 0 2.6-2.6"/></svg></a>\n\t<a href="http://www.montgomeryplanningboard.org/rss/rss.shtm" target="_blank">rss<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1rss{fill:#f29100}.cls-2{fill:#fff}</style></defs><title>social-rss</title><path class="cls-1rss" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M10.9 19.1a2 2 0 0 1 0 2.8 2 2 0 1 1-2.8-2.8 2 2 0 0 1 2.8 0zm-3.4-6.5v2.9a7 7 0 0 1 7 7h2.9a9.9 9.9 0 0 0-9.9-9.9zm0-5.1v2.9a12.2 12.2 0 0 1 12.1 12.1h2.9a15 15 0 0 0-15-15z"/></svg></a>\n</div>\t\t\t\t\t<a href="http://www.mncppc.org" title="Maryland-National Capital Parks and Commission" class="mncppc-link" target="_blank">Maryland-National Capital Park and Planning Commission</a>\r\n\t\t\t\t</div>\r\n\t\t\t</div>\r\n\t\t</div>\r\n\t\t<div id="mobile-footer">\r\n\t\t    <div class="footer-icon">\r\n\t\t    \t<a href="/" class="ico ico-home" title="Back to Home"><svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 31.9 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-home</title><path class="cls-1" d="M26.7 14.7a.9.9 0 0 0-.9.9v12.6h-6.5v-5.1a.9.9 0 0 0-.9-.9h-4.7a.9.9 0 0 0-.9.9v5.1H6.2V16a.9.9 0 0 0-1.8 0v13.1a.9.9 0 0 0 .9.9h8.4a.9.9 0 0 0 .9-.9V24h2.9v5.1a.9.9 0 0 0 .9.9h8.3a.9.9 0 0 0 .9-.9V15.6a.9.9 0 0 0-.9-.9zm0 0l5 .7-7.3-7.3v-5a.9.9 0 0 0-.9-.9.9.9 0 0 0-.9.9v3.2l-6-6A.9.9 0 0 0 16 0a.9.9 0 0 0-.7.3l-15 15a.9.9 0 0 0 0 1.3.9.9 0 0 0 1.3 0L16 2.2l14.4 14.5a.9.9 0 0 0 .7.2l.7-.2a1 1 0 0 0 0-1.3zm0 0"/></svg></a>\r\n\t\t    </div>\r\n\t\t    <div class="footer-icon">\r\n\t\t    <a href="/calendar" class="ico ico-calendar" title="Calendar"><svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-calendar</title><path class="cls-1" d="M27.5 2.8h-4.1V.9a.9.9 0 0 0-1.9 0v1.9h-5.6V.9a.9.9 0 0 0-1.9 0v1.9H8.4V.9a.9.9 0 1 0-1.8 0v1.9H2.5A2.5 2.5 0 0 0 0 5.3v22.2A2.5 2.5 0 0 0 2.5 30h25a2.5 2.5 0 0 0 2.5-2.5V5.3a2.5 2.5 0 0 0-2.5-2.5zm.6 24.7a.6.6 0 0 1-.6.6h-25a.6.6 0 0 1-.6-.6V5.3a.6.6 0 0 1 .6-.6h4.1v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h4.1a.6.6 0 0 1 .6.6v22.2zm0 0"/><path class="cls-1" d="M6.6 11.3h3.7v2.81H6.6zm0 4.6h3.7v2.81H6.6zm0 4.7h3.7v2.81H6.6zm6.5 0h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7zm6.6 9.3h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7z"/></svg></a></div>\r\n\t\t    <div class="footer-icon"><a href="/news" class="ico ico-news" title="News"><svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30.7 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-news</title><path class="cls-1" d="M30 2.5a.8.8 0 0 0-.8.8v21.1a4 4 0 0 1-4 4H5.7a4 4 0 0 1-4-4V1.6h23.2v20.5a.8.8 0 1 0 1.6 0V.8a.8.8 0 0 0-.8-.8H.8a.8.8 0 0 0-.8.8v23.6A5.6 5.6 0 0 0 5.7 30h19.4a5.6 5.6 0 0 0 5.6-5.6V3.3a.8.8 0 0 0-.8-.8zm0 0"/><path class="cls-1" d="M21.9 6.2a.8.8 0 0 0 0-1.6H4.5a.8.8 0 1 0 0 1.6h17.4zm0 14.6H4.5a.8.8 0 1 0 0 1.6h17.4a.8.8 0 0 0 0-1.6zm0 4.1H7.3a.8.8 0 0 0 0 1.6h14.6a.8.8 0 0 0 0-1.6zM3.5 9v8.6a.8.8 0 0 0 .8.8h8.6a.8.8 0 0 0 .8-.8V9a.8.8 0 0 0-.8-.8H4.3a.8.8 0 0 0-.8.8zm1.6.8h7v7h-7v-7zM22.3 9h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 0-1.6zm0 3.6h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 1 0 0-1.6zm.8 4.6a.8.8 0 0 0-.8-.8h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 .8-.8zm0 0"/></svg></a></div>\r\n\t\t    <div class="menu-open">\r\n\t\t    \t<div id="hamburger">\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t</div>\r\n\t\t\t</div>\r\n\t\t</div>\r\n\t\t<div id="siteBody" class="site-body">\r\n<div class="calendar left-sb" data-date="2019-01-29" data-year="2019" data-month="01" data-view="month">\n    <div class="section-headline row">\n        <div class="p-lg-x-l">\n            <h1 class="section-head">Calendar of Events</h1>\n        </div>\n    </div>\n    <div class="sb-flex rel">\n        <div class="col col-sm-desk-12 col-lg-4 side-l col-sb">\n            <div class="sb-wrap">\n                <div class="sidebar">\n\n                    <div class="calendar-box col-sm-desk-6 col-lg-12">\n                        <div class="row controls rel">\n                            <div class="prev f-left">\n                                <a href="?date=2018-12-01"><span class="hv">Prev</span></a>\n                            </div>\n\n                            <div class="title f-center">\n                                <h2></h2>\n                            </div>\n\n                            <div class="next f-right">\n                                <a href="?date=2019-03-01"><span class="hv">Next</span></a>\n                            </div>\n                        </div>\n                        <div class="row days f-center m-y">\n                            <div class="col col-7th">Sun</div>\n                            <div class="col col-7th">Mon</div>\n                            <div class="col col-7th">Tue</div>\n                            <div class="col col-7th">Wed</div>\n                            <div class="col col-7th">Thu</div>\n                            <div class="col col-7th">Fri</div>\n                            <div class="col col-7th">Sat</div>\n                        </div>\n\n                        <div class="js-dates row">\n\n                        </div>\n                    </div>\n                    <div class="p-y">\n                        <h4>Search for An Event</h4>\n                        <div class="neat-search">\n                            <form action="" method="get" class="cal-search">\n                              <label for="search-2">Search Events</label>\n                              <input name="s" id="search-2" class="closed" type="text" placeholder="Search" value=""/>\n                              <button class="search-link" type="submit">Search Events</button>\n                            </form>\n                        </div>\n                    </div>\n                                            <div class="p-y-b inline-block">\n                            <h4>Search by Park</h4>\n                            <div class="neat-search">\n                                <form action="" method="get" class="cal-park-search">\n                                    <label for="search-5">Search Parks</label>\n                                    <input id="search-5" class="closed park-s" type="text" placeholder="Park Name" name="park" value=""/>\n                                    <button class="search-link" href="#search-5" type="submit">Search Parks</button>\n                                </form>\n                            </div>\n                        </div>\n                                            <div class="filter-list accordion col-sm-desk-6 col-lg-12">\n                        <div class="arrow"></div>\n                        <h4>Filter by category</h4>\n                        <a href="?cat=" id="clear" class="clear"><svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 47.5 45.1"><path d="M46.1 29.6c-.4-.1-.9.2-.9.6-.9 4.4-4.4 7.4-8.6 7.4-4 0-7.6-2.6-8.8-6.4h3.8v-1.6h-6.4V36h1.6v-3.1c1.7 3.7 5.5 6.3 9.8 6.3 5 0 9.1-3.5 10.2-8.6 0-.5-.3-.9-.7-1zM45.9 20v3.4c-1.8-3.6-5.4-5.8-9.4-5.8-4.9 0-9.2 3.2-10.6 7.8-.1.4.1.9.5 1 .4.1.9-.1 1-.5 1.1-3.8 4.9-6.6 9-6.6 3.7 0 6.8 2.2 8.3 5.6H41v1.6h6.4V20h-1.5zM30.8 10.5c0 1.7 1.4 3.1 3.1 3.1 1.7 0 3.2-1.4 3.2-3.1 0-1.7-1.4-3.1-3.2-3.1-1.7 0-3.1 1.4-3.1 3.1zm4.7 0c0 .9-.7 1.6-1.6 1.6-.9 0-1.5-.7-1.5-1.6 0-.9.7-1.5 1.5-1.5.9 0 1.6.7 1.6 1.5z"/><path d="M25.7 1.6h9.9l8 7.9v7.3h1.6V8.9l-9-8.9H25L4.8 20.2 0 25l20.1 20.2 6.2-5.9-19.2-19L25.7 1.6zM24 39.2l-3.9 3.7L2.3 25 6 21.3l18 17.9z"/></svg>clear</a>\n\n                        <ul class="filters accordion-wrap">\n                                                            <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1751&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Archaeology                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2230&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Butterfly Show                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2701&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Camp                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2235&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Clean Up                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1738&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Community Events                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1739&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Community Gardens                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2577&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Community Meeting                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=3105&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Earth Month                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2239&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Education                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=3331&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Event Center                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1740&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Events                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1741&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Family                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2328&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Featured                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2232&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Festival                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1746&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Free                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2243&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Gardens                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2242&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Golf                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1733&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Hikes                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1743&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Historical/Cultural                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2240&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                History                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=3215&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                History in the Parks                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2234&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                History Program                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1727&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Hockey                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2241&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Holiday                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1728&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Ice Skating                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1729&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Kids                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2705&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                MLK Day of Service                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1744&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Music                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=82&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Nature                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2901&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                open house                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1753&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Openings/Dedications                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2233&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Other                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2902&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Park Planning                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2916&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Park Projects                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2751&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Planning Board Meeting                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2756&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Planning Board Worksession                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2569&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                pop up                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2754&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Public Hearing                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1750&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Public Meetings                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2231&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Run                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1730&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Special Events                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1731&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Sports                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1734&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Tennis                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2903&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Trail Planning                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2706&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Trail Work                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1735&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Trails                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1736&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Trips                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1747&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Volunteering                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1748&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Weed Warrior                                    </a>\n                                </li>\n                                                        </ul>\n                    </div>\n                </div>\n            </div>\n        </div>\n        <div class="col col-sm-desk-12 col-lg-5 events">\n                <!--p>2 events meet your criteria of [filter item] & [filter item].</p-->\n            <section class="event-list" id="ajax-event-list">\n                <div class="p-y" style="text-align: right">\n                    <a href="#" class="toggle-view toggle-view-day">Switch to Day View</a>                                    </div>\n                <div class="event-item events-2019-01-29 view-m">\n                                        <ul>\n                                                <li>\n                                <a data-id="41841" href="/events/time-for-tots-groundhog-gets-a-say-/">\n                                    <span class="time">\n                                        Wed. January 30th, 2019 10:00am to  11:00am                                    </span>\n                                    <span class="event-name m-view">Time for Tots: &#8220;Groundhog Gets a Say&#8221;</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/black-hill-nature-programs/\' class=\'location\'>Black Hill Nature Programs</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="41962" href="/events/nature-rx-forest-therapy-walks-1-hour-/">\n                                    <span class="time">\n                                        Wed. January 30th, 2019 10:30am to  11:30am                                    </span>\n                                    <span class="event-name m-view">Nature Rx: Forest Therapy Walks (1 hour)</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-nature-center/\' class=\'location\'>Brookside Nature Center</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="41878" href="/events/beat-the-winter-doldrums-camp-for-adults-4/">\n                                    <span class="time">\n                                        Thu. January 31st, 2019 7:00pm to  8:30pm                                    </span>\n                                    <span class="event-name m-view">Beat the Winter Doldrums Camp for Adults!</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-nature-center/\' class=\'location\'>Brookside Nature Center</span>                                    </div>\n                                                                </li>\n                                                </ul>\n                    <div class="pag row">\n                        <div class="col col-6"></div>\n                        <div style="text-align: right;" class="col col-6 a-right"></div>\n                    </div>\n                                    </div>\n            </section>\n            <div class="loading"></div>\n        </div>\n        <div class="col col-sm-desk-12 col-lg-3 p-lg-y-t featured-events">\n                                <div class="popular-events p">\n                        <h3>Featured Events</h3>\n\n                                                    <div class="p-y single-pop-event">\n                                <a href="/events/we-speak-weddings-lodge-at-little-seneca-creek/">\n                                    We Speak Weddings &#8211; Lodge at Little Seneca Creek                                </a><br />\n                                Sun. February 3rd, 2019 10:00am to  1:00pm                            </div>\n                                                        <div class="p-y single-pop-event">\n                                <a href="/events/volunteer-fair-for-montgomery-parks-historic-sites/">\n                                    Join Parks Historic Sites Volunteer Program                                </a><br />\n                                Sat. March 23rd, 2019 9:00am to  3:00pm                            </div>\n                                                </div>\n                            </div>\n    </div>\n</div>\n<style>\n.m-no-time span.time{\n    display:none !important;\n}\n\n.m-no-time span.event-name{\n    padding-left: 0 !important;\n}\n</style>\n<div id="footer" class="main-footer">\r\n                        \r\n                    <section class="hot-topics row">\r\n                        <div class="cc-m">\r\n                            <h3>Topics</h3>\r\n                            <aside class="hot-topic">\r\n                                <div class="menu-hot-topics-container"><ul id="menu-hot-topics" class="menu"><li id="menu-item-14005" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-14005"><a href="/special-events/">Festivals &#038; Events</a></li>\n<li id="menu-item-14007" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-14007"><a href="/classes-camps/">Register Now for Classes &#038; Programs</a></li>\n<li id="menu-item-14008" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-14008"><a href="/support/volunteer/seasonal_volunteering/">Volunteer Events</a></li>\n<li id="menu-item-16517" class="menu-item menu-item-type-custom menu-item-object-custom current-menu-item menu-item-16517"><a href="/calendar/">Calendar of Events</a></li>\n<li id="menu-item-14011" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-14011"><a href="/caring-for-our-parks/wildlife/ticks-lyme-disease/">Ticks and Lyme Disease</a></li>\n<li id="menu-item-16515" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-16515"><a href="/projects/public-input/">Open Town Hall</a></li>\n<li id="menu-item-18385" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-18385"><a href="/activities/history-in-the-parks/">History in the Parks</a></li>\n<li id="menu-item-33409" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-33409"><a href="https://www.montgomeryparks.org/about/parks/pesticides/">Pesticide Reduction &#038; Integrated Pest Management</a></li>\n</ul></div>                            </aside>\r\n                                                    </div>\r\n                    </section>\r\n \r\n                    <section class="main-footer-bottom row">\r\n        <div class="col col-mob-12 col-desk-3">\r\n            <div class="address">\r\n                Montgomery County<br>Department of Parks<br>\r\n                9500 Brunett Ave.<br>\r\n                Silver Spring, MD 20901\r\n            </div>\r\n        </div>\r\n                        <div class="col col-mob-6 col-sm-desk-4 col-desk-2 footer-nav">\r\n                                            <a href="/about/contact-us/">Contact Us</a>\r\n                                                <a href="/about/media-center/press-contacts/">Press</a>\r\n                                                <a href="/about/parks/accessibility/">ADA Accessibility</a>\r\n                                                <a href="/terms/">Terms & Privacy</a>\r\n                                        </div>\r\n                                <div class="col col-mob-6 col-sm-desk-4 col-desk-2 footer-nav">\r\n                                            <a href="/about/park-police/">Park Police</a>\r\n                                                <a href="/about/rules-and-regulations/">Rules and Regulations</a>\r\n                                                <a href="/support/volunteer/">Volunteer</a>\r\n                                                <a href="/activities/">Activities</a>\r\n                                        </div>\r\n                                <div class="col col-mob-6 col-sm-desk-4 col-desk-2 footer-nav">\r\n                                            <a href="/permits/">Permits</a>\r\n                                                <a href="/calendar/">Calendar</a>\r\n                                                <a href="/news/">News</a>\r\n                                                <a href="/weather-alerts/">Weather Alerts</a>\r\n                                        </div>\r\n                        <div class="col col-sm-desk-3 pull-right mob-hide rel">\r\n            SOCIAL<br>\r\n            <div class="share">\n\t<a href="https://www.facebook.com/MontgomeryParks" target="_blank">facebook<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1fb{fill:#547bbc}.cls-2{fill:#fff}</style></defs><title>social-fb</title><path class="cls-1fb" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M11.4 12.5H13V11a4 4 0 0 1 .5-2.3 2.8 2.8 0 0 1 2.4-1.1 9.7 9.7 0 0 1 2.8.3l-.4 2.3-1.3-.3a.9.9 0 0 0-1.1.8v1.8h2.5l-.2 2.2h-2.4v7.8h-2.9v-7.8h-1.5v-2.2z"/></svg></a>\n\t<a href="https://twitter.com/MontgomeryParks" target="_blank">twitter<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1tw{fill:#5ea8dd}.cls-2{fill:#fff}</style></defs><title>social-tw</title><path class="cls-1tw" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M22.5 11.1l-1.8.5a3.1 3.1 0 0 0 1.4-1.7 6.2 6.2 0 0 1-2 .7 3.1 3.1 0 0 0-5.3 2.8 8.8 8.8 0 0 1-6.4-3.2 3.1 3.1 0 0 0 1 4.1 3.1 3.1 0 0 1-1.4-.4 3.1 3.1 0 0 0 2.5 3H9.2a3.1 3.1 0 0 0 2.9 2.1 6.2 6.2 0 0 1-3.8 1.3h-.8A8.8 8.8 0 0 0 21 13.1v-.4a6.2 6.2 0 0 0 1.5-1.6"/></svg></a>\n\t<a href="https://www.instagram.com/montgomeryparks/" target="_blank">instagram<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 291.319 291.319"><path d="M145.659 0c80.44 0 145.66 65.219 145.66 145.66S226.1 291.319 145.66 291.319 0 226.1 0 145.66 65.21 0 145.659 0z" fill="#9b6954"/><path d="M195.93 63.708H95.38c-17.47 0-31.672 14.211-31.672 31.672v100.56c0 17.47 14.211 31.672 31.672 31.672h100.56c17.47 0 31.672-14.211 31.672-31.672V95.38c-.001-17.461-14.212-31.672-31.682-31.672zm9.978 18.326l3.587-.009v27.202l-27.402.091-.091-27.202 23.906-.082zm-60.248 36.205c22.732 0 27.42 21.339 27.42 27.429 0 15.103-12.308 27.411-27.42 27.411-15.121 0-27.42-12.308-27.42-27.411-.01-6.09 4.688-27.429 27.42-27.429zm63.99 75.716c0 8.658-7.037 15.704-15.713 15.704H97.073c-8.667 0-15.713-7.037-15.713-15.704v-66.539h22.759c-2.112 5.198-3.305 12.299-3.305 18.253 0 24.708 20.101 44.818 44.818 44.818s44.808-20.11 44.808-44.818c0-5.954-1.193-13.055-3.296-18.253h22.486v66.539h.02z" fill="#fff"/></svg></a>\n\t<a href="https://www.youtube.com/user/MontgomeryParks" target="_blank">youtube<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1yt{fill:#e9644a}.cls-2{fill:#fff}</style></defs><title>social-yt</title><path class="cls-1yt" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M17.9 15L13 17.9v-5.8zm4.6 2.7v-5.4a2.5 2.5 0 0 0-2.6-2.6h-9.8a2.5 2.5 0 0 0-2.6 2.6v5.4a2.5 2.5 0 0 0 2.6 2.6h9.8a2.5 2.5 0 0 0 2.6-2.6"/></svg></a>\n\t<a href="http://www.montgomeryplanningboard.org/rss/rss.shtm" target="_blank">rss<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1rss{fill:#f29100}.cls-2{fill:#fff}</style></defs><title>social-rss</title><path class="cls-1rss" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M10.9 19.1a2 2 0 0 1 0 2.8 2 2 0 1 1-2.8-2.8 2 2 0 0 1 2.8 0zm-3.4-6.5v2.9a7 7 0 0 1 7 7h2.9a9.9 9.9 0 0 0-9.9-9.9zm0-5.1v2.9a12.2 12.2 0 0 1 12.1 12.1h2.9a15 15 0 0 0-15-15z"/></svg></a>\n</div>        </div>\r\n    </section>\r\n</div>\r\n        <style>\n    .brink-gdpr-cookie-middle{\n      border-right:1px solid rgba(255,255,255,.5);\n    }\n\n    .brink-gdpr-cookie-left a{\n      color:#fff;\n    }\n\n    #brink-gdpr-cookie-accept{\n      background:rgb(0,0,0);\n      color:#fff;\n    }\n\n    #brink-gdpr-cookie-accept:hover, #brink-gdpr-cookie-accept:active{\n      background:rgb(255,255,255);\n      color:#000;\n    }\n\n    .brink-gdpr-cookie-right span:before{\n      border-top:1px solid #fff;\n      border-right:1px solid #fff;\n    }\n\n    .brink-gdpr-cookie-right span:after{\n      border-top:1px solid #fff;\n      border-left:1px solid #fff;\n    }\n    </style>\n    <div id="brink-gdpr-cookie-notice" class="brink-gdpr-cookie-notice bottom" style="background:rgba(102,102,102,.95); color:#fff">\n      <div class="brink-gdpr-container">\n        <div class="brink-gdpr-cookie-left">\n          <span class="h2">Cookies Notice</span>\n            <p>\n              This website uses cookies to provide you with an improved and personalized experience. By using this site you agree to our use of cookies. Please read our <a href="/terms/" target="_blank">cookies policy</a> for more information on the cookies we use and how to delete or block them.            </p>\n          </span>\n        </div>\n        <div class="brink-gdpr-cookie-middle">\n          <a href="#" id="brink-gdpr-cookie-accept">I Accept</a>\n        </div>\n        <div class="brink-gdpr-cookie-right">\n          <a href="#" id="brink-gdpr-cookie-close"></a>\n          <span></span>\n        </div>\n      </div>\n    </div>\n    <script type="text/javascript">\n      window.brink_gdpr_whitelist = [];\n    </script>\n    <script data-cfasync="false" type="text/javascript">if (window.addthis_product === undefined) { window.addthis_product = "wpp"; } if (window.wp_product_version === undefined) { window.wp_product_version = "wpp-6.1.8"; } if (window.wp_blog_version === undefined) { window.wp_blog_version = "5.0.3"; } if (window.addthis_share === undefined) { window.addthis_share = {}; } if (window.addthis_config === undefined) { window.addthis_config = {"data_track_clickback":true,"ui_atversion":"300"}; } if (window.addthis_plugin_info === undefined) { window.addthis_plugin_info = {"info_status":"enabled","cms_name":"WordPress","plugin_name":"Share Buttons by AddThis","plugin_version":"6.1.8","plugin_mode":"AddThis","anonymous_profile_id":"wp-f82eef1f055e09dcb6273ac5462e3333","page_info":{"template":"pages","post_type":""},"sharing_enabled_on_post_via_metabox":false}; } \n                    (function() {\n                      var first_load_interval_id = setInterval(function () {\n                        if (typeof window.addthis !== \'undefined\') {\n                          window.clearInterval(first_load_interval_id);\n                          if (typeof window.addthis_layers !== \'undefined\' && Object.getOwnPropertyNames(window.addthis_layers).length > 0) {\n                            window.addthis.layers(window.addthis_layers);\n                          }\n                          if (Array.isArray(window.addthis_layers_tools)) {\n                            for (i = 0; i < window.addthis_layers_tools.length; i++) {\n                              window.addthis.layers(window.addthis_layers_tools[i]);\n                            }\n                          }\n                        }\n                     },1000)\n                    }());\n                </script> <script data-cfasync="false" type="text/javascript" src="https://s7.addthis.com/js/300/addthis_widget.js#pubid=ra-5a0206a251dc7e59" async="async"></script><script type=\'text/javascript\'>\n/* <![CDATA[ */\nvar countVars = {"disqusShortname":"montgomeryparks"};\n/* ]]> */\n</script>\n<script type=\'text/javascript\' src=\'/wp-content/plugins/disqus-comment-system/public/js/comment_count.js?ver=3.0.16\'></script>\n<script type=\'text/javascript\' src=\'/wp-content/themes/mdparks/dist/js/main.js?ver=885\'></script>\n<script type=\'text/javascript\' src=\'https://js.arcgis.com/3.18/?ver=5.0.3\'></script>\n<script type=\'text/javascript\' src=\'/wp-includes/js/wp-embed.min.js?ver=5.0.3\'></script>\n<script>\r\n  window.addEventListener(\'load\',function(){\r\n    jQuery(\'[href*=".pdf"]\').click(function(){\r\n      __gaTracker(\'send\',\'event\',\'PDF\',\'click\',jQuery(this).attr(\'href\'));\r\n    })\r\n  });\r\n</script>\r\n<img height="1" width="1" style="border-style:none;" alt="" src="//googleads.g.doubleclick.net/pagead/viewthroughconversion/974628909/?guid=ON&amp;script=0"/>\r\n\r\n</body>\r\n\r\n</html>\r\n<!-- AWS -->\r\n'
 
 event_page_content = b'<!DOCTYPE html>\r\n<html lang="en-US" xmlns:fb="https://www.facebook.com/2008/fbml" xmlns:addthis="https://www.addthis.com/help/api-spec"  prefix="og: http://ogp.me/ns#" class="no-js">\r\n\t<head>\r\n\t\t<meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">\r\n\t\t<script src="https://use.typekit.net/bxi2fgb.js"></script>\r\n\t\t<script>try{Typekit.load({ async: true });}catch(e){}</script>\r\n\t\t<meta charset="UTF-8">\r\n\t\t<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=12.0, minimum-scale=.25, user-scalable=yes"/>\r\n\t\t<!-- CHECK META CONTENT -->\r\n\t\t<title>Maple Sugaring Days - Montgomery Parks</title>\r\n\r\n        <script>\r\n          // helpful for understanding dojoConfig.packages vs. dojoConfig.paths:\r\n          // http://www.sitepen.com/blog/2013/06/20/dojo-faq-what-is-the-difference-packages-vs-paths-vs-aliases/\r\n          var dojoConfig = {\r\n            paths: {\r\n              extras: "/wp-content/themes/mdparks/dist/js/extras"\r\n            }\r\n          };\r\n        </script>\r\n\r\n\t\t<link rel="pingback" href="/xmlrpc.php">\r\n\t\t\t\t<style>#skipNav{display: block;background-color:#fff;padding:1em;position: absolute;top: 0;left: -9999px;}#skipNav:focus{left:0;}</style>\r\n\r\n\t\t\n<!-- This site is optimized with the Yoast SEO plugin v9.5 - https://yoast.com/wordpress/plugins/seo/ -->\n<link rel="canonical" href="/events/maple-sugaring-days-147/" />\n<meta property="og:locale" content="en_US" />\n<meta property="og:type" content="article" />\n<meta property="og:title" content="Maple Sugaring Days - Montgomery Parks" />\n<meta property="og:description" content="February is Maple Sugaring Month at Brookside Nature Center. Every Saturday and Sunday you&#8217;ll have an opportunity to experience an..." />\n<meta property="og:url" content="/events/maple-sugaring-days-147/" />\n<meta property="og:site_name" content="Montgomery Parks" />\n<meta property="article:tag" content="Nature" />\n<meta name="twitter:card" content="summary" />\n<meta name="twitter:description" content="February is Maple Sugaring Month at Brookside Nature Center. Every Saturday and Sunday you&#8217;ll have an opportunity to experience an..." />\n<meta name="twitter:title" content="Maple Sugaring Days - Montgomery Parks" />\n<!-- / Yoast SEO plugin. -->\n\n<link rel=\'dns-prefetch\' href=\'//js.arcgis.com\' />\n<link rel=\'dns-prefetch\' href=\'//s.w.org\' />\n<link rel=\'dns-prefetch\' href=\'//s3.amazonaws.com\' />\n<link rel="alternate" type="application/rss+xml" title="Montgomery Parks &raquo; Feed" href="/feed/" />\n<link rel="alternate" type="application/rss+xml" title="Montgomery Parks &raquo; Comments Feed" href="/comments/feed/" />\n<link rel="alternate" type="application/rss+xml" title="Montgomery Parks &raquo; Maple Sugaring Days Comments Feed" href="/events/maple-sugaring-days-147/feed/" />\n<!-- This site uses the Google Analytics by MonsterInsights plugin v7.4.2 - Using Analytics tracking - https://www.monsterinsights.com/ -->\r\n<script type="text/javascript" data-cfasync="false">\r\n\tvar mi_version         = \'7.4.2\';\r\n\tvar mi_track_user      = true;\r\n\tvar mi_no_track_reason = \'\';\r\n\t\r\n\tvar disableStr = \'ga-disable-UA-493554-2\';\r\n\r\n\t/* Function to detect opted out users */\r\n\tfunction __gaTrackerIsOptedOut() {\r\n\t\treturn document.cookie.indexOf(disableStr + \'=true\') > -1;\r\n\t}\r\n\r\n\t/* Disable tracking if the opt-out cookie exists. */\r\n\tif ( __gaTrackerIsOptedOut() ) {\r\n\t\twindow[disableStr] = true;\r\n\t}\r\n\r\n\t/* Opt-out function */\r\n\tfunction __gaTrackerOptout() {\r\n\t  document.cookie = disableStr + \'=true; expires=Thu, 31 Dec 2099 23:59:59 UTC; path=/\';\r\n\t  window[disableStr] = true;\r\n\t}\r\n\t\r\n\tif ( mi_track_user ) {\r\n\t\t(function(i,s,o,g,r,a,m){i[\'GoogleAnalyticsObject\']=r;i[r]=i[r]||function(){\r\n\t\t\t(i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),\r\n\t\t\tm=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)\r\n\t\t})(window,document,\'script\',\'//www.google-analytics.com/analytics.js\',\'__gaTracker\');\r\n\r\n\t\t__gaTracker(\'create\', \'UA-493554-2\', \'auto\');\n\t\t__gaTracker(\'set\', \'forceSSL\', true);\n\t\t__gaTracker(\'set\', \'anonymizeIp\', true);\n\t\t__gaTracker(\'require\', \'linkid\', \'linkid.js\');\n\t\t__gaTracker(\'send\',\'pageview\');\n\t} else {\r\n\t\tconsole.log( "" );\r\n\t\t(function() {\r\n\t\t\t/* https://developers.google.com/analytics/devguides/collection/analyticsjs/ */\r\n\t\t\tvar noopfn = function() {\r\n\t\t\t\treturn null;\r\n\t\t\t};\r\n\t\t\tvar noopnullfn = function() {\r\n\t\t\t\treturn null;\r\n\t\t\t};\r\n\t\t\tvar Tracker = function() {\r\n\t\t\t\treturn null;\r\n\t\t\t};\r\n\t\t\tvar p = Tracker.prototype;\r\n\t\t\tp.get = noopfn;\r\n\t\t\tp.set = noopfn;\r\n\t\t\tp.send = noopfn;\r\n\t\t\tvar __gaTracker = function() {\r\n\t\t\t\tvar len = arguments.length;\r\n\t\t\t\tif ( len === 0 ) {\r\n\t\t\t\t\treturn;\r\n\t\t\t\t}\r\n\t\t\t\tvar f = arguments[len-1];\r\n\t\t\t\tif ( typeof f !== \'object\' || f === null || typeof f.hitCallback !== \'function\' ) {\r\n\t\t\t\t\tconsole.log( \'Not running function __gaTracker(\' + arguments[0] + " ....) because you are not being tracked. " + mi_no_track_reason );\r\n\t\t\t\t\treturn;\r\n\t\t\t\t}\r\n\t\t\t\ttry {\r\n\t\t\t\t\tf.hitCallback();\r\n\t\t\t\t} catch (ex) {\r\n\r\n\t\t\t\t}\r\n\t\t\t};\r\n\t\t\t__gaTracker.create = function() {\r\n\t\t\t\treturn new Tracker();\r\n\t\t\t};\r\n\t\t\t__gaTracker.getByName = noopnullfn;\r\n\t\t\t__gaTracker.getAll = function() {\r\n\t\t\t\treturn [];\r\n\t\t\t};\r\n\t\t\t__gaTracker.remove = noopfn;\r\n\t\t\twindow[\'__gaTracker\'] = __gaTracker;\r\n\t\t\t\t\t})();\r\n\t\t}\r\n</script>\r\n<!-- / Google Analytics by MonsterInsights -->\r\n<link rel=\'stylesheet\' id=\'wp-block-library-group-css\' href=\'/wp-content/plugins/bwp-minify/min/?f=wp-includes/css/dist/block-library/style.min.css,wp-content/themes/mdparks/dist/css/main.min.css,wp-content/plugins/addthis/frontend/build/addthis_wordpress_public.min.css\' type=\'text/css\' media=\'all\' />\r\n<link rel=\'stylesheet\' id=\'brink_gdpr_css-group-css\' href=\'/wp-content/plugins/bwp-minify/min/?f=wp-content/plugins/brink-gdpr/css/brink-gdpr.css\' type=\'text/css\' media=\'screen\' />\r\n<link rel=\'stylesheet\' id=\'esri-css-css\'  href=\'https://js.arcgis.com/3.18/esri/css/esri.css?ver=5.0.3\' type=\'text/css\' media=\'all\' />\n<script type=\'text/javascript\'>\n/* <![CDATA[ */\nvar monsterinsights_frontend = {"js_events_tracking":"true","download_extensions":"doc,exe,js,pdf,ppt,tgz,zip,xls","inbound_paths":"[]","home_url":"http:\\/\\/staging.montgomeryparks.org","hash_tracking":"false"};\n/* ]]> */\n</script>\n<script type=\'text/javascript\' src=\'/wp-content/plugins/google-analytics-for-wordpress/assets/js/frontend.min.js?ver=7.4.2\'></script>\n<script type=\'text/javascript\' src=\'/wp-content/plugins/brink-gdpr/js/brink_gdpr.js?ver=5.0.3\'></script>\n<link rel=\'https://api.w.org/\' href=\'/wp-json/\' />\n<link rel="EditURI" type="application/rsd+xml" title="RSD" href="/xmlrpc.php?rsd" />\n<link rel="wlwmanifest" type="application/wlwmanifest+xml" href="/wp-includes/wlwmanifest.xml" /> \n<meta name="generator" content="WordPress 5.0.3" />\n<link rel=\'shortlink\' href=\'/?p=41896\' />\n<link rel="alternate" type="application/json+oembed" href="/wp-json/oembed/1.0/embed?url=http%3A%2F%2Fstaging.montgomeryparks.org%2Fevents%2Fmaple-sugaring-days-147%2F" />\n<link rel="alternate" type="text/xml+oembed" href="/wp-json/oembed/1.0/embed?url=http%3A%2F%2Fstaging.montgomeryparks.org%2Fevents%2Fmaple-sugaring-days-147%2F&#038;format=xml" />\n<script type="text/javascript">\nvar ajaxurl = \'/wp-admin/admin-ajax.php/\';\n</script>\n\t\t<style type="text/css">.recentcomments a{display:inline !important;padding:0 !important;margin:0 !important;}</style>\n\t\t<link rel="icon" href="/uploads/2016/07/cropped-favicon-490x490.png" sizes="32x32" />\n<link rel="icon" href="/uploads/2016/07/cropped-favicon-490x490.png" sizes="192x192" />\n<link rel="apple-touch-icon-precomposed" href="/uploads/2016/07/cropped-favicon-490x490.png" />\n<meta name="msapplication-TileImage" content="/uploads/2016/07/cropped-favicon-490x490.png" />\n\r\n\t\t\t</head>\r\n\r\n\t<body class="events-template-default single single-events postid-41896 group-blog">\r\n\r\n\r\n\r\n\t\t<div id="page" class="page">\r\n\t\t\t<div class="main-header parks-header">\r\n\t\t    <div class="mncppc">\r\n\t\t        <a href="http://www.mncppc.org" title="Maryland-National Capital Parks and Commission" class="mncppc-link" target="_blank">Maryland-National Capital Park and Planning Commission</a>\r\n\t\t        <ul class="sites-links">\r\n\t\t\t\t\t<li class="parks current"><a href="/parks-home">Parks</a></li>\r\n\t\t            <li class="planning"><a href="http://montgomeryplanning.org" target="_blank">Planning</a></li>\r\n\t\t            <li class="board"><a href="http://montgomeryplanningboard.org" target="_blank">Planning Board</a></li>\r\n\t\t        </ul>\r\n\t\t    </div>\r\n\t\t    <div class="branding">\r\n\t\t    \t<div class="col col-tab-9 col-sm-desk-6 col-lg-5 left-wrap p-x">\r\n\t\t    \t\t<div class="mid-v">\r\n\t\t\t\t        <div class="site-icon">\r\n\t\t\t\t        \t<a href="/" class="parks-logo t-indent" title="Back to Home">Montgomery Parks</a>\r\n\t\t\t\t        </div>\r\n\t\t\t\t        <div class="site-title">\r\n\t\t\t\t        \t<a href="/" class="parks-logo" title="Back to Home">\r\n\t\t\t\t        \t<div class="title-wrap">\r\n\t\t\t\t\t            <p class="montco">Montgomery County, MD</p>\r\n\t\t\t\t\t            <h2 class="name">Montgomery Parks</h2>\r\n\t\t\t\t            </div>\r\n\t\t\t\t            </a>\r\n\t\t\t\t        </div>\r\n\t\t\t        </div>\r\n\t\t        </div>\r\n\t\t\t\t<div class="col col-tab-3 col-desk-2 viewing-options">\r\n\t\t\t\t<!-- \t<div class="head-col-wrap"> -->\r\n\t\t\t\t\t\t<div class="options-wrap">\r\n\t\t\t\t\t\t\t<div id="google_translate_element" tabindex="-1"></div>\r\n\t\t\t\t\t\t\t<script type="text/javascript">\r\n                            function fireEvent(element,event){\r\n                                if (document.createEventObject){\r\n                                    // dispatch for IE\r\n                                    var evt = document.createEventObject();\r\n                                    return element.fireEvent(\'on\'+event,evt)\r\n                                }\r\n                                else{\r\n                                    // dispatch for firefox + others\r\n                                    var evt = document.createEvent("HTMLEvents");\r\n                                    evt.initEvent(event, true, true ); // event type,bubbling,cancelable\r\n                                    return !element.dispatchEvent(evt);\r\n                                }\r\n                             }\r\n\r\n                            function googleTranslateElementInit() {\r\n                                new google.translate.TranslateElement({pageLanguage: \'en\', includedLanguages: \'af,ach,ak,am,ar,az,be,bem,bg,bh,bn,br,bs,ca,chr,ckb,co,crs,cs,cy,da,de,ee,el,en,eo,es,es-419,et,eu,fa,fi,fo,fr,fy,ga,gaa,gd,gl,gn,gu,ha,haw,hi,hr,ht,hu,hy,ia,id,ig,is,it,iw,ja,jw,ka,kg,kk,km,kn,ko,kri,ku,ky,la,lg,ln,lo,loz,lt,lua,lv,mfe,mg,mi,mk,ml,mn,mo,mr,ms,mt,ne,nl,nn,no,nso,ny,nyn,oc,om,or,pa,pcm,pl,ps,pt-BR,pt-PT,qu,rm,rn,ro,ru,rw,sd,sh,si,sk,sl,sn,so,sq,sr,sr-ME,st,su,sv,sw,ta,te,tg,th,ti,tk,tl,tn,to,tr,tt,tum,tw,ug,uk,ur,uz,vi,wo,xh,xx-bork,xx-elmer,xx-hacker,xx-klingon,xx-pirate,yi,yo,zh-CN,zh-TW,zu\', layout: google.translate.TranslateElement.InlineLayout}, \'google_translate_element\');\r\n                                setTimeout(function () {\r\n                                    $(\'select.goog-te-combo\').selectize({\r\n                                        onChange: function (val) {\r\n                                            $(\'select.goog-te-combo\').val(val);\r\n                                            fireEvent($(\'select.goog-te-combo\').get(0), \'change\');\r\n                                        }\r\n                                    });\r\n                                    // $(\'#google_translate_element .selectize-input > input\').attr(\'id\', \'google_translate_element\');\r\n                                    // $(\'#google_translate_element .selectize-input > input\').wrap("<label for=\'google_translate_element\'></label>");\r\n                                }, 1000);\r\n                            }\r\n\t\t\t\t\t\t\t</script>\r\n\t\t\t\t\t\t\t<script type="text/javascript" src="//translate.google.com/translate_a/element.js?cb=googleTranslateElementInit"></script>\r\n\t\t\t\t\t\t\t<span class=\'divide-line\'>|</span> <div class="open-wrap"><a class="modal-open" href="#">Text Size</a></div>\r\n\t\t\t\t\t\t</div>\r\n<!-- \t\t\t\t\t</div>\r\n -->\t\t\t\t</div>\r\n\t\t\t\t<div class="mob-hide col-desk-2 col-lg-3 rel">\r\n\t\t\t\t\t<div class="neat-search">\r\n                        <form action="/">\r\n                        \t<label for="search_2">Search the Site</label>\r\n    \t\t\t\t\t\t<input id="search_2" class="closed" type="text" name="s" placeholder="Search" />\r\n    \t\t\t\t\t\t<button type="submit" class="search-link t-indent">Search</button>\r\n                        </form>\r\n\t\t\t\t\t</div>\r\n\t\t\t\t</div>\r\n\t\t\t\t<div class="social col col-sm-desk-3 col-desk-2 rel mob-hide">\r\n\t\t\t\t\t<div class="share">\n\t<a href="https://www.facebook.com/MontgomeryParks" target="_blank">facebook<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1fb{fill:#547bbc}.cls-2{fill:#fff}</style></defs><title>social-fb</title><path class="cls-1fb" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M11.4 12.5H13V11a4 4 0 0 1 .5-2.3 2.8 2.8 0 0 1 2.4-1.1 9.7 9.7 0 0 1 2.8.3l-.4 2.3-1.3-.3a.9.9 0 0 0-1.1.8v1.8h2.5l-.2 2.2h-2.4v7.8h-2.9v-7.8h-1.5v-2.2z"/></svg></a>\n\t<a href="https://twitter.com/MontgomeryParks" target="_blank">twitter<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1tw{fill:#5ea8dd}.cls-2{fill:#fff}</style></defs><title>social-tw</title><path class="cls-1tw" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M22.5 11.1l-1.8.5a3.1 3.1 0 0 0 1.4-1.7 6.2 6.2 0 0 1-2 .7 3.1 3.1 0 0 0-5.3 2.8 8.8 8.8 0 0 1-6.4-3.2 3.1 3.1 0 0 0 1 4.1 3.1 3.1 0 0 1-1.4-.4 3.1 3.1 0 0 0 2.5 3H9.2a3.1 3.1 0 0 0 2.9 2.1 6.2 6.2 0 0 1-3.8 1.3h-.8A8.8 8.8 0 0 0 21 13.1v-.4a6.2 6.2 0 0 0 1.5-1.6"/></svg></a>\n\t<a href="https://www.instagram.com/montgomeryparks/" target="_blank">instagram<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 291.319 291.319"><path d="M145.659 0c80.44 0 145.66 65.219 145.66 145.66S226.1 291.319 145.66 291.319 0 226.1 0 145.66 65.21 0 145.659 0z" fill="#9b6954"/><path d="M195.93 63.708H95.38c-17.47 0-31.672 14.211-31.672 31.672v100.56c0 17.47 14.211 31.672 31.672 31.672h100.56c17.47 0 31.672-14.211 31.672-31.672V95.38c-.001-17.461-14.212-31.672-31.682-31.672zm9.978 18.326l3.587-.009v27.202l-27.402.091-.091-27.202 23.906-.082zm-60.248 36.205c22.732 0 27.42 21.339 27.42 27.429 0 15.103-12.308 27.411-27.42 27.411-15.121 0-27.42-12.308-27.42-27.411-.01-6.09 4.688-27.429 27.42-27.429zm63.99 75.716c0 8.658-7.037 15.704-15.713 15.704H97.073c-8.667 0-15.713-7.037-15.713-15.704v-66.539h22.759c-2.112 5.198-3.305 12.299-3.305 18.253 0 24.708 20.101 44.818 44.818 44.818s44.808-20.11 44.808-44.818c0-5.954-1.193-13.055-3.296-18.253h22.486v66.539h.02z" fill="#fff"/></svg></a>\n\t<a href="https://www.youtube.com/user/MontgomeryParks" target="_blank">youtube<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1yt{fill:#e9644a}.cls-2{fill:#fff}</style></defs><title>social-yt</title><path class="cls-1yt" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M17.9 15L13 17.9v-5.8zm4.6 2.7v-5.4a2.5 2.5 0 0 0-2.6-2.6h-9.8a2.5 2.5 0 0 0-2.6 2.6v5.4a2.5 2.5 0 0 0 2.6 2.6h9.8a2.5 2.5 0 0 0 2.6-2.6"/></svg></a>\n\t<a href="http://www.montgomeryplanningboard.org/rss/rss.shtm" target="_blank">rss<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1rss{fill:#f29100}.cls-2{fill:#fff}</style></defs><title>social-rss</title><path class="cls-1rss" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M10.9 19.1a2 2 0 0 1 0 2.8 2 2 0 1 1-2.8-2.8 2 2 0 0 1 2.8 0zm-3.4-6.5v2.9a7 7 0 0 1 7 7h2.9a9.9 9.9 0 0 0-9.9-9.9zm0-5.1v2.9a12.2 12.2 0 0 1 12.1 12.1h2.9a15 15 0 0 0-15-15z"/></svg></a>\n</div>\t\t\t\t</div>\r\n\r\n\t\t    </div>\r\n\t\t    <div class="nav-primary count-7">\r\n\t\t    \t<div class="icon-wrap rel">\r\n\t\t\t        <div class="ico-link">\r\n\t\t\t        \t<a href="/" class="ico ico-home" title="Back to Home">Back to Home<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 31.9 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-home</title><path class="cls-1" d="M26.7 14.7a.9.9 0 0 0-.9.9v12.6h-6.5v-5.1a.9.9 0 0 0-.9-.9h-4.7a.9.9 0 0 0-.9.9v5.1H6.2V16a.9.9 0 0 0-1.8 0v13.1a.9.9 0 0 0 .9.9h8.4a.9.9 0 0 0 .9-.9V24h2.9v5.1a.9.9 0 0 0 .9.9h8.3a.9.9 0 0 0 .9-.9V15.6a.9.9 0 0 0-.9-.9zm0 0l5 .7-7.3-7.3v-5a.9.9 0 0 0-.9-.9.9.9 0 0 0-.9.9v3.2l-6-6A.9.9 0 0 0 16 0a.9.9 0 0 0-.7.3l-15 15a.9.9 0 0 0 0 1.3.9.9 0 0 0 1.3 0L16 2.2l14.4 14.5a.9.9 0 0 0 .7.2l.7-.2a1 1 0 0 0 0-1.3zm0 0"/></svg></a>\r\n\t\t\t        </div>\r\n\t\t\t        <div class="ico-link">\r\n\t\t\t        \t<a href="/calendar/?v=1" class="ico ico-calendar" title="Calendar of Events">Calendar of Events<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-calendar</title><path class="cls-1" d="M27.5 2.8h-4.1V.9a.9.9 0 0 0-1.9 0v1.9h-5.6V.9a.9.9 0 0 0-1.9 0v1.9H8.4V.9a.9.9 0 1 0-1.8 0v1.9H2.5A2.5 2.5 0 0 0 0 5.3v22.2A2.5 2.5 0 0 0 2.5 30h25a2.5 2.5 0 0 0 2.5-2.5V5.3a2.5 2.5 0 0 0-2.5-2.5zm.6 24.7a.6.6 0 0 1-.6.6h-25a.6.6 0 0 1-.6-.6V5.3a.6.6 0 0 1 .6-.6h4.1v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h4.1a.6.6 0 0 1 .6.6v22.2zm0 0"/><path class="cls-1" d="M6.6 11.3h3.7v2.81H6.6zm0 4.6h3.7v2.81H6.6zm0 4.7h3.7v2.81H6.6zm6.5 0h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7zm6.6 9.3h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7z"/></svg></a>\r\n\t\t\t        </div>\r\n\t\t\t        <div class="ico-link">\r\n\t\t\t        \t<a href="/news/" class="ico ico-news" title="Latest News">Latest News<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30.7 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-news</title><path class="cls-1" d="M30 2.5a.8.8 0 0 0-.8.8v21.1a4 4 0 0 1-4 4H5.7a4 4 0 0 1-4-4V1.6h23.2v20.5a.8.8 0 1 0 1.6 0V.8a.8.8 0 0 0-.8-.8H.8a.8.8 0 0 0-.8.8v23.6A5.6 5.6 0 0 0 5.7 30h19.4a5.6 5.6 0 0 0 5.6-5.6V3.3a.8.8 0 0 0-.8-.8zm0 0"/><path class="cls-1" d="M21.9 6.2a.8.8 0 0 0 0-1.6H4.5a.8.8 0 1 0 0 1.6h17.4zm0 14.6H4.5a.8.8 0 1 0 0 1.6h17.4a.8.8 0 0 0 0-1.6zm0 4.1H7.3a.8.8 0 0 0 0 1.6h14.6a.8.8 0 0 0 0-1.6zM3.5 9v8.6a.8.8 0 0 0 .8.8h8.6a.8.8 0 0 0 .8-.8V9a.8.8 0 0 0-.8-.8H4.3a.8.8 0 0 0-.8.8zm1.6.8h7v7h-7v-7zM22.3 9h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 0-1.6zm0 3.6h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 1 0 0-1.6zm.8 4.6a.8.8 0 0 0-.8-.8h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 .8-.8zm0 0"/></svg></a>\r\n\t\t\t        </div>\r\n\t\t        </div>\r\n\t\t        <div class="menu-primary-container"><ul id="menu-primary" class="menu"><li id="menu-item-2177" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-2177"><a href="/activities/">Things To Do</a>\n<ul class="sub-menu">\n\t<li id="menu-item-664" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-664"><a href="/activities/">Activities</a></li>\n\t<li id="menu-item-8304" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-8304"><a href="/classes-camps/">Classes &#038; Camps</a></li>\n\t<li id="menu-item-950" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-950"><a href="/special-events/">Special Events &#038; Festivals</a></li>\n</ul>\n</li>\n<li id="menu-item-853" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-853"><a href="/parks-directory/">Parks &#038; Trails Directory</a>\n<ul class="sub-menu">\n\t<li id="menu-item-1089" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1089"><a href="/parks-trails/parks/">Parks</a></li>\n\t<li id="menu-item-39673" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-39673"><a href="https://www.montgomeryparks.org/activities/park-trails/">Trails</a></li>\n\t<li id="menu-item-1092" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1092"><a href="/parks-trails/gardens/">Gardens</a></li>\n\t<li id="menu-item-2941" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2941"><a href="/parks-trails/facilities/">Facilities</a></li>\n\t<li id="menu-item-1090" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1090"><a href="/parks-trails/nature-centers/">Nature Centers</a></li>\n\t<li id="menu-item-2360" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2360"><a href="/parks-trails/event-centers/">Event Centers</a></li>\n</ul>\n</li>\n<li id="menu-item-837" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-837"><a href="/services/">Permits, Rentals &#038; Requests</a>\n<ul class="sub-menu">\n\t<li id="menu-item-2850" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2850"><a href="/services/permits-rentals/">Permits &#038; Rentals</a></li>\n\t<li id="menu-item-838" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-838"><a href="/services/report-a-problem/">Parks Information &#038; Customer Service</a></li>\n</ul>\n</li>\n<li id="menu-item-839" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-839"><a href="/caring-for-our-parks/">Caring for Our Parks</a>\n<ul class="sub-menu">\n\t<li id="menu-item-840" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-840"><a href="/caring-for-our-parks/culture-heritage/">Culture &#038; Heritage</a></li>\n\t<li id="menu-item-842" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-842"><a href="/caring-for-our-parks/natural-spaces/">Natural Spaces</a></li>\n\t<li id="menu-item-959" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-959"><a href="/caring-for-our-parks/parks-and-trails/">Park and Trail Planning</a></li>\n\t<li id="menu-item-31505" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-31505"><a href="/caring-for-our-parks/sustainability/">Sustainability</a></li>\n\t<li id="menu-item-844" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-844"><a href="/caring-for-our-parks/wildlife/">Living with Wildlife</a></li>\n</ul>\n</li>\n<li id="menu-item-122" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-122"><a href="/projects/">Parks &#038; Projects</a>\n<ul class="sub-menu">\n\t<li id="menu-item-26054" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-26054"><a href="/projects/">Projects</a></li>\n\t<li id="menu-item-2503" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2503"><a href="/projects/capital-improvements-program/">Funding</a></li>\n\t<li id="menu-item-16631" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-16631"><a href="/projects/public-input/">Open Town Hall</a></li>\n\t<li id="menu-item-912" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-912"><a href="/projects/bids-procurement/">Bids / Procurement</a></li>\n\t<li id="menu-item-897" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-897"><a href="/projects/resources/">Resources</a></li>\n</ul>\n</li>\n<li id="menu-item-123" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-123"><a href="/support/">Support the Parks</a>\n<ul class="sub-menu">\n\t<li id="menu-item-19531" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-19531"><a href="/support/advocacy/">Advocacy</a></li>\n\t<li id="menu-item-907" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-907"><a href="/support/donate/">Donate</a></li>\n\t<li id="menu-item-15740" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-15740"><a href="/support/partnerships/">Partnerships</a></li>\n\t<li id="menu-item-900" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-900"><a href="/support/sponsorships/">Sponsorships</a></li>\n\t<li id="menu-item-901" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-901"><a href="/support/volunteer/">Volunteer</a></li>\n</ul>\n</li>\n<li id="menu-item-1047" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-1047"><a href="/about/parks/">About</a>\n<ul class="sub-menu">\n\t<li id="menu-item-660" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-660"><a href="/about/parks/">About the Parks</a></li>\n\t<li id="menu-item-659" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-659"><a href="/about/divisions/">Divisions &#038; Offices</a></li>\n\t<li id="menu-item-30188" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-30188"><a href="/about/media-center/">Media Center</a></li>\n\t<li id="menu-item-2592" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2592"><a href="/about/park-police/">Police</a></li>\n\t<li id="menu-item-655" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-655"><a href="/about/employment/">Employment</a></li>\n\t<li id="menu-item-932" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-932"><a href="/about/contact-us/">Contact Us</a></li>\n</ul>\n</li>\n</ul></div>\t\t    </div>\r\n\t\t</div>\r\n\t\t<div id="modal">\r\n\t    \t<div class="close">\xc3\x97</div>\r\n\t\t\t<div class="modal-wrapper">\r\n\t\t\t\t<div class="close">\xc3\x97</div>\r\n\t\t\t\tTo change the text size you can use your web browser\'s settings. Most browsers include functionality to let you increase or decrease the text on a web page. For example, to increase text size using:<br><br>\r\n\r\n\t\t\t\t<strong>Chrome</strong><br>\r\n\r\n\t\t\t\tIn the menu to the right of the address bar, select and set Zoom level. Menu > Zoom > +<br><br>\r\n\r\n\t\t\t\t<strong>Firefox</strong><br>\r\n\r\n\t\t\t\tIn the View menu, select Zoom. View > Zoom > Zoom In<br><br>\r\n\r\n\t\t\t\t<strong>Internet Explorer</strong><br>\r\n\r\n\t\t\t\tIn the View menu, select Text Size. View > Text Size > Largest<br><br>\r\n\r\n\t\t\t\t<strong>Safari</strong><br>\r\n\r\n\t\t\t\tIn the View menu, select Zoom In. View > Zoom In\r\n\t\t\t\tMacintosh Shortcut: Command+\r\n\t\t\t</div>\r\n\t\t</div>\r\n\t\t<div class="mobile-nav count-7">\r\n\t\t\t<div class="just-click"></div>\r\n\t\t\t<div class="upper-half">\r\n\t\t\t\t<div class="menu-primary-container"><ul id="menu-primary-1" class="menu"><li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-2177"><a href="/activities/">Things To Do</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-664"><a href="/activities/">Activities</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-8304"><a href="/classes-camps/">Classes &#038; Camps</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-950"><a href="/special-events/">Special Events &#038; Festivals</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-853"><a href="/parks-directory/">Parks &#038; Trails Directory</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1089"><a href="/parks-trails/parks/">Parks</a></li>\n\t<li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-39673"><a href="https://www.montgomeryparks.org/activities/park-trails/">Trails</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1092"><a href="/parks-trails/gardens/">Gardens</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2941"><a href="/parks-trails/facilities/">Facilities</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1090"><a href="/parks-trails/nature-centers/">Nature Centers</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2360"><a href="/parks-trails/event-centers/">Event Centers</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-837"><a href="/services/">Permits, Rentals &#038; Requests</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2850"><a href="/services/permits-rentals/">Permits &#038; Rentals</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-838"><a href="/services/report-a-problem/">Parks Information &#038; Customer Service</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-839"><a href="/caring-for-our-parks/">Caring for Our Parks</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-840"><a href="/caring-for-our-parks/culture-heritage/">Culture &#038; Heritage</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-842"><a href="/caring-for-our-parks/natural-spaces/">Natural Spaces</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-959"><a href="/caring-for-our-parks/parks-and-trails/">Park and Trail Planning</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-31505"><a href="/caring-for-our-parks/sustainability/">Sustainability</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-844"><a href="/caring-for-our-parks/wildlife/">Living with Wildlife</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-122"><a href="/projects/">Parks &#038; Projects</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-26054"><a href="/projects/">Projects</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2503"><a href="/projects/capital-improvements-program/">Funding</a></li>\n\t<li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-16631"><a href="/projects/public-input/">Open Town Hall</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-912"><a href="/projects/bids-procurement/">Bids / Procurement</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-897"><a href="/projects/resources/">Resources</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-123"><a href="/support/">Support the Parks</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-19531"><a href="/support/advocacy/">Advocacy</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-907"><a href="/support/donate/">Donate</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-15740"><a href="/support/partnerships/">Partnerships</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-900"><a href="/support/sponsorships/">Sponsorships</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-901"><a href="/support/volunteer/">Volunteer</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-1047"><a href="/about/parks/">About</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-660"><a href="/about/parks/">About the Parks</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-659"><a href="/about/divisions/">Divisions &#038; Offices</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-30188"><a href="/about/media-center/">Media Center</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2592"><a href="/about/park-police/">Police</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-655"><a href="/about/employment/">Employment</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-932"><a href="/about/contact-us/">Contact Us</a></li>\n</ul>\n</li>\n</ul></div>\t\t\t</div>\r\n\t\t\t<div class="lower-half rel">\r\n\t\t\t\t<div class="neat-search">\r\n\t\t\t\t\t<label for="m_search" />\r\n\t\t\t\t\t<input id="m_search" class="closed" type="text" placeholder="Search" />\r\n\t\t\t\t\t<a class="search-link t-indent" href="#m_search">Search</a>\r\n\t\t\t\t</div>\r\n\t\t\t\t<div class="page-nav">\r\n\t\t\t\t\t<a href="http://www.montgomeryparks.org/" target="_blank">Parks</a> &#8226;\r\n\t\t\t\t\t<a href="http://montgomeryplanning.org/" target="_blank">Planning</a> &#8226;\r\n\t\t\t\t\t<a href="http://montgomeryplanningboard.org/" target="_blank">Planning Board</a>\r\n\t\t\t\t</div>\r\n\t\t\t\t\t<div class="share">\n\t<a href="https://www.facebook.com/MontgomeryParks" target="_blank">facebook<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1fb{fill:#547bbc}.cls-2{fill:#fff}</style></defs><title>social-fb</title><path class="cls-1fb" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M11.4 12.5H13V11a4 4 0 0 1 .5-2.3 2.8 2.8 0 0 1 2.4-1.1 9.7 9.7 0 0 1 2.8.3l-.4 2.3-1.3-.3a.9.9 0 0 0-1.1.8v1.8h2.5l-.2 2.2h-2.4v7.8h-2.9v-7.8h-1.5v-2.2z"/></svg></a>\n\t<a href="https://twitter.com/MontgomeryParks" target="_blank">twitter<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1tw{fill:#5ea8dd}.cls-2{fill:#fff}</style></defs><title>social-tw</title><path class="cls-1tw" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M22.5 11.1l-1.8.5a3.1 3.1 0 0 0 1.4-1.7 6.2 6.2 0 0 1-2 .7 3.1 3.1 0 0 0-5.3 2.8 8.8 8.8 0 0 1-6.4-3.2 3.1 3.1 0 0 0 1 4.1 3.1 3.1 0 0 1-1.4-.4 3.1 3.1 0 0 0 2.5 3H9.2a3.1 3.1 0 0 0 2.9 2.1 6.2 6.2 0 0 1-3.8 1.3h-.8A8.8 8.8 0 0 0 21 13.1v-.4a6.2 6.2 0 0 0 1.5-1.6"/></svg></a>\n\t<a href="https://www.instagram.com/montgomeryparks/" target="_blank">instagram<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 291.319 291.319"><path d="M145.659 0c80.44 0 145.66 65.219 145.66 145.66S226.1 291.319 145.66 291.319 0 226.1 0 145.66 65.21 0 145.659 0z" fill="#9b6954"/><path d="M195.93 63.708H95.38c-17.47 0-31.672 14.211-31.672 31.672v100.56c0 17.47 14.211 31.672 31.672 31.672h100.56c17.47 0 31.672-14.211 31.672-31.672V95.38c-.001-17.461-14.212-31.672-31.682-31.672zm9.978 18.326l3.587-.009v27.202l-27.402.091-.091-27.202 23.906-.082zm-60.248 36.205c22.732 0 27.42 21.339 27.42 27.429 0 15.103-12.308 27.411-27.42 27.411-15.121 0-27.42-12.308-27.42-27.411-.01-6.09 4.688-27.429 27.42-27.429zm63.99 75.716c0 8.658-7.037 15.704-15.713 15.704H97.073c-8.667 0-15.713-7.037-15.713-15.704v-66.539h22.759c-2.112 5.198-3.305 12.299-3.305 18.253 0 24.708 20.101 44.818 44.818 44.818s44.808-20.11 44.808-44.818c0-5.954-1.193-13.055-3.296-18.253h22.486v66.539h.02z" fill="#fff"/></svg></a>\n\t<a href="https://www.youtube.com/user/MontgomeryParks" target="_blank">youtube<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1yt{fill:#e9644a}.cls-2{fill:#fff}</style></defs><title>social-yt</title><path class="cls-1yt" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M17.9 15L13 17.9v-5.8zm4.6 2.7v-5.4a2.5 2.5 0 0 0-2.6-2.6h-9.8a2.5 2.5 0 0 0-2.6 2.6v5.4a2.5 2.5 0 0 0 2.6 2.6h9.8a2.5 2.5 0 0 0 2.6-2.6"/></svg></a>\n\t<a href="http://www.montgomeryplanningboard.org/rss/rss.shtm" target="_blank">rss<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1rss{fill:#f29100}.cls-2{fill:#fff}</style></defs><title>social-rss</title><path class="cls-1rss" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M10.9 19.1a2 2 0 0 1 0 2.8 2 2 0 1 1-2.8-2.8 2 2 0 0 1 2.8 0zm-3.4-6.5v2.9a7 7 0 0 1 7 7h2.9a9.9 9.9 0 0 0-9.9-9.9zm0-5.1v2.9a12.2 12.2 0 0 1 12.1 12.1h2.9a15 15 0 0 0-15-15z"/></svg></a>\n</div>\t\t\t\t\t<a href="http://www.mncppc.org" title="Maryland-National Capital Parks and Commission" class="mncppc-link" target="_blank">Maryland-National Capital Park and Planning Commission</a>\r\n\t\t\t\t</div>\r\n\t\t\t</div>\r\n\t\t</div>\r\n\t\t<div id="mobile-footer">\r\n\t\t    <div class="footer-icon">\r\n\t\t    \t<a href="/" class="ico ico-home" title="Back to Home"><svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 31.9 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-home</title><path class="cls-1" d="M26.7 14.7a.9.9 0 0 0-.9.9v12.6h-6.5v-5.1a.9.9 0 0 0-.9-.9h-4.7a.9.9 0 0 0-.9.9v5.1H6.2V16a.9.9 0 0 0-1.8 0v13.1a.9.9 0 0 0 .9.9h8.4a.9.9 0 0 0 .9-.9V24h2.9v5.1a.9.9 0 0 0 .9.9h8.3a.9.9 0 0 0 .9-.9V15.6a.9.9 0 0 0-.9-.9zm0 0l5 .7-7.3-7.3v-5a.9.9 0 0 0-.9-.9.9.9 0 0 0-.9.9v3.2l-6-6A.9.9 0 0 0 16 0a.9.9 0 0 0-.7.3l-15 15a.9.9 0 0 0 0 1.3.9.9 0 0 0 1.3 0L16 2.2l14.4 14.5a.9.9 0 0 0 .7.2l.7-.2a1 1 0 0 0 0-1.3zm0 0"/></svg></a>\r\n\t\t    </div>\r\n\t\t    <div class="footer-icon">\r\n\t\t    <a href="/calendar" class="ico ico-calendar" title="Calendar"><svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-calendar</title><path class="cls-1" d="M27.5 2.8h-4.1V.9a.9.9 0 0 0-1.9 0v1.9h-5.6V.9a.9.9 0 0 0-1.9 0v1.9H8.4V.9a.9.9 0 1 0-1.8 0v1.9H2.5A2.5 2.5 0 0 0 0 5.3v22.2A2.5 2.5 0 0 0 2.5 30h25a2.5 2.5 0 0 0 2.5-2.5V5.3a2.5 2.5 0 0 0-2.5-2.5zm.6 24.7a.6.6 0 0 1-.6.6h-25a.6.6 0 0 1-.6-.6V5.3a.6.6 0 0 1 .6-.6h4.1v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h4.1a.6.6 0 0 1 .6.6v22.2zm0 0"/><path class="cls-1" d="M6.6 11.3h3.7v2.81H6.6zm0 4.6h3.7v2.81H6.6zm0 4.7h3.7v2.81H6.6zm6.5 0h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7zm6.6 9.3h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7z"/></svg></a></div>\r\n\t\t    <div class="footer-icon"><a href="/news" class="ico ico-news" title="News"><svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30.7 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-news</title><path class="cls-1" d="M30 2.5a.8.8 0 0 0-.8.8v21.1a4 4 0 0 1-4 4H5.7a4 4 0 0 1-4-4V1.6h23.2v20.5a.8.8 0 1 0 1.6 0V.8a.8.8 0 0 0-.8-.8H.8a.8.8 0 0 0-.8.8v23.6A5.6 5.6 0 0 0 5.7 30h19.4a5.6 5.6 0 0 0 5.6-5.6V3.3a.8.8 0 0 0-.8-.8zm0 0"/><path class="cls-1" d="M21.9 6.2a.8.8 0 0 0 0-1.6H4.5a.8.8 0 1 0 0 1.6h17.4zm0 14.6H4.5a.8.8 0 1 0 0 1.6h17.4a.8.8 0 0 0 0-1.6zm0 4.1H7.3a.8.8 0 0 0 0 1.6h14.6a.8.8 0 0 0 0-1.6zM3.5 9v8.6a.8.8 0 0 0 .8.8h8.6a.8.8 0 0 0 .8-.8V9a.8.8 0 0 0-.8-.8H4.3a.8.8 0 0 0-.8.8zm1.6.8h7v7h-7v-7zM22.3 9h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 0-1.6zm0 3.6h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 1 0 0-1.6zm.8 4.6a.8.8 0 0 0-.8-.8h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 .8-.8zm0 0"/></svg></a></div>\r\n\t\t    <div class="menu-open">\r\n\t\t    \t<div id="hamburger">\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t</div>\r\n\t\t\t</div>\r\n\t\t</div>\r\n\t\t<div id="siteBody" class="site-body">\r\n         <div class="section">\n             <div class="section-headline row rel">\n               <div class="container rel">\n                   <div class="section-head-wrap">\n                       <h1 class=\'section-head\'>Maple Sugaring Days</h1>\n                                                  Sun. February 3rd, 2019 10:00am to  11:00am                            \n                                                                                                            <div class="location-wrap">\n                                            <span class="before"></span> <a href="/parks-and-trails/brookside-nature-center/" class="location"><strong>Brookside Nature Center</strong></a>\n                                        </div>\n                                                                                                \n                    </div>\n                   <a class="back-to-calendar" href="/calendar/"><svg id="Layer_1" data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 21.1 30"><defs><style>.cls-1{fill:#258440}</style><linearGradient id="linear-gradient" x1="11" y1="30" x2="11" gradientUnits="userSpaceOnUse"><stop offset="0" stop-color="#79ba00"/><stop offset="1" stop-color="#18358B"/></linearGradient></defs><title>arrow-left</title><path id="arrow_left" data-name="arrow left" class="cls-1" d="M17.2 30a4.3 4.3 0 0 1-3-1.3L.5 15 14.2 1.3a4.3 4.3 0 1 1 6.1 6.1L12.6 15l7.7 7.7a4.3 4.3 0 0 1-3.1 7.3z" transform="translate(-.5)"/></svg>BACK TO CALENDAR</a>\n               </div>\n             </div>\n             <div class="rel fix-footer">\n                 <div class="row">\n                                          <div class="col col-9 main">\n                        \n                                                 <div class="p-y">\n                             <div class="p-y-b">\n                             <div class="linkem">\n                                 <div class="at-above-post addthis_tool" data-url="/events/maple-sugaring-days-147/"></div><div><a class="addthis_button" href="//addthis.com/bookmark.php?v=" /events/maple-sugaring-days-147/><img src="//cache.addthis.com/cachefly/static/btn/v2/lg-share-en.gif" width="125" height="16" alt="Bookmark and Share" style="border:0"/></a></div><p>February is Maple Sugaring Month at Brookside Nature Center. Every Saturday and Sunday you&#8217;ll have an opportunity to experience an American tradition: maple sugaring! Watch the whole maple sugaring process from start to finish. See sap drip from trees and taste it. Watch us boil it down into sweet maple syrup, then sample a tasty treat. Join in the fun and activities and learn something new at this family-friendly program! Space is limited so pre-registration is encouraged.</p>\n<p>Intended for ages 3&#038;up</p>\n<!-- AddThis Advanced Settings above via filter on the_content --><!-- AddThis Advanced Settings below via filter on the_content --><!-- AddThis Advanced Settings generic via filter on the_content --><!-- AddThis Share Buttons above via filter on the_content --><!-- AddThis Share Buttons below via filter on the_content --><div class="at-below-post addthis_tool" data-url="/events/maple-sugaring-days-147/"></div><!-- AddThis Share Buttons generic via filter on the_content -->                                                                     <p class="inline-list">\n        \t\t\t\t\t\t\t\tThis event will be held at\n        \t\t\t\t\t\t\t            \t\t\t\t\t\t\t                \t\t\t\t\t\t\t            <a href="/parks-and-trails/brookside-nature-center/">Brookside Nature Center</a>\n        \t\t\t\t\t\t\t                    \t\t\t\t\t\t\t                                        </p>\n    \t\t\t\t\t\t\t\n                                                                  \t<br>\n                                 \t<a href="http://apm.activecommunities.com/montgomerycounty/Activity_Search/53395" target="_blank" class="btn">\n                                        Register                                    </a>\n                                                                  </div>\n                                 <div class="row">\n</div>\n                                 <div class="home-alert-wrapper">\n    </div>\n                             </div>\n                         </div>\n                     </div>\n                     <div class="col col-3">\n                         <div class="sb-wrap">\n                             <div class="sidebar">\n                                 <aside class="info-box">\n\n                                                                      <dl>\n                                                                             <dt>Dates</dt>\n                                                                            <dd>\n                                            <p>\n                                                02/03/2019                                             </p>\n                                        </dd>\n                                     </dl>\n                                                                        <dl>\n                                        <dt>Time</dt>\n                                        <dd>\n                                            <p>\n                                                Sun. February 3rd, 2019 10:00am to  11:00am                                            </p>\n                                        </dd>\n                                     </dl>\n                                                                         <dl>\n                                        <dt>Location</dt>\n                                        <dd class="inline-list">\n\n                                                                                                                <a href="/parks-and-trails/brookside-nature-center/" class="a-theme">Brookside Nature Center</a></strong>\n                                        \n                                        </dd>\n                                    </dl>\n                                                                                                                                                <dl>\n                                        <dt>Fee</dt>\n                                        <dd>\n\n                                        <p>$7</p>\n\n                                        </dd>\n                                    </dl>\n                                                                        <dl>\n                                        \n                                    </dl>\n                                                                        <dl class="sync-cal-container p-y"\n                                        data-title="Maple Sugaring Days"\n                                        data-start="February 3, 2019 10:00"\n                                        data-end="February 3, 2019 11:00"\n                                        data-address=""\n                                        data-desc=""\n                                    >\n                                        <dt>\n                                            <div class="sync-cal"></div>\n                                        </dt>\n                                    </dl>\n                                </aside>\n                                <aside class="action-box">\n    <h2>Sign up for Updates</h2>\n    <p></p>\n    <div class="row">\n        \n                <div class=\'gf_browser_unknown gform_wrapper\' id=\'gform_wrapper_5\' ><form method=\'post\' enctype=\'multipart/form-data\'  id=\'gform_5\'  action=\'/events/maple-sugaring-days-147/\'>\n                        <div class=\'gform_body\'><ul id=\'gform_fields_5\' class=\'gform_fields top_label form_sublabel_below description_below\'><li id=\'field_5_1\'  class=\'gfield gfield_contains_required field_sublabel_below field_description_below gfield_visibility_visible\' ><label class=\'gfield_label gfield_label_before_complex\'  >Name<span class=\'gfield_required\'>*</span></label><div class=\'ginput_complex ginput_container no_prefix has_first_name no_middle_name has_last_name no_suffix gf_name_has_2 ginput_container_name\' id=\'input_5_1\'>\n                            \n                            <span id=\'input_5_1_3_container\' class=\'name_first\' >\n                                                    <input type=\'text\' name=\'input_1.3\' id=\'input_5_1_3\' value=\'\' aria-label=\'First name\'   aria-required="true" aria-invalid="false" />\n                                                    <label for=\'input_5_1_3\' >First</label>\n                                                </span>\n                            \n                            <span id=\'input_5_1_6_container\' class=\'name_last\' >\n                                                    <input type=\'text\' name=\'input_1.6\' id=\'input_5_1_6\' value=\'\' aria-label=\'Last name\'   aria-required="true" aria-invalid="false" />\n                                                    <label for=\'input_5_1_6\' >Last</label>\n                                                </span>\n                            \n                        </div></li><li id=\'field_5_2\'  class=\'gfield gfield_contains_required field_sublabel_below field_description_below gfield_visibility_visible\' ><label class=\'gfield_label\' for=\'input_5_2\' >Email<span class=\'gfield_required\'>*</span></label><div class=\'ginput_container ginput_container_email\'>\n                            <input name=\'input_2\' id=\'input_5_2\' type=\'text\' value=\'\' class=\'medium\'    aria-required="true" aria-invalid="false"/>\n                        </div></li><li id=\'field_5_4\'  class=\'gfield gfield_contains_required field_sublabel_below field_description_below gfield_visibility_visible\' ><label class=\'gfield_label\'  >What type of information would you like to receive?<span class=\'gfield_required\'>*</span></label><div class=\'ginput_container ginput_container_checkbox\'><ul class=\'gfield_checkbox\' id=\'input_5_4\'><li class=\'gchoice_5_4_1\'>\n\t\t\t\t\t\t\t\t<input name=\'input_4.1\' type=\'checkbox\'  value=\'General News\'  id=\'choice_5_4_1\'   />\n\t\t\t\t\t\t\t\t<label for=\'choice_5_4_1\' id=\'label_5_4_1\'>General News</label>\n\t\t\t\t\t\t\t</li><li class=\'gchoice_5_4_2\'>\n\t\t\t\t\t\t\t\t<input name=\'input_4.2\' type=\'checkbox\'  value=\'Nature Centers &amp; Gardens\'  id=\'choice_5_4_2\'   />\n\t\t\t\t\t\t\t\t<label for=\'choice_5_4_2\' id=\'label_5_4_2\'>Nature Centers & Gardens</label>\n\t\t\t\t\t\t\t</li><li class=\'gchoice_5_4_3\'>\n\t\t\t\t\t\t\t\t<input name=\'input_4.3\' type=\'checkbox\'  value=\'Sporting &amp; Outdoors (Tennis, Ice, Golf, Archery, Field Sports, Camping)\'  id=\'choice_5_4_3\'   />\n\t\t\t\t\t\t\t\t<label for=\'choice_5_4_3\' id=\'label_5_4_3\'>Sporting & Outdoors (Tennis, Ice, Golf, Archery, Field Sports, Camping)</label>\n\t\t\t\t\t\t\t</li><li class=\'gchoice_5_4_4\'>\n\t\t\t\t\t\t\t\t<input name=\'input_4.4\' type=\'checkbox\'  value=\'History &amp; Archaeology Programs\'  id=\'choice_5_4_4\'   />\n\t\t\t\t\t\t\t\t<label for=\'choice_5_4_4\' id=\'label_5_4_4\'>History & Archaeology Programs</label>\n\t\t\t\t\t\t\t</li></ul></div><div class=\'gfield_description\'>Select any that apply. General news is delivered monthly while specialized information is delivered quarterly.</div></li><li id=\'field_5_5\'  class=\'gfield gfield_contains_required field_sublabel_below field_description_below gfield_visibility_visible\' ><label class=\'gfield_label\'  >Please check this box to confirm your opt-in to receiving newsletters from Montgomery Parks. Thank you.<span class=\'gfield_required\'>*</span></label><div class=\'ginput_container ginput_container_checkbox\'><ul class=\'gfield_checkbox\' id=\'input_5_5\'><li class=\'gchoice_5_5_1\'>\n\t\t\t\t\t\t\t\t<input name=\'input_5.1\' type=\'checkbox\'  value=\'Agreeing to opt-in\'  id=\'choice_5_5_1\'   />\n\t\t\t\t\t\t\t\t<label for=\'choice_5_5_1\' id=\'label_5_5_1\'>Agreeing to opt-in</label>\n\t\t\t\t\t\t\t</li></ul></div></li>\n                            </ul></div>\n        <div class=\'gform_footer top_label\'> <input type=\'submit\' id=\'gform_submit_button_5\' class=\'gform_button button\' value=\'Submit\'  onclick=\'if(window["gf_submitting_5"]){return false;}  window["gf_submitting_5"]=true;  \' onkeypress=\'if( event.keyCode == 13 ){ if(window["gf_submitting_5"]){return false;} window["gf_submitting_5"]=true;  jQuery("#gform_5").trigger("submit",[true]); }\' /> \n            <input type=\'hidden\' class=\'gform_hidden\' name=\'is_submit_5\' value=\'1\' />\n            <input type=\'hidden\' class=\'gform_hidden\' name=\'gform_submit\' value=\'5\' />\n            \n            <input type=\'hidden\' class=\'gform_hidden\' name=\'gform_unique_id\' value=\'\' />\n            <input type=\'hidden\' class=\'gform_hidden\' name=\'state_5\' value=\'WyJbXSIsIjIxMzlhM2M3YjI5NmY4NjM5MzgxYzAwYzQ0ZjFlOGZhIl0=\' />\n            <input type=\'hidden\' class=\'gform_hidden\' name=\'gform_target_page_number_5\' id=\'gform_target_page_number_5\' value=\'0\' />\n            <input type=\'hidden\' class=\'gform_hidden\' name=\'gform_source_page_number_5\' id=\'gform_source_page_number_5\' value=\'1\' />\n            <input type=\'hidden\' name=\'gform_field_values\' value=\'\' />\n            \n        </div>\n                        </form>\n                        </div><script type=\'text/javascript\'> jQuery(document).bind(\'gform_post_render\', function(event, formId, currentPage){if(formId == 5) {} } );jQuery(document).bind(\'gform_post_conditional_logic\', function(event, formId, fields, isInit){} );</script><script type=\'text/javascript\'> jQuery(document).ready(function(){jQuery(document).trigger(\'gform_post_render\', [5, 1]) } ); </script>    </div>\n    </aside>\n                                                             </div>\n                         </div>\n                     </div>\n                 </div>\n             </div>\n         </div>\n     \n <div id="footer" class="main-footer">\r\n        <section class="main-footer-bottom row">\r\n        <div class="col col-mob-12 col-desk-3">\r\n            <div class="address">\r\n                Montgomery County<br>Department of Parks<br>\r\n                9500 Brunett Ave.<br>\r\n                Silver Spring, MD 20901\r\n            </div>\r\n        </div>\r\n                        <div class="col col-mob-6 col-sm-desk-4 col-desk-2 footer-nav">\r\n                                            <a href="/about/contact-us/">Contact Us</a>\r\n                                                <a href="/about/media-center/press-contacts/">Press</a>\r\n                                                <a href="/about/parks/accessibility/">ADA Accessibility</a>\r\n                                                <a href="/terms/">Terms & Privacy</a>\r\n                                        </div>\r\n                                <div class="col col-mob-6 col-sm-desk-4 col-desk-2 footer-nav">\r\n                                            <a href="/about/park-police/">Park Police</a>\r\n                                                <a href="/about/rules-and-regulations/">Rules and Regulations</a>\r\n                                                <a href="/support/volunteer/">Volunteer</a>\r\n                                                <a href="/activities/">Activities</a>\r\n                                        </div>\r\n                                <div class="col col-mob-6 col-sm-desk-4 col-desk-2 footer-nav">\r\n                                            <a href="/permits/">Permits</a>\r\n                                                <a href="/calendar/">Calendar</a>\r\n                                                <a href="/news/">News</a>\r\n                                                <a href="/weather-alerts/">Weather Alerts</a>\r\n                                        </div>\r\n                        <div class="col col-sm-desk-3 pull-right mob-hide rel">\r\n            SOCIAL<br>\r\n            <div class="share">\n\t<a href="https://www.facebook.com/MontgomeryParks" target="_blank">facebook<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1fb{fill:#547bbc}.cls-2{fill:#fff}</style></defs><title>social-fb</title><path class="cls-1fb" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M11.4 12.5H13V11a4 4 0 0 1 .5-2.3 2.8 2.8 0 0 1 2.4-1.1 9.7 9.7 0 0 1 2.8.3l-.4 2.3-1.3-.3a.9.9 0 0 0-1.1.8v1.8h2.5l-.2 2.2h-2.4v7.8h-2.9v-7.8h-1.5v-2.2z"/></svg></a>\n\t<a href="https://twitter.com/MontgomeryParks" target="_blank">twitter<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1tw{fill:#5ea8dd}.cls-2{fill:#fff}</style></defs><title>social-tw</title><path class="cls-1tw" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M22.5 11.1l-1.8.5a3.1 3.1 0 0 0 1.4-1.7 6.2 6.2 0 0 1-2 .7 3.1 3.1 0 0 0-5.3 2.8 8.8 8.8 0 0 1-6.4-3.2 3.1 3.1 0 0 0 1 4.1 3.1 3.1 0 0 1-1.4-.4 3.1 3.1 0 0 0 2.5 3H9.2a3.1 3.1 0 0 0 2.9 2.1 6.2 6.2 0 0 1-3.8 1.3h-.8A8.8 8.8 0 0 0 21 13.1v-.4a6.2 6.2 0 0 0 1.5-1.6"/></svg></a>\n\t<a href="https://www.instagram.com/montgomeryparks/" target="_blank">instagram<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 291.319 291.319"><path d="M145.659 0c80.44 0 145.66 65.219 145.66 145.66S226.1 291.319 145.66 291.319 0 226.1 0 145.66 65.21 0 145.659 0z" fill="#9b6954"/><path d="M195.93 63.708H95.38c-17.47 0-31.672 14.211-31.672 31.672v100.56c0 17.47 14.211 31.672 31.672 31.672h100.56c17.47 0 31.672-14.211 31.672-31.672V95.38c-.001-17.461-14.212-31.672-31.682-31.672zm9.978 18.326l3.587-.009v27.202l-27.402.091-.091-27.202 23.906-.082zm-60.248 36.205c22.732 0 27.42 21.339 27.42 27.429 0 15.103-12.308 27.411-27.42 27.411-15.121 0-27.42-12.308-27.42-27.411-.01-6.09 4.688-27.429 27.42-27.429zm63.99 75.716c0 8.658-7.037 15.704-15.713 15.704H97.073c-8.667 0-15.713-7.037-15.713-15.704v-66.539h22.759c-2.112 5.198-3.305 12.299-3.305 18.253 0 24.708 20.101 44.818 44.818 44.818s44.808-20.11 44.808-44.818c0-5.954-1.193-13.055-3.296-18.253h22.486v66.539h.02z" fill="#fff"/></svg></a>\n\t<a href="https://www.youtube.com/user/MontgomeryParks" target="_blank">youtube<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1yt{fill:#e9644a}.cls-2{fill:#fff}</style></defs><title>social-yt</title><path class="cls-1yt" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M17.9 15L13 17.9v-5.8zm4.6 2.7v-5.4a2.5 2.5 0 0 0-2.6-2.6h-9.8a2.5 2.5 0 0 0-2.6 2.6v5.4a2.5 2.5 0 0 0 2.6 2.6h9.8a2.5 2.5 0 0 0 2.6-2.6"/></svg></a>\n\t<a href="http://www.montgomeryplanningboard.org/rss/rss.shtm" target="_blank">rss<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1rss{fill:#f29100}.cls-2{fill:#fff}</style></defs><title>social-rss</title><path class="cls-1rss" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M10.9 19.1a2 2 0 0 1 0 2.8 2 2 0 1 1-2.8-2.8 2 2 0 0 1 2.8 0zm-3.4-6.5v2.9a7 7 0 0 1 7 7h2.9a9.9 9.9 0 0 0-9.9-9.9zm0-5.1v2.9a12.2 12.2 0 0 1 12.1 12.1h2.9a15 15 0 0 0-15-15z"/></svg></a>\n</div>        </div>\r\n    </section>\r\n</div>\r\n        <style>\n    .brink-gdpr-cookie-middle{\n      border-right:1px solid rgba(255,255,255,.5);\n    }\n\n    .brink-gdpr-cookie-left a{\n      color:#fff;\n    }\n\n    #brink-gdpr-cookie-accept{\n      background:rgb(0,0,0);\n      color:#fff;\n    }\n\n    #brink-gdpr-cookie-accept:hover, #brink-gdpr-cookie-accept:active{\n      background:rgb(255,255,255);\n      color:#000;\n    }\n\n    .brink-gdpr-cookie-right span:before{\n      border-top:1px solid #fff;\n      border-right:1px solid #fff;\n    }\n\n    .brink-gdpr-cookie-right span:after{\n      border-top:1px solid #fff;\n      border-left:1px solid #fff;\n    }\n    </style>\n    <div id="brink-gdpr-cookie-notice" class="brink-gdpr-cookie-notice bottom" style="background:rgba(102,102,102,.95); color:#fff">\n      <div class="brink-gdpr-container">\n        <div class="brink-gdpr-cookie-left">\n          <span class="h2">Cookies Notice</span>\n            <p>\n              This website uses cookies to provide you with an improved and personalized experience. By using this site you agree to our use of cookies. Please read our <a href="/terms/" target="_blank">cookies policy</a> for more information on the cookies we use and how to delete or block them.            </p>\n          </span>\n        </div>\n        <div class="brink-gdpr-cookie-middle">\n          <a href="#" id="brink-gdpr-cookie-accept">I Accept</a>\n        </div>\n        <div class="brink-gdpr-cookie-right">\n          <a href="#" id="brink-gdpr-cookie-close"></a>\n          <span></span>\n        </div>\n      </div>\n    </div>\n    <script type="text/javascript">\n      window.brink_gdpr_whitelist = [];\n    </script>\n    <script data-cfasync="false" type="text/javascript">if (window.addthis_product === undefined) { window.addthis_product = "wpp"; } if (window.wp_product_version === undefined) { window.wp_product_version = "wpp-6.1.8"; } if (window.wp_blog_version === undefined) { window.wp_blog_version = "5.0.3"; } if (window.addthis_share === undefined) { window.addthis_share = {}; } if (window.addthis_config === undefined) { window.addthis_config = {"data_track_clickback":true,"ui_atversion":"300"}; } if (window.addthis_plugin_info === undefined) { window.addthis_plugin_info = {"info_status":"enabled","cms_name":"WordPress","plugin_name":"Share Buttons by AddThis","plugin_version":"6.1.8","plugin_mode":"AddThis","anonymous_profile_id":"wp-f82eef1f055e09dcb6273ac5462e3333","page_info":{"template":"posts","post_type":"events"},"sharing_enabled_on_post_via_metabox":false}; } \n                    (function() {\n                      var first_load_interval_id = setInterval(function () {\n                        if (typeof window.addthis !== \'undefined\') {\n                          window.clearInterval(first_load_interval_id);\n                          if (typeof window.addthis_layers !== \'undefined\' && Object.getOwnPropertyNames(window.addthis_layers).length > 0) {\n                            window.addthis.layers(window.addthis_layers);\n                          }\n                          if (Array.isArray(window.addthis_layers_tools)) {\n                            for (i = 0; i < window.addthis_layers_tools.length; i++) {\n                              window.addthis.layers(window.addthis_layers_tools[i]);\n                            }\n                          }\n                        }\n                     },1000)\n                    }());\n                </script> <script data-cfasync="false" type="text/javascript" src="https://s7.addthis.com/js/300/addthis_widget.js#pubid=ra-5a0206a251dc7e59" async="async"></script><script type=\'text/javascript\'>\n/* <![CDATA[ */\nvar countVars = {"disqusShortname":"montgomeryparks"};\n/* ]]> */\n</script>\n<script type=\'text/javascript\' src=\'/wp-content/plugins/disqus-comment-system/public/js/comment_count.js?ver=3.0.16\'></script>\n<script type=\'text/javascript\'>\n/* <![CDATA[ */\nvar embedVars = {"disqusConfig":{"integration":"wordpress 3.0.16"},"disqusIdentifier":"41896 https:\\/\\/www.montgomeryparks.org?p=41896","disqusShortname":"montgomeryparks","disqusTitle":"Maple Sugaring Days","disqusUrl":"http:\\/\\/staging.montgomeryparks.org\\/events\\/maple-sugaring-days-147\\/","postId":"41896"};\n/* ]]> */\n</script>\n<script type=\'text/javascript\' src=\'/wp-content/plugins/disqus-comment-system/public/js/comment_embed.js?ver=3.0.16\'></script>\n<script type=\'text/javascript\' src=\'/wp-content/themes/mdparks/dist/js/main.js?ver=659\'></script>\n<script type=\'text/javascript\' src=\'https://js.arcgis.com/3.18/?ver=5.0.3\'></script>\n<script type=\'text/javascript\' src=\'/wp-includes/js/wp-embed.min.js?ver=5.0.3\'></script>\n<script type=\'text/javascript\' src=\'/wp-includes/js/jquery/jquery.js?ver=1.12.4\'></script>\n<script type=\'text/javascript\' src=\'/wp-includes/js/jquery/jquery-migrate.min.js?ver=1.4.1\'></script>\n<link rel=\'stylesheet\' id=\'gforms_reset_css-group-css\' href=\'/wp-content/plugins/bwp-minify/min/?f=wp-content/plugins/gravityforms/css/formreset.min.css,wp-content/plugins/gravityforms/css/formsmain.min.css,wp-content/plugins/gravityforms/css/readyclass.min.css,wp-content/plugins/gravityforms/css/browsers.min.css\' type=\'text/css\' media=\'all\' />\r\n<script>\r\n  window.addEventListener(\'load\',function(){\r\n    jQuery(\'[href*=".pdf"]\').click(function(){\r\n      __gaTracker(\'send\',\'event\',\'PDF\',\'click\',jQuery(this).attr(\'href\'));\r\n    })\r\n  });\r\n</script>\r\n<img height="1" width="1" style="border-style:none;" alt="" src="//googleads.g.doubleclick.net/pagead/viewthroughconversion/974628909/?guid=ON&amp;script=0"/>\r\n\r\n</body>\r\n\r\n</html>\r\n<!-- AWS -->\r\n'
@@ -10,7 +65,10 @@ calendar_page_no_events_content = b'<!DOCTYPE html>\r\n<html lang="en-US" xmlns:
 
 calendar_page_next_page_content = b'<!DOCTYPE html>\r\n<html lang="en-US" xmlns:fb="https://www.facebook.com/2008/fbml" xmlns:addthis="https://www.addthis.com/help/api-spec"  prefix="og: http://ogp.me/ns#" class="no-js">\r\n\t<head>\r\n\t\t<meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">\r\n\t\t<script src="https://use.typekit.net/bxi2fgb.js"></script>\r\n\t\t<script>try{Typekit.load({ async: true });}catch(e){}</script>\r\n\t\t<meta charset="UTF-8">\r\n\t\t<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=12.0, minimum-scale=.25, user-scalable=yes"/>\r\n\t\t<!-- CHECK META CONTENT -->\r\n\t\t<title>Calendar - Montgomery Parks</title>\r\n\r\n        <script>\r\n          // helpful for understanding dojoConfig.packages vs. dojoConfig.paths:\r\n          // http://www.sitepen.com/blog/2013/06/20/dojo-faq-what-is-the-difference-packages-vs-paths-vs-aliases/\r\n          var dojoConfig = {\r\n            paths: {\r\n              extras: "/wp-content/themes/mdparks/dist/js/extras"\r\n            }\r\n          };\r\n        </script>\r\n\r\n\t\t<link rel="pingback" href="/xmlrpc.php">\r\n\t\t\t\t<style>#skipNav{display: block;background-color:#fff;padding:1em;position: absolute;top: 0;left: -9999px;}#skipNav:focus{left:0;}</style>\r\n\r\n\t\t\n<!-- This site is optimized with the Yoast SEO plugin v9.5 - https://yoast.com/wordpress/plugins/seo/ -->\n<link rel="canonical" href="/calendar/" />\n<meta property="og:locale" content="en_US" />\n<meta property="og:type" content="article" />\n<meta property="og:title" content="Calendar - Montgomery Parks" />\n<meta property="og:url" content="/calendar/" />\n<meta property="og:site_name" content="Montgomery Parks" />\n<meta name="twitter:card" content="summary" />\n<meta name="twitter:title" content="Calendar - Montgomery Parks" />\n<!-- / Yoast SEO plugin. -->\n\n<link rel=\'dns-prefetch\' href=\'//js.arcgis.com\' />\n<link rel=\'dns-prefetch\' href=\'//s.w.org\' />\n<link rel=\'dns-prefetch\' href=\'//s3.amazonaws.com\' />\n<link rel="alternate" type="application/rss+xml" title="Montgomery Parks &raquo; Feed" href="/feed/" />\n<link rel="alternate" type="application/rss+xml" title="Montgomery Parks &raquo; Comments Feed" href="/comments/feed/" />\n<!-- This site uses the Google Analytics by MonsterInsights plugin v7.4.2 - Using Analytics tracking - https://www.monsterinsights.com/ -->\r\n<script type="text/javascript" data-cfasync="false">\r\n\tvar mi_version         = \'7.4.2\';\r\n\tvar mi_track_user      = true;\r\n\tvar mi_no_track_reason = \'\';\r\n\t\r\n\tvar disableStr = \'ga-disable-UA-493554-2\';\r\n\r\n\t/* Function to detect opted out users */\r\n\tfunction __gaTrackerIsOptedOut() {\r\n\t\treturn document.cookie.indexOf(disableStr + \'=true\') > -1;\r\n\t}\r\n\r\n\t/* Disable tracking if the opt-out cookie exists. */\r\n\tif ( __gaTrackerIsOptedOut() ) {\r\n\t\twindow[disableStr] = true;\r\n\t}\r\n\r\n\t/* Opt-out function */\r\n\tfunction __gaTrackerOptout() {\r\n\t  document.cookie = disableStr + \'=true; expires=Thu, 31 Dec 2099 23:59:59 UTC; path=/\';\r\n\t  window[disableStr] = true;\r\n\t}\r\n\t\r\n\tif ( mi_track_user ) {\r\n\t\t(function(i,s,o,g,r,a,m){i[\'GoogleAnalyticsObject\']=r;i[r]=i[r]||function(){\r\n\t\t\t(i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),\r\n\t\t\tm=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)\r\n\t\t})(window,document,\'script\',\'//www.google-analytics.com/analytics.js\',\'__gaTracker\');\r\n\r\n\t\t__gaTracker(\'create\', \'UA-493554-2\', \'auto\');\n\t\t__gaTracker(\'set\', \'forceSSL\', true);\n\t\t__gaTracker(\'set\', \'anonymizeIp\', true);\n\t\t__gaTracker(\'require\', \'linkid\', \'linkid.js\');\n\t\t__gaTracker(\'send\',\'pageview\');\n\t} else {\r\n\t\tconsole.log( "" );\r\n\t\t(function() {\r\n\t\t\t/* https://developers.google.com/analytics/devguides/collection/analyticsjs/ */\r\n\t\t\tvar noopfn = function() {\r\n\t\t\t\treturn null;\r\n\t\t\t};\r\n\t\t\tvar noopnullfn = function() {\r\n\t\t\t\treturn null;\r\n\t\t\t};\r\n\t\t\tvar Tracker = function() {\r\n\t\t\t\treturn null;\r\n\t\t\t};\r\n\t\t\tvar p = Tracker.prototype;\r\n\t\t\tp.get = noopfn;\r\n\t\t\tp.set = noopfn;\r\n\t\t\tp.send = noopfn;\r\n\t\t\tvar __gaTracker = function() {\r\n\t\t\t\tvar len = arguments.length;\r\n\t\t\t\tif ( len === 0 ) {\r\n\t\t\t\t\treturn;\r\n\t\t\t\t}\r\n\t\t\t\tvar f = arguments[len-1];\r\n\t\t\t\tif ( typeof f !== \'object\' || f === null || typeof f.hitCallback !== \'function\' ) {\r\n\t\t\t\t\tconsole.log( \'Not running function __gaTracker(\' + arguments[0] + " ....) because you are not being tracked. " + mi_no_track_reason );\r\n\t\t\t\t\treturn;\r\n\t\t\t\t}\r\n\t\t\t\ttry {\r\n\t\t\t\t\tf.hitCallback();\r\n\t\t\t\t} catch (ex) {\r\n\r\n\t\t\t\t}\r\n\t\t\t};\r\n\t\t\t__gaTracker.create = function() {\r\n\t\t\t\treturn new Tracker();\r\n\t\t\t};\r\n\t\t\t__gaTracker.getByName = noopnullfn;\r\n\t\t\t__gaTracker.getAll = function() {\r\n\t\t\t\treturn [];\r\n\t\t\t};\r\n\t\t\t__gaTracker.remove = noopfn;\r\n\t\t\twindow[\'__gaTracker\'] = __gaTracker;\r\n\t\t\t\t\t})();\r\n\t\t}\r\n</script>\r\n<!-- / Google Analytics by MonsterInsights -->\r\n<link rel=\'stylesheet\' id=\'wp-block-library-group-css\' href=\'/wp-content/plugins/bwp-minify/min/?f=wp-includes/css/dist/block-library/style.min.css,wp-content/themes/mdparks/dist/css/main.min.css,wp-content/plugins/addthis/frontend/build/addthis_wordpress_public.min.css\' type=\'text/css\' media=\'all\' />\r\n<link rel=\'stylesheet\' id=\'brink_gdpr_css-group-css\' href=\'/wp-content/plugins/bwp-minify/min/?f=wp-content/plugins/brink-gdpr/css/brink-gdpr.css\' type=\'text/css\' media=\'screen\' />\r\n<link rel=\'stylesheet\' id=\'esri-css-css\'  href=\'https://js.arcgis.com/3.18/esri/css/esri.css?ver=5.0.3\' type=\'text/css\' media=\'all\' />\n<script type=\'text/javascript\'>\n/* <![CDATA[ */\nvar monsterinsights_frontend = {"js_events_tracking":"true","download_extensions":"doc,exe,js,pdf,ppt,tgz,zip,xls","inbound_paths":"[]","home_url":"http:\\/\\/staging.montgomeryparks.org","hash_tracking":"false"};\n/* ]]> */\n</script>\n<script type=\'text/javascript\' src=\'/wp-content/plugins/google-analytics-for-wordpress/assets/js/frontend.min.js?ver=7.4.2\'></script>\n<script type=\'text/javascript\' src=\'/wp-content/plugins/brink-gdpr/js/brink_gdpr.js?ver=5.0.3\'></script>\n<link rel=\'https://api.w.org/\' href=\'/wp-json/\' />\n<link rel="EditURI" type="application/rsd+xml" title="RSD" href="/xmlrpc.php?rsd" />\n<link rel="wlwmanifest" type="application/wlwmanifest+xml" href="/wp-includes/wlwmanifest.xml" /> \n<meta name="generator" content="WordPress 5.0.3" />\n<link rel=\'shortlink\' href=\'/?p=32\' />\n<link rel="alternate" type="application/json+oembed" href="/wp-json/oembed/1.0/embed?url=http%3A%2F%2Fstaging.montgomeryparks.org%2Fcalendar%2F" />\n<link rel="alternate" type="text/xml+oembed" href="/wp-json/oembed/1.0/embed?url=http%3A%2F%2Fstaging.montgomeryparks.org%2Fcalendar%2F&#038;format=xml" />\n<script type="text/javascript">\nvar ajaxurl = \'/wp-admin/admin-ajax.php/\';\n</script>\n\t\t<style type="text/css">.recentcomments a{display:inline !important;padding:0 !important;margin:0 !important;}</style>\n\t\t<link rel="icon" href="/uploads/2016/07/cropped-favicon-490x490.png" sizes="32x32" />\n<link rel="icon" href="/uploads/2016/07/cropped-favicon-490x490.png" sizes="192x192" />\n<link rel="apple-touch-icon-precomposed" href="/uploads/2016/07/cropped-favicon-490x490.png" />\n<meta name="msapplication-TileImage" content="/uploads/2016/07/cropped-favicon-490x490.png" />\n\r\n\t\t\t</head>\r\n\r\n\t<body class="page-template page-template-page-calendar page-template-page-calendar-php page page-id-32 group-blog">\r\n\r\n\r\n\r\n\t\t<div id="page" class="page">\r\n\t\t\t<div class="main-header parks-header">\r\n\t\t    <div class="mncppc">\r\n\t\t        <a href="http://www.mncppc.org" title="Maryland-National Capital Parks and Commission" class="mncppc-link" target="_blank">Maryland-National Capital Park and Planning Commission</a>\r\n\t\t        <ul class="sites-links">\r\n\t\t\t\t\t<li class="parks current"><a href="/parks-home">Parks</a></li>\r\n\t\t            <li class="planning"><a href="http://montgomeryplanning.org" target="_blank">Planning</a></li>\r\n\t\t            <li class="board"><a href="http://montgomeryplanningboard.org" target="_blank">Planning Board</a></li>\r\n\t\t        </ul>\r\n\t\t    </div>\r\n\t\t    <div class="branding">\r\n\t\t    \t<div class="col col-tab-9 col-sm-desk-6 col-lg-5 left-wrap p-x">\r\n\t\t    \t\t<div class="mid-v">\r\n\t\t\t\t        <div class="site-icon">\r\n\t\t\t\t        \t<a href="/" class="parks-logo t-indent" title="Back to Home">Montgomery Parks</a>\r\n\t\t\t\t        </div>\r\n\t\t\t\t        <div class="site-title">\r\n\t\t\t\t        \t<a href="/" class="parks-logo" title="Back to Home">\r\n\t\t\t\t        \t<div class="title-wrap">\r\n\t\t\t\t\t            <p class="montco">Montgomery County, MD</p>\r\n\t\t\t\t\t            <h2 class="name">Montgomery Parks</h2>\r\n\t\t\t\t            </div>\r\n\t\t\t\t            </a>\r\n\t\t\t\t        </div>\r\n\t\t\t        </div>\r\n\t\t        </div>\r\n\t\t\t\t<div class="col col-tab-3 col-desk-2 viewing-options">\r\n\t\t\t\t<!-- \t<div class="head-col-wrap"> -->\r\n\t\t\t\t\t\t<div class="options-wrap">\r\n\t\t\t\t\t\t\t<div id="google_translate_element" tabindex="-1"></div>\r\n\t\t\t\t\t\t\t<script type="text/javascript">\r\n                            function fireEvent(element,event){\r\n                                if (document.createEventObject){\r\n                                    // dispatch for IE\r\n                                    var evt = document.createEventObject();\r\n                                    return element.fireEvent(\'on\'+event,evt)\r\n                                }\r\n                                else{\r\n                                    // dispatch for firefox + others\r\n                                    var evt = document.createEvent("HTMLEvents");\r\n                                    evt.initEvent(event, true, true ); // event type,bubbling,cancelable\r\n                                    return !element.dispatchEvent(evt);\r\n                                }\r\n                             }\r\n\r\n                            function googleTranslateElementInit() {\r\n                                new google.translate.TranslateElement({pageLanguage: \'en\', includedLanguages: \'af,ach,ak,am,ar,az,be,bem,bg,bh,bn,br,bs,ca,chr,ckb,co,crs,cs,cy,da,de,ee,el,en,eo,es,es-419,et,eu,fa,fi,fo,fr,fy,ga,gaa,gd,gl,gn,gu,ha,haw,hi,hr,ht,hu,hy,ia,id,ig,is,it,iw,ja,jw,ka,kg,kk,km,kn,ko,kri,ku,ky,la,lg,ln,lo,loz,lt,lua,lv,mfe,mg,mi,mk,ml,mn,mo,mr,ms,mt,ne,nl,nn,no,nso,ny,nyn,oc,om,or,pa,pcm,pl,ps,pt-BR,pt-PT,qu,rm,rn,ro,ru,rw,sd,sh,si,sk,sl,sn,so,sq,sr,sr-ME,st,su,sv,sw,ta,te,tg,th,ti,tk,tl,tn,to,tr,tt,tum,tw,ug,uk,ur,uz,vi,wo,xh,xx-bork,xx-elmer,xx-hacker,xx-klingon,xx-pirate,yi,yo,zh-CN,zh-TW,zu\', layout: google.translate.TranslateElement.InlineLayout}, \'google_translate_element\');\r\n                                setTimeout(function () {\r\n                                    $(\'select.goog-te-combo\').selectize({\r\n                                        onChange: function (val) {\r\n                                            $(\'select.goog-te-combo\').val(val);\r\n                                            fireEvent($(\'select.goog-te-combo\').get(0), \'change\');\r\n                                        }\r\n                                    });\r\n                                    // $(\'#google_translate_element .selectize-input > input\').attr(\'id\', \'google_translate_element\');\r\n                                    // $(\'#google_translate_element .selectize-input > input\').wrap("<label for=\'google_translate_element\'></label>");\r\n                                }, 1000);\r\n                            }\r\n\t\t\t\t\t\t\t</script>\r\n\t\t\t\t\t\t\t<script type="text/javascript" src="//translate.google.com/translate_a/element.js?cb=googleTranslateElementInit"></script>\r\n\t\t\t\t\t\t\t<span class=\'divide-line\'>|</span> <div class="open-wrap"><a class="modal-open" href="#">Text Size</a></div>\r\n\t\t\t\t\t\t</div>\r\n<!-- \t\t\t\t\t</div>\r\n -->\t\t\t\t</div>\r\n\t\t\t\t<div class="mob-hide col-desk-2 col-lg-3 rel">\r\n\t\t\t\t\t<div class="neat-search">\r\n                        <form action="/">\r\n                        \t<label for="search_2">Search the Site</label>\r\n    \t\t\t\t\t\t<input id="search_2" class="closed" type="text" name="s" placeholder="Search" />\r\n    \t\t\t\t\t\t<button type="submit" class="search-link t-indent">Search</button>\r\n                        </form>\r\n\t\t\t\t\t</div>\r\n\t\t\t\t</div>\r\n\t\t\t\t<div class="social col col-sm-desk-3 col-desk-2 rel mob-hide">\r\n\t\t\t\t\t<div class="share">\n\t<a href="https://www.facebook.com/MontgomeryParks" target="_blank">facebook<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1fb{fill:#547bbc}.cls-2{fill:#fff}</style></defs><title>social-fb</title><path class="cls-1fb" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M11.4 12.5H13V11a4 4 0 0 1 .5-2.3 2.8 2.8 0 0 1 2.4-1.1 9.7 9.7 0 0 1 2.8.3l-.4 2.3-1.3-.3a.9.9 0 0 0-1.1.8v1.8h2.5l-.2 2.2h-2.4v7.8h-2.9v-7.8h-1.5v-2.2z"/></svg></a>\n\t<a href="https://twitter.com/MontgomeryParks" target="_blank">twitter<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1tw{fill:#5ea8dd}.cls-2{fill:#fff}</style></defs><title>social-tw</title><path class="cls-1tw" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M22.5 11.1l-1.8.5a3.1 3.1 0 0 0 1.4-1.7 6.2 6.2 0 0 1-2 .7 3.1 3.1 0 0 0-5.3 2.8 8.8 8.8 0 0 1-6.4-3.2 3.1 3.1 0 0 0 1 4.1 3.1 3.1 0 0 1-1.4-.4 3.1 3.1 0 0 0 2.5 3H9.2a3.1 3.1 0 0 0 2.9 2.1 6.2 6.2 0 0 1-3.8 1.3h-.8A8.8 8.8 0 0 0 21 13.1v-.4a6.2 6.2 0 0 0 1.5-1.6"/></svg></a>\n\t<a href="https://www.instagram.com/montgomeryparks/" target="_blank">instagram<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 291.319 291.319"><path d="M145.659 0c80.44 0 145.66 65.219 145.66 145.66S226.1 291.319 145.66 291.319 0 226.1 0 145.66 65.21 0 145.659 0z" fill="#9b6954"/><path d="M195.93 63.708H95.38c-17.47 0-31.672 14.211-31.672 31.672v100.56c0 17.47 14.211 31.672 31.672 31.672h100.56c17.47 0 31.672-14.211 31.672-31.672V95.38c-.001-17.461-14.212-31.672-31.682-31.672zm9.978 18.326l3.587-.009v27.202l-27.402.091-.091-27.202 23.906-.082zm-60.248 36.205c22.732 0 27.42 21.339 27.42 27.429 0 15.103-12.308 27.411-27.42 27.411-15.121 0-27.42-12.308-27.42-27.411-.01-6.09 4.688-27.429 27.42-27.429zm63.99 75.716c0 8.658-7.037 15.704-15.713 15.704H97.073c-8.667 0-15.713-7.037-15.713-15.704v-66.539h22.759c-2.112 5.198-3.305 12.299-3.305 18.253 0 24.708 20.101 44.818 44.818 44.818s44.808-20.11 44.808-44.818c0-5.954-1.193-13.055-3.296-18.253h22.486v66.539h.02z" fill="#fff"/></svg></a>\n\t<a href="https://www.youtube.com/user/MontgomeryParks" target="_blank">youtube<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1yt{fill:#e9644a}.cls-2{fill:#fff}</style></defs><title>social-yt</title><path class="cls-1yt" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M17.9 15L13 17.9v-5.8zm4.6 2.7v-5.4a2.5 2.5 0 0 0-2.6-2.6h-9.8a2.5 2.5 0 0 0-2.6 2.6v5.4a2.5 2.5 0 0 0 2.6 2.6h9.8a2.5 2.5 0 0 0 2.6-2.6"/></svg></a>\n\t<a href="http://www.montgomeryplanningboard.org/rss/rss.shtm" target="_blank">rss<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1rss{fill:#f29100}.cls-2{fill:#fff}</style></defs><title>social-rss</title><path class="cls-1rss" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M10.9 19.1a2 2 0 0 1 0 2.8 2 2 0 1 1-2.8-2.8 2 2 0 0 1 2.8 0zm-3.4-6.5v2.9a7 7 0 0 1 7 7h2.9a9.9 9.9 0 0 0-9.9-9.9zm0-5.1v2.9a12.2 12.2 0 0 1 12.1 12.1h2.9a15 15 0 0 0-15-15z"/></svg></a>\n</div>\t\t\t\t</div>\r\n\r\n\t\t    </div>\r\n\t\t    <div class="nav-primary count-7">\r\n\t\t    \t<div class="icon-wrap rel">\r\n\t\t\t        <div class="ico-link">\r\n\t\t\t        \t<a href="/" class="ico ico-home" title="Back to Home">Back to Home<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 31.9 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-home</title><path class="cls-1" d="M26.7 14.7a.9.9 0 0 0-.9.9v12.6h-6.5v-5.1a.9.9 0 0 0-.9-.9h-4.7a.9.9 0 0 0-.9.9v5.1H6.2V16a.9.9 0 0 0-1.8 0v13.1a.9.9 0 0 0 .9.9h8.4a.9.9 0 0 0 .9-.9V24h2.9v5.1a.9.9 0 0 0 .9.9h8.3a.9.9 0 0 0 .9-.9V15.6a.9.9 0 0 0-.9-.9zm0 0l5 .7-7.3-7.3v-5a.9.9 0 0 0-.9-.9.9.9 0 0 0-.9.9v3.2l-6-6A.9.9 0 0 0 16 0a.9.9 0 0 0-.7.3l-15 15a.9.9 0 0 0 0 1.3.9.9 0 0 0 1.3 0L16 2.2l14.4 14.5a.9.9 0 0 0 .7.2l.7-.2a1 1 0 0 0 0-1.3zm0 0"/></svg></a>\r\n\t\t\t        </div>\r\n\t\t\t        <div class="ico-link">\r\n\t\t\t        \t<a href="/calendar/?v=1" class="ico ico-calendar" title="Calendar of Events">Calendar of Events<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-calendar</title><path class="cls-1" d="M27.5 2.8h-4.1V.9a.9.9 0 0 0-1.9 0v1.9h-5.6V.9a.9.9 0 0 0-1.9 0v1.9H8.4V.9a.9.9 0 1 0-1.8 0v1.9H2.5A2.5 2.5 0 0 0 0 5.3v22.2A2.5 2.5 0 0 0 2.5 30h25a2.5 2.5 0 0 0 2.5-2.5V5.3a2.5 2.5 0 0 0-2.5-2.5zm.6 24.7a.6.6 0 0 1-.6.6h-25a.6.6 0 0 1-.6-.6V5.3a.6.6 0 0 1 .6-.6h4.1v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h4.1a.6.6 0 0 1 .6.6v22.2zm0 0"/><path class="cls-1" d="M6.6 11.3h3.7v2.81H6.6zm0 4.6h3.7v2.81H6.6zm0 4.7h3.7v2.81H6.6zm6.5 0h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7zm6.6 9.3h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7z"/></svg></a>\r\n\t\t\t        </div>\r\n\t\t\t        <div class="ico-link">\r\n\t\t\t        \t<a href="/news/" class="ico ico-news" title="Latest News">Latest News<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30.7 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-news</title><path class="cls-1" d="M30 2.5a.8.8 0 0 0-.8.8v21.1a4 4 0 0 1-4 4H5.7a4 4 0 0 1-4-4V1.6h23.2v20.5a.8.8 0 1 0 1.6 0V.8a.8.8 0 0 0-.8-.8H.8a.8.8 0 0 0-.8.8v23.6A5.6 5.6 0 0 0 5.7 30h19.4a5.6 5.6 0 0 0 5.6-5.6V3.3a.8.8 0 0 0-.8-.8zm0 0"/><path class="cls-1" d="M21.9 6.2a.8.8 0 0 0 0-1.6H4.5a.8.8 0 1 0 0 1.6h17.4zm0 14.6H4.5a.8.8 0 1 0 0 1.6h17.4a.8.8 0 0 0 0-1.6zm0 4.1H7.3a.8.8 0 0 0 0 1.6h14.6a.8.8 0 0 0 0-1.6zM3.5 9v8.6a.8.8 0 0 0 .8.8h8.6a.8.8 0 0 0 .8-.8V9a.8.8 0 0 0-.8-.8H4.3a.8.8 0 0 0-.8.8zm1.6.8h7v7h-7v-7zM22.3 9h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 0-1.6zm0 3.6h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 1 0 0-1.6zm.8 4.6a.8.8 0 0 0-.8-.8h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 .8-.8zm0 0"/></svg></a>\r\n\t\t\t        </div>\r\n\t\t        </div>\r\n\t\t        <div class="menu-primary-container"><ul id="menu-primary" class="menu"><li id="menu-item-2177" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-2177"><a href="/activities/">Things To Do</a>\n<ul class="sub-menu">\n\t<li id="menu-item-664" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-664"><a href="/activities/">Activities</a></li>\n\t<li id="menu-item-8304" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-8304"><a href="/classes-camps/">Classes &#038; Camps</a></li>\n\t<li id="menu-item-950" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-950"><a href="/special-events/">Special Events &#038; Festivals</a></li>\n</ul>\n</li>\n<li id="menu-item-853" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-853"><a href="/parks-directory/">Parks &#038; Trails Directory</a>\n<ul class="sub-menu">\n\t<li id="menu-item-1089" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1089"><a href="/parks-trails/parks/">Parks</a></li>\n\t<li id="menu-item-39673" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-39673"><a href="https://www.montgomeryparks.org/activities/park-trails/">Trails</a></li>\n\t<li id="menu-item-1092" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1092"><a href="/parks-trails/gardens/">Gardens</a></li>\n\t<li id="menu-item-2941" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2941"><a href="/parks-trails/facilities/">Facilities</a></li>\n\t<li id="menu-item-1090" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1090"><a href="/parks-trails/nature-centers/">Nature Centers</a></li>\n\t<li id="menu-item-2360" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2360"><a href="/parks-trails/event-centers/">Event Centers</a></li>\n</ul>\n</li>\n<li id="menu-item-837" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-837"><a href="/services/">Permits, Rentals &#038; Requests</a>\n<ul class="sub-menu">\n\t<li id="menu-item-2850" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2850"><a href="/services/permits-rentals/">Permits &#038; Rentals</a></li>\n\t<li id="menu-item-838" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-838"><a href="/services/report-a-problem/">Parks Information &#038; Customer Service</a></li>\n</ul>\n</li>\n<li id="menu-item-839" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-839"><a href="/caring-for-our-parks/">Caring for Our Parks</a>\n<ul class="sub-menu">\n\t<li id="menu-item-840" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-840"><a href="/caring-for-our-parks/culture-heritage/">Culture &#038; Heritage</a></li>\n\t<li id="menu-item-842" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-842"><a href="/caring-for-our-parks/natural-spaces/">Natural Spaces</a></li>\n\t<li id="menu-item-959" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-959"><a href="/caring-for-our-parks/parks-and-trails/">Park and Trail Planning</a></li>\n\t<li id="menu-item-31505" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-31505"><a href="/caring-for-our-parks/sustainability/">Sustainability</a></li>\n\t<li id="menu-item-844" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-844"><a href="/caring-for-our-parks/wildlife/">Living with Wildlife</a></li>\n</ul>\n</li>\n<li id="menu-item-122" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-122"><a href="/projects/">Parks &#038; Projects</a>\n<ul class="sub-menu">\n\t<li id="menu-item-26054" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-26054"><a href="/projects/">Projects</a></li>\n\t<li id="menu-item-2503" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2503"><a href="/projects/capital-improvements-program/">Funding</a></li>\n\t<li id="menu-item-16631" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-16631"><a href="/projects/public-input/">Open Town Hall</a></li>\n\t<li id="menu-item-912" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-912"><a href="/projects/bids-procurement/">Bids / Procurement</a></li>\n\t<li id="menu-item-897" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-897"><a href="/projects/resources/">Resources</a></li>\n</ul>\n</li>\n<li id="menu-item-123" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-123"><a href="/support/">Support the Parks</a>\n<ul class="sub-menu">\n\t<li id="menu-item-19531" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-19531"><a href="/support/advocacy/">Advocacy</a></li>\n\t<li id="menu-item-907" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-907"><a href="/support/donate/">Donate</a></li>\n\t<li id="menu-item-15740" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-15740"><a href="/support/partnerships/">Partnerships</a></li>\n\t<li id="menu-item-900" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-900"><a href="/support/sponsorships/">Sponsorships</a></li>\n\t<li id="menu-item-901" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-901"><a href="/support/volunteer/">Volunteer</a></li>\n</ul>\n</li>\n<li id="menu-item-1047" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-1047"><a href="/about/parks/">About</a>\n<ul class="sub-menu">\n\t<li id="menu-item-660" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-660"><a href="/about/parks/">About the Parks</a></li>\n\t<li id="menu-item-659" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-659"><a href="/about/divisions/">Divisions &#038; Offices</a></li>\n\t<li id="menu-item-30188" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-30188"><a href="/about/media-center/">Media Center</a></li>\n\t<li id="menu-item-2592" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2592"><a href="/about/park-police/">Police</a></li>\n\t<li id="menu-item-655" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-655"><a href="/about/employment/">Employment</a></li>\n\t<li id="menu-item-932" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-932"><a href="/about/contact-us/">Contact Us</a></li>\n</ul>\n</li>\n</ul></div>\t\t    </div>\r\n\t\t</div>\r\n\t\t<div id="modal">\r\n\t    \t<div class="close">\xc3\x97</div>\r\n\t\t\t<div class="modal-wrapper">\r\n\t\t\t\t<div class="close">\xc3\x97</div>\r\n\t\t\t\tTo change the text size you can use your web browser\'s settings. Most browsers include functionality to let you increase or decrease the text on a web page. For example, to increase text size using:<br><br>\r\n\r\n\t\t\t\t<strong>Chrome</strong><br>\r\n\r\n\t\t\t\tIn the menu to the right of the address bar, select and set Zoom level. Menu > Zoom > +<br><br>\r\n\r\n\t\t\t\t<strong>Firefox</strong><br>\r\n\r\n\t\t\t\tIn the View menu, select Zoom. View > Zoom > Zoom In<br><br>\r\n\r\n\t\t\t\t<strong>Internet Explorer</strong><br>\r\n\r\n\t\t\t\tIn the View menu, select Text Size. View > Text Size > Largest<br><br>\r\n\r\n\t\t\t\t<strong>Safari</strong><br>\r\n\r\n\t\t\t\tIn the View menu, select Zoom In. View > Zoom In\r\n\t\t\t\tMacintosh Shortcut: Command+\r\n\t\t\t</div>\r\n\t\t</div>\r\n\t\t<div class="mobile-nav count-7">\r\n\t\t\t<div class="just-click"></div>\r\n\t\t\t<div class="upper-half">\r\n\t\t\t\t<div class="menu-primary-container"><ul id="menu-primary-1" class="menu"><li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-2177"><a href="/activities/">Things To Do</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-664"><a href="/activities/">Activities</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-8304"><a href="/classes-camps/">Classes &#038; Camps</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-950"><a href="/special-events/">Special Events &#038; Festivals</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-853"><a href="/parks-directory/">Parks &#038; Trails Directory</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1089"><a href="/parks-trails/parks/">Parks</a></li>\n\t<li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-39673"><a href="https://www.montgomeryparks.org/activities/park-trails/">Trails</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1092"><a href="/parks-trails/gardens/">Gardens</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2941"><a href="/parks-trails/facilities/">Facilities</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1090"><a href="/parks-trails/nature-centers/">Nature Centers</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2360"><a href="/parks-trails/event-centers/">Event Centers</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-837"><a href="/services/">Permits, Rentals &#038; Requests</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2850"><a href="/services/permits-rentals/">Permits &#038; Rentals</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-838"><a href="/services/report-a-problem/">Parks Information &#038; Customer Service</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-839"><a href="/caring-for-our-parks/">Caring for Our Parks</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-840"><a href="/caring-for-our-parks/culture-heritage/">Culture &#038; Heritage</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-842"><a href="/caring-for-our-parks/natural-spaces/">Natural Spaces</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-959"><a href="/caring-for-our-parks/parks-and-trails/">Park and Trail Planning</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-31505"><a href="/caring-for-our-parks/sustainability/">Sustainability</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-844"><a href="/caring-for-our-parks/wildlife/">Living with Wildlife</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-122"><a href="/projects/">Parks &#038; Projects</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-26054"><a href="/projects/">Projects</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2503"><a href="/projects/capital-improvements-program/">Funding</a></li>\n\t<li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-16631"><a href="/projects/public-input/">Open Town Hall</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-912"><a href="/projects/bids-procurement/">Bids / Procurement</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-897"><a href="/projects/resources/">Resources</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-123"><a href="/support/">Support the Parks</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-19531"><a href="/support/advocacy/">Advocacy</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-907"><a href="/support/donate/">Donate</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-15740"><a href="/support/partnerships/">Partnerships</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-900"><a href="/support/sponsorships/">Sponsorships</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-901"><a href="/support/volunteer/">Volunteer</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-1047"><a href="/about/parks/">About</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-660"><a href="/about/parks/">About the Parks</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-659"><a href="/about/divisions/">Divisions &#038; Offices</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-30188"><a href="/about/media-center/">Media Center</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2592"><a href="/about/park-police/">Police</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-655"><a href="/about/employment/">Employment</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-932"><a href="/about/contact-us/">Contact Us</a></li>\n</ul>\n</li>\n</ul></div>\t\t\t</div>\r\n\t\t\t<div class="lower-half rel">\r\n\t\t\t\t<div class="neat-search">\r\n\t\t\t\t\t<label for="m_search" />\r\n\t\t\t\t\t<input id="m_search" class="closed" type="text" placeholder="Search" />\r\n\t\t\t\t\t<a class="search-link t-indent" href="#m_search">Search</a>\r\n\t\t\t\t</div>\r\n\t\t\t\t<div class="page-nav">\r\n\t\t\t\t\t<a href="http://www.montgomeryparks.org/" target="_blank">Parks</a> &#8226;\r\n\t\t\t\t\t<a href="http://montgomeryplanning.org/" target="_blank">Planning</a> &#8226;\r\n\t\t\t\t\t<a href="http://montgomeryplanningboard.org/" target="_blank">Planning Board</a>\r\n\t\t\t\t</div>\r\n\t\t\t\t\t<div class="share">\n\t<a href="https://www.facebook.com/MontgomeryParks" target="_blank">facebook<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1fb{fill:#547bbc}.cls-2{fill:#fff}</style></defs><title>social-fb</title><path class="cls-1fb" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M11.4 12.5H13V11a4 4 0 0 1 .5-2.3 2.8 2.8 0 0 1 2.4-1.1 9.7 9.7 0 0 1 2.8.3l-.4 2.3-1.3-.3a.9.9 0 0 0-1.1.8v1.8h2.5l-.2 2.2h-2.4v7.8h-2.9v-7.8h-1.5v-2.2z"/></svg></a>\n\t<a href="https://twitter.com/MontgomeryParks" target="_blank">twitter<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1tw{fill:#5ea8dd}.cls-2{fill:#fff}</style></defs><title>social-tw</title><path class="cls-1tw" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M22.5 11.1l-1.8.5a3.1 3.1 0 0 0 1.4-1.7 6.2 6.2 0 0 1-2 .7 3.1 3.1 0 0 0-5.3 2.8 8.8 8.8 0 0 1-6.4-3.2 3.1 3.1 0 0 0 1 4.1 3.1 3.1 0 0 1-1.4-.4 3.1 3.1 0 0 0 2.5 3H9.2a3.1 3.1 0 0 0 2.9 2.1 6.2 6.2 0 0 1-3.8 1.3h-.8A8.8 8.8 0 0 0 21 13.1v-.4a6.2 6.2 0 0 0 1.5-1.6"/></svg></a>\n\t<a href="https://www.instagram.com/montgomeryparks/" target="_blank">instagram<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 291.319 291.319"><path d="M145.659 0c80.44 0 145.66 65.219 145.66 145.66S226.1 291.319 145.66 291.319 0 226.1 0 145.66 65.21 0 145.659 0z" fill="#9b6954"/><path d="M195.93 63.708H95.38c-17.47 0-31.672 14.211-31.672 31.672v100.56c0 17.47 14.211 31.672 31.672 31.672h100.56c17.47 0 31.672-14.211 31.672-31.672V95.38c-.001-17.461-14.212-31.672-31.682-31.672zm9.978 18.326l3.587-.009v27.202l-27.402.091-.091-27.202 23.906-.082zm-60.248 36.205c22.732 0 27.42 21.339 27.42 27.429 0 15.103-12.308 27.411-27.42 27.411-15.121 0-27.42-12.308-27.42-27.411-.01-6.09 4.688-27.429 27.42-27.429zm63.99 75.716c0 8.658-7.037 15.704-15.713 15.704H97.073c-8.667 0-15.713-7.037-15.713-15.704v-66.539h22.759c-2.112 5.198-3.305 12.299-3.305 18.253 0 24.708 20.101 44.818 44.818 44.818s44.808-20.11 44.808-44.818c0-5.954-1.193-13.055-3.296-18.253h22.486v66.539h.02z" fill="#fff"/></svg></a>\n\t<a href="https://www.youtube.com/user/MontgomeryParks" target="_blank">youtube<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1yt{fill:#e9644a}.cls-2{fill:#fff}</style></defs><title>social-yt</title><path class="cls-1yt" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M17.9 15L13 17.9v-5.8zm4.6 2.7v-5.4a2.5 2.5 0 0 0-2.6-2.6h-9.8a2.5 2.5 0 0 0-2.6 2.6v5.4a2.5 2.5 0 0 0 2.6 2.6h9.8a2.5 2.5 0 0 0 2.6-2.6"/></svg></a>\n\t<a href="http://www.montgomeryplanningboard.org/rss/rss.shtm" target="_blank">rss<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1rss{fill:#f29100}.cls-2{fill:#fff}</style></defs><title>social-rss</title><path class="cls-1rss" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M10.9 19.1a2 2 0 0 1 0 2.8 2 2 0 1 1-2.8-2.8 2 2 0 0 1 2.8 0zm-3.4-6.5v2.9a7 7 0 0 1 7 7h2.9a9.9 9.9 0 0 0-9.9-9.9zm0-5.1v2.9a12.2 12.2 0 0 1 12.1 12.1h2.9a15 15 0 0 0-15-15z"/></svg></a>\n</div>\t\t\t\t\t<a href="http://www.mncppc.org" title="Maryland-National Capital Parks and Commission" class="mncppc-link" target="_blank">Maryland-National Capital Park and Planning Commission</a>\r\n\t\t\t\t</div>\r\n\t\t\t</div>\r\n\t\t</div>\r\n\t\t<div id="mobile-footer">\r\n\t\t    <div class="footer-icon">\r\n\t\t    \t<a href="/" class="ico ico-home" title="Back to Home"><svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 31.9 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-home</title><path class="cls-1" d="M26.7 14.7a.9.9 0 0 0-.9.9v12.6h-6.5v-5.1a.9.9 0 0 0-.9-.9h-4.7a.9.9 0 0 0-.9.9v5.1H6.2V16a.9.9 0 0 0-1.8 0v13.1a.9.9 0 0 0 .9.9h8.4a.9.9 0 0 0 .9-.9V24h2.9v5.1a.9.9 0 0 0 .9.9h8.3a.9.9 0 0 0 .9-.9V15.6a.9.9 0 0 0-.9-.9zm0 0l5 .7-7.3-7.3v-5a.9.9 0 0 0-.9-.9.9.9 0 0 0-.9.9v3.2l-6-6A.9.9 0 0 0 16 0a.9.9 0 0 0-.7.3l-15 15a.9.9 0 0 0 0 1.3.9.9 0 0 0 1.3 0L16 2.2l14.4 14.5a.9.9 0 0 0 .7.2l.7-.2a1 1 0 0 0 0-1.3zm0 0"/></svg></a>\r\n\t\t    </div>\r\n\t\t    <div class="footer-icon">\r\n\t\t    <a href="/calendar" class="ico ico-calendar" title="Calendar"><svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-calendar</title><path class="cls-1" d="M27.5 2.8h-4.1V.9a.9.9 0 0 0-1.9 0v1.9h-5.6V.9a.9.9 0 0 0-1.9 0v1.9H8.4V.9a.9.9 0 1 0-1.8 0v1.9H2.5A2.5 2.5 0 0 0 0 5.3v22.2A2.5 2.5 0 0 0 2.5 30h25a2.5 2.5 0 0 0 2.5-2.5V5.3a2.5 2.5 0 0 0-2.5-2.5zm.6 24.7a.6.6 0 0 1-.6.6h-25a.6.6 0 0 1-.6-.6V5.3a.6.6 0 0 1 .6-.6h4.1v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h4.1a.6.6 0 0 1 .6.6v22.2zm0 0"/><path class="cls-1" d="M6.6 11.3h3.7v2.81H6.6zm0 4.6h3.7v2.81H6.6zm0 4.7h3.7v2.81H6.6zm6.5 0h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7zm6.6 9.3h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7z"/></svg></a></div>\r\n\t\t    <div class="footer-icon"><a href="/news" class="ico ico-news" title="News"><svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30.7 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-news</title><path class="cls-1" d="M30 2.5a.8.8 0 0 0-.8.8v21.1a4 4 0 0 1-4 4H5.7a4 4 0 0 1-4-4V1.6h23.2v20.5a.8.8 0 1 0 1.6 0V.8a.8.8 0 0 0-.8-.8H.8a.8.8 0 0 0-.8.8v23.6A5.6 5.6 0 0 0 5.7 30h19.4a5.6 5.6 0 0 0 5.6-5.6V3.3a.8.8 0 0 0-.8-.8zm0 0"/><path class="cls-1" d="M21.9 6.2a.8.8 0 0 0 0-1.6H4.5a.8.8 0 1 0 0 1.6h17.4zm0 14.6H4.5a.8.8 0 1 0 0 1.6h17.4a.8.8 0 0 0 0-1.6zm0 4.1H7.3a.8.8 0 0 0 0 1.6h14.6a.8.8 0 0 0 0-1.6zM3.5 9v8.6a.8.8 0 0 0 .8.8h8.6a.8.8 0 0 0 .8-.8V9a.8.8 0 0 0-.8-.8H4.3a.8.8 0 0 0-.8.8zm1.6.8h7v7h-7v-7zM22.3 9h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 0-1.6zm0 3.6h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 1 0 0-1.6zm.8 4.6a.8.8 0 0 0-.8-.8h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 .8-.8zm0 0"/></svg></a></div>\r\n\t\t    <div class="menu-open">\r\n\t\t    \t<div id="hamburger">\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t</div>\r\n\t\t\t</div>\r\n\t\t</div>\r\n\t\t<div id="siteBody" class="site-body">\r\n<div class="calendar left-sb" data-date="2019-01-31" data-year="2019" data-month="01" data-view="month">\n    <div class="section-headline row">\n        <div class="p-lg-x-l">\n            <h1 class="section-head">Calendar of Events</h1>\n        </div>\n    </div>\n    <div class="sb-flex rel">\n        <div class="col col-sm-desk-12 col-lg-4 side-l col-sb">\n            <div class="sb-wrap">\n                <div class="sidebar">\n\n                    <div class="calendar-box col-sm-desk-6 col-lg-12">\n                        <div class="row controls rel">\n                            <div class="prev f-left">\n                                <a href="?cat=2243&v=0&date=2018-12-01"><span class="hv">Prev</span></a>\n                            </div>\n\n                            <div class="title f-center">\n                                <h2></h2>\n                            </div>\n\n                            <div class="next f-right">\n                                <a href="?cat=2243&v=0&date=2019-03-01"><span class="hv">Next</span></a>\n                            </div>\n                        </div>\n                        <div class="row days f-center m-y">\n                            <div class="col col-7th">Sun</div>\n                            <div class="col col-7th">Mon</div>\n                            <div class="col col-7th">Tue</div>\n                            <div class="col col-7th">Wed</div>\n                            <div class="col col-7th">Thu</div>\n                            <div class="col col-7th">Fri</div>\n                            <div class="col col-7th">Sat</div>\n                        </div>\n\n                        <div class="js-dates row">\n\n                        </div>\n                    </div>\n                    <div class="p-y">\n                        <h4>Search for An Event</h4>\n                        <div class="neat-search">\n                            <form action="" method="get" class="cal-search">\n                              <label for="search-2">Search Events</label>\n                              <input name="s" id="search-2" class="closed" type="text" placeholder="Search" value=""/>\n                              <button class="search-link" type="submit">Search Events</button>\n                            </form>\n                        </div>\n                    </div>\n                                            <div class="p-y-b inline-block">\n                            <h4>Search by Park</h4>\n                            <div class="neat-search">\n                                <form action="" method="get" class="cal-park-search">\n                                    <label for="search-5">Search Parks</label>\n                                    <input id="search-5" class="closed park-s" type="text" placeholder="Park Name" name="park" value=""/>\n                                    <button class="search-link" href="#search-5" type="submit">Search Parks</button>\n                                </form>\n                            </div>\n                        </div>\n                                            <div class="filter-list accordion col-sm-desk-6 col-lg-12">\n                        <div class="arrow"></div>\n                        <h4>Filter by category</h4>\n                        <a href="?cat=&v=0" id="clear" class="clear"><svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 47.5 45.1"><path d="M46.1 29.6c-.4-.1-.9.2-.9.6-.9 4.4-4.4 7.4-8.6 7.4-4 0-7.6-2.6-8.8-6.4h3.8v-1.6h-6.4V36h1.6v-3.1c1.7 3.7 5.5 6.3 9.8 6.3 5 0 9.1-3.5 10.2-8.6 0-.5-.3-.9-.7-1zM45.9 20v3.4c-1.8-3.6-5.4-5.8-9.4-5.8-4.9 0-9.2 3.2-10.6 7.8-.1.4.1.9.5 1 .4.1.9-.1 1-.5 1.1-3.8 4.9-6.6 9-6.6 3.7 0 6.8 2.2 8.3 5.6H41v1.6h6.4V20h-1.5zM30.8 10.5c0 1.7 1.4 3.1 3.1 3.1 1.7 0 3.2-1.4 3.2-3.1 0-1.7-1.4-3.1-3.2-3.1-1.7 0-3.1 1.4-3.1 3.1zm4.7 0c0 .9-.7 1.6-1.6 1.6-.9 0-1.5-.7-1.5-1.6 0-.9.7-1.5 1.5-1.5.9 0 1.6.7 1.6 1.5z"/><path d="M25.7 1.6h9.9l8 7.9v7.3h1.6V8.9l-9-8.9H25L4.8 20.2 0 25l20.1 20.2 6.2-5.9-19.2-19L25.7 1.6zM24 39.2l-3.9 3.7L2.3 25 6 21.3l18 17.9z"/></svg>clear</a>\n\n                        <ul class="filters accordion-wrap">\n                                                            <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1751&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Archaeology                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2230&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Butterfly Show                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2701&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Camp                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2235&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Clean Up                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1738&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Community Events                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1739&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Community Gardens                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2577&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Community Meeting                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=3105&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Earth Month                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2239&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Education                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=3331&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Event Center                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1740&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Events                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1741&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Family                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2328&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Featured                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2232&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Festival                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1746&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Free                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn selected"\n                                       href="?cat=2243&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Gardens                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2242&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Golf                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1733&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Hikes                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1743&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Historical/Cultural                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2240&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                History                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=3215&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                History in the Parks                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2234&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                History Program                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1727&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Hockey                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2241&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Holiday                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1728&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Ice Skating                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1729&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Kids                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2705&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                MLK Day of Service                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1744&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Music                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=82&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Nature                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2901&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                open house                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1753&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Openings/Dedications                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2233&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Other                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2902&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Park Planning                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2916&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Park Projects                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2751&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Planning Board Meeting                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2756&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Planning Board Worksession                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2569&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                pop up                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2754&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Public Hearing                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1750&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Public Meetings                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2231&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Run                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1730&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Special Events                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1731&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Sports                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1734&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Tennis                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2903&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Trail Planning                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2706&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Trail Work                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1735&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Trails                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1736&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Trips                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1747&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Volunteering                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1748&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Weed Warrior                                    </a>\n                                </li>\n                                                        </ul>\n                    </div>\n                </div>\n            </div>\n        </div>\n        <div class="col col-sm-desk-12 col-lg-5 events">\n                <!--p>221 events meet your criteria of [filter item] & [filter item].</p-->\n            <section class="event-list" id="ajax-event-list">\n                <div class="p-y" style="text-align: right">\n                    <a href="#" class="toggle-view toggle-view-day">Switch to Day View</a>                                    </div>\n                <div class="event-item events-2019-01-31 view-m">\n                                        <ul>\n                                                <li>\n                                <a data-id="44559" href="/events/how-to-get-your-orchid-to-rebloom-3/">\n                                    <span class="time">\n                                        Sat. March 16th, 2019 1:00pm to  2:30pm                                    </span>\n                                    <span class="event-name m-view">How to Get Your Orchid to Rebloom</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44558" href="/events/hands-on-introduction-to-plant-propagation/">\n                                    <span class="time">\n                                        Sat. March 23rd, 2019 10:00am to  12:00pm                                    </span>\n                                    <span class="event-name m-view">Hands-On Introduction to Plant Propagation</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44537" href="/events/advanced-composition/">\n                                    <span class="time">\n                                        Sat. March 23rd, 2019 10:00am to  3:30pm                                    </span>\n                                    <span class="event-name m-view">Advanced Composition</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44560" href="/events/introduction-to-landscape-design-session-i-2/">\n                                    <span class="time">\n                                        Sat. March 30th, 2019 10:00am to  12:00pm                                    </span>\n                                    <span class="event-name m-view">Introduction to Landscape Design: Session I</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44577" href="/events/speaker-series-undaunted-daffodils/">\n                                    <span class="time">\n                                        Sun. March 31st, 2019 3:00pm to  4:30pm                                    </span>\n                                    <span class="event-name m-view">Speaker Series: Undaunted Daffodils</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44562" href="/events/make-and-take-rain-barrels-5/">\n                                    <span class="time">\n                                        Thu. April 4th, 2019 4:00pm to  7:00pm                                    </span>\n                                    <span class="event-name m-view">Make and Take Rain Barrels</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44563" href="/events/make-and-take-rain-barrels-6/">\n                                    <span class="time">\n                                        Sun. April 7th, 2019 10:00am to  1:00pm                                    </span>\n                                    <span class="event-name m-view">Make and Take Rain Barrels</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44571" href="/events/ornamental-flowering-trees-shrubs/">\n                                    <span class="time">\n                                        Tue. April 9th, 2019 1:30pm to  3:30pm                                    </span>\n                                    <span class="event-name m-view">Ornamental Flowering Trees &#038; Shrubs</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44561" href="/events/introduction-to-landscape-design-session-ii-2/">\n                                    <span class="time">\n                                        Sat. April 13th, 2019 10:00am to  12:00pm                                    </span>\n                                    <span class="event-name m-view">Introduction to Landscape Design: Session II</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44542" href="/events/brookside-gardens-greenhouse-tour-2/">\n                                    <span class="time">\n                                        Sun. April 14th, 2019 2:00pm to  3:15pm                                    </span>\n                                    <span class="event-name m-view">Brookside Gardens Greenhouse Tour</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44578" href="/events/spring-at-sherwood-gardens-rawlings-conservatory/">\n                                    <span class="time">\n                                        Wed. April 17th, 2019 9:00am to  4:30pm                                    </span>\n                                    <span class="event-name m-view">Spring at Sherwood Gardens &#038; Rawlings Conservatory</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44548" href="/events/designing-for-dry-shade-2/">\n                                    <span class="time">\n                                        Thu. April 18th, 2019 2:00pm to  3:30pm                                    </span>\n                                    <span class="event-name m-view">Designing For Dry Shade</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44574" href="/events/shinrin-yoku-the-art-of-forest-bathing/">\n                                    <span class="time">\n                                        Sat. April 20th, 2019 10:00am to  12:30pm                                    </span>\n                                    <span class="event-name m-view">Shinrin-yoku: The Art of Forest Bathing</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44567" href="/events/miniature-perennial-container-3/">\n                                    <span class="time">\n                                        Tue. April 23rd, 2019 10:00am to  11:30am                                    </span>\n                                    <span class="event-name m-view">Miniature Perennial Container</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44545" href="/events/brookside-gardens-spring-tour-3/">\n                                    <span class="time">\n                                        Wed. April 24th, 2019 1:00pm to  2:30pm                                    </span>\n                                    <span class="event-name m-view">Brookside Gardens Spring Tour</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44541" href="/events/brookside-gardens-greenhouse-tour/">\n                                    <span class="time">\n                                        Wed. April 24th, 2019 2:00pm to  3:15pm                                    </span>\n                                    <span class="event-name m-view">Brookside Gardens Greenhouse Tour</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44573" href="/events/perfect-plant-marriages/">\n                                    <span class="time">\n                                        Sat. April 27th, 2019 10:00am to  11:30am                                    </span>\n                                    <span class="event-name m-view">Perfect Plant Marriages</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44539" href="/events/brookside-gardens-azalea-rhododendron-tour/">\n                                    <span class="time">\n                                        Wed. May 1st, 2019 1:00pm to  2:30pm                                    </span>\n                                    <span class="event-name m-view">Brookside Gardens Azalea &#038; Rhododendron Tour</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44546" href="/events/conservation-landscapes-and-rain-gardens-5/">\n                                    <span class="time">\n                                        Wed. May 1st, 2019 4:00pm to  7:00pm                                    </span>\n                                    <span class="event-name m-view">Conservation Landscapes and Rain Gardens</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44547" href="/events/conservation-landscapes-and-rain-gardens-8/">\n                                    <span class="time">\n                                        Fri. May 3rd, 2019 10:00am to  1:00pm                                    </span>\n                                    <span class="event-name m-view">Conservation Landscapes and Rain Gardens</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44580" href="/events/summer-annual-container-3/">\n                                    <span class="time">\n                                        Sat. May 4th, 2019 10:00am to  11:30am                                    </span>\n                                    <span class="event-name m-view">Summer Annual Container</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44538" href="/events/art-nature-journaling-with-mixed-media/">\n                                    <span class="time">\n                                        Sat. May 4th, 2019 10:00am to  2:30pm                                    </span>\n                                    <span class="event-name m-view">Art &#038; Nature Journaling with Mixed Media</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44570" href="/events/mother-s-day-floral-design/">\n                                    <span class="time">\n                                        Sat. May 11th, 2019 10:00am to  12:00pm                                    </span>\n                                    <span class="event-name m-view">Mother&#8217;s Day Floral Design</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44572" href="/events/peppers-nursery-red-mill-pond-garden/">\n                                    <span class="time">\n                                        Thu. May 16th, 2019 8:00am to  6:00pm                                    </span>\n                                    <span class="event-name m-view">Peppers Nursery &#038; Red Mill Pond Garden</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44566" href="/events/miniature-hostas/">\n                                    <span class="time">\n                                        Wed. May 22nd, 2019 10:00am to  11:30am                                    </span>\n                                    <span class="event-name m-view">Miniature Hostas</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44581" href="/events/sustainable-gardening-practices/">\n                                    <span class="time">\n                                        Sat. June 1st, 2019 10:00am to  11:30am                                    </span>\n                                    <span class="event-name m-view">Sustainable Gardening Practices</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                        <li>\n                                <a data-id="44569" href="/events/morris-arboretum-trip/">\n                                    <span class="time">\n                                        Wed. June 12th, 2019 8:00am to  6:00pm                                    </span>\n                                    <span class="event-name m-view">Morris Arboretum Trip</span>\n                                </a>\n\n                                                                    <div class=\'location\'>\n                                        @\n                                        <span href=\'/parks-and-trails/brookside-gardens/\' class=\'location\'>Brookside Gardens</span>                                    </div>\n                                                                </li>\n                                                </ul>\n                    <div class="pag row">\n                        <div class="col col-6"></div>\n                        <div style="text-align: right;" class="col col-6 a-right"></div>\n                    </div>\n                                    </div>\n            </section>\n            <div class="loading"></div>\n        </div>\n        <div class="col col-sm-desk-12 col-lg-3 p-lg-y-t featured-events">\n                                <div class="popular-events p">\n                        <h3>Featured Events</h3>\n\n                                                    <div class="p-y single-pop-event">\n                                <a href="/events/we-speak-weddings-lodge-at-little-seneca-creek/">\n                                    We Speak Weddings &#8211; Lodge at Little Seneca Creek                                </a><br />\n                                Sun. February 3rd, 2019 10:00am to  1:00pm                            </div>\n                                                        <div class="p-y single-pop-event">\n                                <a href="/events/wedding-vendor-showcase-2/">\n                                    Wedding Vendor Showcase                                </a><br />\n                                Sun. March 3rd, 2019 11:00am to  2:00pm                            </div>\n                                                        <div class="p-y single-pop-event">\n                                <a href="/events/volunteer-fair-for-montgomery-parks-historic-sites/">\n                                    Join Parks Historic Sites Volunteer Program                                </a><br />\n                                Sat. March 23rd, 2019 9:00am to  3:00pm                            </div>\n                                                </div>\n                            </div>\n    </div>\n</div>\n<style>\n.m-no-time span.time{\n    display:none !important;\n}\n\n.m-no-time span.event-name{\n    padding-left: 0 !important;\n}\n</style>\n<div id="footer" class="main-footer">\r\n                        \r\n                    <section class="hot-topics row">\r\n                        <div class="cc-m">\r\n                            <h3>Topics</h3>\r\n                            <aside class="hot-topic">\r\n                                <div class="menu-hot-topics-container"><ul id="menu-hot-topics" class="menu"><li id="menu-item-14005" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-14005"><a href="/special-events/">Festivals &#038; Events</a></li>\n<li id="menu-item-14007" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-14007"><a href="/classes-camps/">Register Now for Classes &#038; Programs</a></li>\n<li id="menu-item-14008" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-14008"><a href="/support/volunteer/seasonal_volunteering/">Volunteer Events</a></li>\n<li id="menu-item-16517" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-16517"><a href="/calendar/">Calendar of Events</a></li>\n<li id="menu-item-14011" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-14011"><a href="/caring-for-our-parks/wildlife/ticks-lyme-disease/">Ticks and Lyme Disease</a></li>\n<li id="menu-item-16515" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-16515"><a href="/projects/public-input/">Open Town Hall</a></li>\n<li id="menu-item-18385" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-18385"><a href="/activities/history-in-the-parks/">History in the Parks</a></li>\n<li id="menu-item-33409" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-33409"><a href="https://www.montgomeryparks.org/about/parks/pesticides/">Pesticide Reduction &#038; Integrated Pest Management</a></li>\n</ul></div>                            </aside>\r\n                                                    </div>\r\n                    </section>\r\n \r\n                    <section class="main-footer-bottom row">\r\n        <div class="col col-mob-12 col-desk-3">\r\n            <div class="address">\r\n                Montgomery County<br>Department of Parks<br>\r\n                9500 Brunett Ave.<br>\r\n                Silver Spring, MD 20901\r\n            </div>\r\n        </div>\r\n                        <div class="col col-mob-6 col-sm-desk-4 col-desk-2 footer-nav">\r\n                                            <a href="/about/contact-us/">Contact Us</a>\r\n                                                <a href="/about/media-center/press-contacts/">Press</a>\r\n                                                <a href="/about/parks/accessibility/">ADA Accessibility</a>\r\n                                                <a href="/terms/">Terms & Privacy</a>\r\n                                        </div>\r\n                                <div class="col col-mob-6 col-sm-desk-4 col-desk-2 footer-nav">\r\n                                            <a href="/about/park-police/">Park Police</a>\r\n                                                <a href="/about/rules-and-regulations/">Rules and Regulations</a>\r\n                                                <a href="/support/volunteer/">Volunteer</a>\r\n                                                <a href="/activities/">Activities</a>\r\n                                        </div>\r\n                                <div class="col col-mob-6 col-sm-desk-4 col-desk-2 footer-nav">\r\n                                            <a href="/permits/">Permits</a>\r\n                                                <a href="/calendar/">Calendar</a>\r\n                                                <a href="/news/">News</a>\r\n                                                <a href="/weather-alerts/">Weather Alerts</a>\r\n                                        </div>\r\n                        <div class="col col-sm-desk-3 pull-right mob-hide rel"><a href="/calendar/page/2/?cat=1727&amp;v=0&amp;_=1548894797807">Next Page</a>\r\n            SOCIAL<br>\r\n            <div class="share">\n\t<a href="https://www.facebook.com/MontgomeryParks" target="_blank">facebook<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1fb{fill:#547bbc}.cls-2{fill:#fff}</style></defs><title>social-fb</title><path class="cls-1fb" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M11.4 12.5H13V11a4 4 0 0 1 .5-2.3 2.8 2.8 0 0 1 2.4-1.1 9.7 9.7 0 0 1 2.8.3l-.4 2.3-1.3-.3a.9.9 0 0 0-1.1.8v1.8h2.5l-.2 2.2h-2.4v7.8h-2.9v-7.8h-1.5v-2.2z"/></svg></a>\n\t<a href="https://twitter.com/MontgomeryParks" target="_blank">twitter<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1tw{fill:#5ea8dd}.cls-2{fill:#fff}</style></defs><title>social-tw</title><path class="cls-1tw" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M22.5 11.1l-1.8.5a3.1 3.1 0 0 0 1.4-1.7 6.2 6.2 0 0 1-2 .7 3.1 3.1 0 0 0-5.3 2.8 8.8 8.8 0 0 1-6.4-3.2 3.1 3.1 0 0 0 1 4.1 3.1 3.1 0 0 1-1.4-.4 3.1 3.1 0 0 0 2.5 3H9.2a3.1 3.1 0 0 0 2.9 2.1 6.2 6.2 0 0 1-3.8 1.3h-.8A8.8 8.8 0 0 0 21 13.1v-.4a6.2 6.2 0 0 0 1.5-1.6"/></svg></a>\n\t<a href="https://www.instagram.com/montgomeryparks/" target="_blank">instagram<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 291.319 291.319"><path d="M145.659 0c80.44 0 145.66 65.219 145.66 145.66S226.1 291.319 145.66 291.319 0 226.1 0 145.66 65.21 0 145.659 0z" fill="#9b6954"/><path d="M195.93 63.708H95.38c-17.47 0-31.672 14.211-31.672 31.672v100.56c0 17.47 14.211 31.672 31.672 31.672h100.56c17.47 0 31.672-14.211 31.672-31.672V95.38c-.001-17.461-14.212-31.672-31.682-31.672zm9.978 18.326l3.587-.009v27.202l-27.402.091-.091-27.202 23.906-.082zm-60.248 36.205c22.732 0 27.42 21.339 27.42 27.429 0 15.103-12.308 27.411-27.42 27.411-15.121 0-27.42-12.308-27.42-27.411-.01-6.09 4.688-27.429 27.42-27.429zm63.99 75.716c0 8.658-7.037 15.704-15.713 15.704H97.073c-8.667 0-15.713-7.037-15.713-15.704v-66.539h22.759c-2.112 5.198-3.305 12.299-3.305 18.253 0 24.708 20.101 44.818 44.818 44.818s44.808-20.11 44.808-44.818c0-5.954-1.193-13.055-3.296-18.253h22.486v66.539h.02z" fill="#fff"/></svg></a>\n\t<a href="https://www.youtube.com/user/MontgomeryParks" target="_blank">youtube<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1yt{fill:#e9644a}.cls-2{fill:#fff}</style></defs><title>social-yt</title><path class="cls-1yt" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M17.9 15L13 17.9v-5.8zm4.6 2.7v-5.4a2.5 2.5 0 0 0-2.6-2.6h-9.8a2.5 2.5 0 0 0-2.6 2.6v5.4a2.5 2.5 0 0 0 2.6 2.6h9.8a2.5 2.5 0 0 0 2.6-2.6"/></svg></a>\n\t<a href="http://www.montgomeryplanningboard.org/rss/rss.shtm" target="_blank">rss<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1rss{fill:#f29100}.cls-2{fill:#fff}</style></defs><title>social-rss</title><path class="cls-1rss" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M10.9 19.1a2 2 0 0 1 0 2.8 2 2 0 1 1-2.8-2.8 2 2 0 0 1 2.8 0zm-3.4-6.5v2.9a7 7 0 0 1 7 7h2.9a9.9 9.9 0 0 0-9.9-9.9zm0-5.1v2.9a12.2 12.2 0 0 1 12.1 12.1h2.9a15 15 0 0 0-15-15z"/></svg></a>\n</div>        </div>\r\n    </section>\r\n</div>\r\n        <style>\n    .brink-gdpr-cookie-middle{\n      border-right:1px solid rgba(255,255,255,.5);\n    }\n\n    .brink-gdpr-cookie-left a{\n      color:#fff;\n    }\n\n    #brink-gdpr-cookie-accept{\n      background:rgb(0,0,0);\n      color:#fff;\n    }\n\n    #brink-gdpr-cookie-accept:hover, #brink-gdpr-cookie-accept:active{\n      background:rgb(255,255,255);\n      color:#000;\n    }\n\n    .brink-gdpr-cookie-right span:before{\n      border-top:1px solid #fff;\n      border-right:1px solid #fff;\n    }\n\n    .brink-gdpr-cookie-right span:after{\n      border-top:1px solid #fff;\n      border-left:1px solid #fff;\n    }\n    </style>\n    <div id="brink-gdpr-cookie-notice" class="brink-gdpr-cookie-notice bottom" style="background:rgba(102,102,102,.95); color:#fff">\n      <div class="brink-gdpr-container">\n        <div class="brink-gdpr-cookie-left">\n          <span class="h2">Cookies Notice</span>\n            <p>\n              This website uses cookies to provide you with an improved and personalized experience. By using this site you agree to our use of cookies. Please read our <a href="/terms/" target="_blank">cookies policy</a> for more information on the cookies we use and how to delete or block them.            </p>\n          </span>\n        </div>\n        <div class="brink-gdpr-cookie-middle">\n          <a href="#" id="brink-gdpr-cookie-accept">I Accept</a>\n        </div>\n        <div class="brink-gdpr-cookie-right">\n          <a href="#" id="brink-gdpr-cookie-close"></a>\n          <span></span>\n        </div>\n      </div>\n    </div>\n    <script type="text/javascript">\n      window.brink_gdpr_whitelist = [];\n    </script>\n    <script data-cfasync="false" type="text/javascript">if (window.addthis_product === undefined) { window.addthis_product = "wpp"; } if (window.wp_product_version === undefined) { window.wp_product_version = "wpp-6.1.8"; } if (window.wp_blog_version === undefined) { window.wp_blog_version = "5.0.3"; } if (window.addthis_share === undefined) { window.addthis_share = {}; } if (window.addthis_config === undefined) { window.addthis_config = {"data_track_clickback":true,"ui_atversion":"300"}; } if (window.addthis_plugin_info === undefined) { window.addthis_plugin_info = {"info_status":"enabled","cms_name":"WordPress","plugin_name":"Share Buttons by AddThis","plugin_version":"6.1.8","plugin_mode":"AddThis","anonymous_profile_id":"wp-f82eef1f055e09dcb6273ac5462e3333","page_info":{"template":"pages","post_type":""},"sharing_enabled_on_post_via_metabox":false}; } \n                    (function() {\n                      var first_load_interval_id = setInterval(function () {\n                        if (typeof window.addthis !== \'undefined\') {\n                          window.clearInterval(first_load_interval_id);\n                          if (typeof window.addthis_layers !== \'undefined\' && Object.getOwnPropertyNames(window.addthis_layers).length > 0) {\n                            window.addthis.layers(window.addthis_layers);\n                          }\n                          if (Array.isArray(window.addthis_layers_tools)) {\n                            for (i = 0; i < window.addthis_layers_tools.length; i++) {\n                              window.addthis.layers(window.addthis_layers_tools[i]);\n                            }\n                          }\n                        }\n                     },1000)\n                    }());\n                </script> <script data-cfasync="false" type="text/javascript" src="https://s7.addthis.com/js/300/addthis_widget.js#pubid=ra-5a0206a251dc7e59" async="async"></script><script type=\'text/javascript\'>\n/* <![CDATA[ */\nvar countVars = {"disqusShortname":"montgomeryparks"};\n/* ]]> */\n</script>\n<script type=\'text/javascript\' src=\'/wp-content/plugins/disqus-comment-system/public/js/comment_count.js?ver=3.0.16\'></script>\n<script type=\'text/javascript\' src=\'/wp-content/themes/mdparks/dist/js/main.js?ver=641\'></script>\n<script type=\'text/javascript\' src=\'https://js.arcgis.com/3.18/?ver=5.0.3\'></script>\n<script type=\'text/javascript\' src=\'/wp-includes/js/wp-embed.min.js?ver=5.0.3\'></script>\n<script>\r\n  window.addEventListener(\'load\',function(){\r\n    jQuery(\'[href*=".pdf"]\').click(function(){\r\n      __gaTracker(\'send\',\'event\',\'PDF\',\'click\',jQuery(this).attr(\'href\'));\r\n    })\r\n  });\r\n</script>\r\n<img height="1" width="1" style="border-style:none;" alt="" src="//googleads.g.doubleclick.net/pagead/viewthroughconversion/974628909/?guid=ON&amp;script=0"/>\r\n\r\n</body>\r\n\r\n</html>\r\n<!-- AWS -->\r\n'
 
-open_house_event = {'Event Start Date': 'Sat. March 23rd, 2019', 'Event Start Time': '9:00am', 'Event End Time': '3:00pm', 'Event Website': 'https://www.montgomeryparks.org/events/events/volunteer-fair-for-montgomery-parks-historic-sites/', 'Event Name': 'Join Parks Historic Sites Volunteer Program', 'Event Venue Name': '', 'Event Cost': '', 'Event Description': 'Help bring history to life! Share your love of history, the great outdoors, agriculture & old buildings with visitors of all ages and become a trained Montgomery Parks volunteer tour guide. Parks needs your help at these historic sites: Oakley Cabin African American Museum (Olney), Josiah Henson Park (North Bethesda), Woodlawn Manor Cultural Park (Sandy Spring), Kingsley Schoolhouse (Clarksburg) and the Agricultural History Farm Park (Derwood).', 'Event Category': 'open house'}
+open_house_event = [{'Event Start Date': '2019-03-23', 
+'Event End Date': '2019-03-23',
+'Event Start Time': '09:00:00', 'Event End Time': '15:00:00', 'Event Website': 'https://www.montgomeryparks.org/events/volunteer-fair-for-montgomery-parks-historic-sites/', 'Event Name': 'Join Parks Historic Sites Volunteer Program', 'Event Venue Name': '', 'Event Cost': '', 'Event Description': 'Help bring history to life! Share your love of history, the great outdoors, agriculture & old buildings with visitors of all ages and become a trained Montgomery Parks volunteer tour guide. Parks needs your help at these historic sites: Oakley Cabin African American Museum (Olney), Josiah Henson Park (North Bethesda), Woodlawn Manor Cultural Park (Sandy Spring), Kingsley Schoolhouse (Clarksburg) and the Agricultural History Farm Park (Derwood).', 'Event Category': 'open house','Event Currency Symbol':'$','Timezone': 'America/New_York',
+'Event Organizers': '','All Day Event':False}]
 
 single_event_calendar_page_content = b'<!DOCTYPE html>\r\n<html lang="en-US" xmlns:fb="https://www.facebook.com/2008/fbml" xmlns:addthis="https://www.addthis.com/help/api-spec"  prefix="og: http://ogp.me/ns#" class="no-js">\r\n\t<head>\r\n\t\t<meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">\r\n\t\t<script src="https://use.typekit.net/bxi2fgb.js"></script>\r\n\t\t<script>try{Typekit.load({ async: true });}catch(e){}</script>\r\n\t\t<meta charset="UTF-8">\r\n\t\t<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=12.0, minimum-scale=.25, user-scalable=yes"/>\r\n\t\t<!-- CHECK META CONTENT -->\r\n\t\t<title>Calendar - Montgomery Parks</title>\r\n\r\n        <script>\r\n          // helpful for understanding dojoConfig.packages vs. dojoConfig.paths:\r\n          // http://www.sitepen.com/blog/2013/06/20/dojo-faq-what-is-the-difference-packages-vs-paths-vs-aliases/\r\n          var dojoConfig = {\r\n            paths: {\r\n              extras: "/wp-content/themes/mdparks/dist/js/extras"\r\n            }\r\n          };\r\n        </script>\r\n\r\n\t\t<link rel="pingback" href="/xmlrpc.php">\r\n\t\t\t\t<style>#skipNav{display: block;background-color:#fff;padding:1em;position: absolute;top: 0;left: -9999px;}#skipNav:focus{left:0;}</style>\r\n\r\n\t\t\n<!-- This site is optimized with the Yoast SEO plugin v9.5 - https://yoast.com/wordpress/plugins/seo/ -->\n<link rel="canonical" href="/calendar/" />\n<meta property="og:locale" content="en_US" />\n<meta property="og:type" content="article" />\n<meta property="og:title" content="Calendar - Montgomery Parks" />\n<meta property="og:url" content="/calendar/" />\n<meta property="og:site_name" content="Montgomery Parks" />\n<meta name="twitter:card" content="summary" />\n<meta name="twitter:title" content="Calendar - Montgomery Parks" />\n<!-- / Yoast SEO plugin. -->\n\n<link rel=\'dns-prefetch\' href=\'//js.arcgis.com\' />\n<link rel=\'dns-prefetch\' href=\'//s.w.org\' />\n<link rel=\'dns-prefetch\' href=\'//s3.amazonaws.com\' />\n<link rel="alternate" type="application/rss+xml" title="Montgomery Parks &raquo; Feed" href="/feed/" />\n<link rel="alternate" type="application/rss+xml" title="Montgomery Parks &raquo; Comments Feed" href="/comments/feed/" />\n<!-- This site uses the Google Analytics by MonsterInsights plugin v7.4.2 - Using Analytics tracking - https://www.monsterinsights.com/ -->\r\n<script type="text/javascript" data-cfasync="false">\r\n\tvar mi_version         = \'7.4.2\';\r\n\tvar mi_track_user      = true;\r\n\tvar mi_no_track_reason = \'\';\r\n\t\r\n\tvar disableStr = \'ga-disable-UA-493554-2\';\r\n\r\n\t/* Function to detect opted out users */\r\n\tfunction __gaTrackerIsOptedOut() {\r\n\t\treturn document.cookie.indexOf(disableStr + \'=true\') > -1;\r\n\t}\r\n\r\n\t/* Disable tracking if the opt-out cookie exists. */\r\n\tif ( __gaTrackerIsOptedOut() ) {\r\n\t\twindow[disableStr] = true;\r\n\t}\r\n\r\n\t/* Opt-out function */\r\n\tfunction __gaTrackerOptout() {\r\n\t  document.cookie = disableStr + \'=true; expires=Thu, 31 Dec 2099 23:59:59 UTC; path=/\';\r\n\t  window[disableStr] = true;\r\n\t}\r\n\t\r\n\tif ( mi_track_user ) {\r\n\t\t(function(i,s,o,g,r,a,m){i[\'GoogleAnalyticsObject\']=r;i[r]=i[r]||function(){\r\n\t\t\t(i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),\r\n\t\t\tm=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)\r\n\t\t})(window,document,\'script\',\'//www.google-analytics.com/analytics.js\',\'__gaTracker\');\r\n\r\n\t\t__gaTracker(\'create\', \'UA-493554-2\', \'auto\');\n\t\t__gaTracker(\'set\', \'forceSSL\', true);\n\t\t__gaTracker(\'set\', \'anonymizeIp\', true);\n\t\t__gaTracker(\'require\', \'linkid\', \'linkid.js\');\n\t\t__gaTracker(\'send\',\'pageview\');\n\t} else {\r\n\t\tconsole.log( "" );\r\n\t\t(function() {\r\n\t\t\t/* https://developers.google.com/analytics/devguides/collection/analyticsjs/ */\r\n\t\t\tvar noopfn = function() {\r\n\t\t\t\treturn null;\r\n\t\t\t};\r\n\t\t\tvar noopnullfn = function() {\r\n\t\t\t\treturn null;\r\n\t\t\t};\r\n\t\t\tvar Tracker = function() {\r\n\t\t\t\treturn null;\r\n\t\t\t};\r\n\t\t\tvar p = Tracker.prototype;\r\n\t\t\tp.get = noopfn;\r\n\t\t\tp.set = noopfn;\r\n\t\t\tp.send = noopfn;\r\n\t\t\tvar __gaTracker = function() {\r\n\t\t\t\tvar len = arguments.length;\r\n\t\t\t\tif ( len === 0 ) {\r\n\t\t\t\t\treturn;\r\n\t\t\t\t}\r\n\t\t\t\tvar f = arguments[len-1];\r\n\t\t\t\tif ( typeof f !== \'object\' || f === null || typeof f.hitCallback !== \'function\' ) {\r\n\t\t\t\t\tconsole.log( \'Not running function __gaTracker(\' + arguments[0] + " ....) because you are not being tracked. " + mi_no_track_reason );\r\n\t\t\t\t\treturn;\r\n\t\t\t\t}\r\n\t\t\t\ttry {\r\n\t\t\t\t\tf.hitCallback();\r\n\t\t\t\t} catch (ex) {\r\n\r\n\t\t\t\t}\r\n\t\t\t};\r\n\t\t\t__gaTracker.create = function() {\r\n\t\t\t\treturn new Tracker();\r\n\t\t\t};\r\n\t\t\t__gaTracker.getByName = noopnullfn;\r\n\t\t\t__gaTracker.getAll = function() {\r\n\t\t\t\treturn [];\r\n\t\t\t};\r\n\t\t\t__gaTracker.remove = noopfn;\r\n\t\t\twindow[\'__gaTracker\'] = __gaTracker;\r\n\t\t\t\t\t})();\r\n\t\t}\r\n</script>\r\n<!-- / Google Analytics by MonsterInsights -->\r\n<link rel=\'stylesheet\' id=\'wp-block-library-group-css\' href=\'/wp-content/plugins/bwp-minify/min/?f=wp-includes/css/dist/block-library/style.min.css,wp-content/themes/mdparks/dist/css/main.min.css,wp-content/plugins/addthis/frontend/build/addthis_wordpress_public.min.css\' type=\'text/css\' media=\'all\' />\r\n<link rel=\'stylesheet\' id=\'brink_gdpr_css-group-css\' href=\'/wp-content/plugins/bwp-minify/min/?f=wp-content/plugins/brink-gdpr/css/brink-gdpr.css\' type=\'text/css\' media=\'screen\' />\r\n<link rel=\'stylesheet\' id=\'esri-css-css\'  href=\'https://js.arcgis.com/3.18/esri/css/esri.css?ver=5.0.3\' type=\'text/css\' media=\'all\' />\n<script type=\'text/javascript\'>\n/* <![CDATA[ */\nvar monsterinsights_frontend = {"js_events_tracking":"true","download_extensions":"doc,exe,js,pdf,ppt,tgz,zip,xls","inbound_paths":"[]","home_url":"http:\\/\\/staging.montgomeryparks.org","hash_tracking":"false"};\n/* ]]> */\n</script>\n<script type=\'text/javascript\' src=\'/wp-content/plugins/google-analytics-for-wordpress/assets/js/frontend.min.js?ver=7.4.2\'></script>\n<script type=\'text/javascript\' src=\'/wp-content/plugins/brink-gdpr/js/brink_gdpr.js?ver=5.0.3\'></script>\n<link rel=\'https://api.w.org/\' href=\'/wp-json/\' />\n<link rel="EditURI" type="application/rsd+xml" title="RSD" href="/xmlrpc.php?rsd" />\n<link rel="wlwmanifest" type="application/wlwmanifest+xml" href="/wp-includes/wlwmanifest.xml" /> \n<meta name="generator" content="WordPress 5.0.3" />\n<link rel=\'shortlink\' href=\'/?p=32\' />\n<link rel="alternate" type="application/json+oembed" href="/wp-json/oembed/1.0/embed?url=http%3A%2F%2Fstaging.montgomeryparks.org%2Fcalendar%2F" />\n<link rel="alternate" type="text/xml+oembed" href="/wp-json/oembed/1.0/embed?url=http%3A%2F%2Fstaging.montgomeryparks.org%2Fcalendar%2F&#038;format=xml" />\n<script type="text/javascript">\nvar ajaxurl = \'/wp-admin/admin-ajax.php/\';\n</script>\n\t\t<style type="text/css">.recentcomments a{display:inline !important;padding:0 !important;margin:0 !important;}</style>\n\t\t<link rel="icon" href="/uploads/2016/07/cropped-favicon-490x490.png" sizes="32x32" />\n<link rel="icon" href="/uploads/2016/07/cropped-favicon-490x490.png" sizes="192x192" />\n<link rel="apple-touch-icon-precomposed" href="/uploads/2016/07/cropped-favicon-490x490.png" />\n<meta name="msapplication-TileImage" content="/uploads/2016/07/cropped-favicon-490x490.png" />\n\r\n\t\t\t</head>\r\n\r\n\t<body class="page-template page-template-page-calendar page-template-page-calendar-php page page-id-32 group-blog">\r\n\r\n\r\n\r\n\t\t<div id="page" class="page">\r\n\t\t\t<div class="main-header parks-header">\r\n\t\t    <div class="mncppc">\r\n\t\t        <a href="http://www.mncppc.org" title="Maryland-National Capital Parks and Commission" class="mncppc-link" target="_blank">Maryland-National Capital Park and Planning Commission</a>\r\n\t\t        <ul class="sites-links">\r\n\t\t\t\t\t<li class="parks current"><a href="/parks-home">Parks</a></li>\r\n\t\t            <li class="planning"><a href="http://montgomeryplanning.org" target="_blank">Planning</a></li>\r\n\t\t            <li class="board"><a href="http://montgomeryplanningboard.org" target="_blank">Planning Board</a></li>\r\n\t\t        </ul>\r\n\t\t    </div>\r\n\t\t    <div class="branding">\r\n\t\t    \t<div class="col col-tab-9 col-sm-desk-6 col-lg-5 left-wrap p-x">\r\n\t\t    \t\t<div class="mid-v">\r\n\t\t\t\t        <div class="site-icon">\r\n\t\t\t\t        \t<a href="/" class="parks-logo t-indent" title="Back to Home">Montgomery Parks</a>\r\n\t\t\t\t        </div>\r\n\t\t\t\t        <div class="site-title">\r\n\t\t\t\t        \t<a href="/" class="parks-logo" title="Back to Home">\r\n\t\t\t\t        \t<div class="title-wrap">\r\n\t\t\t\t\t            <p class="montco">Montgomery County, MD</p>\r\n\t\t\t\t\t            <h2 class="name">Montgomery Parks</h2>\r\n\t\t\t\t            </div>\r\n\t\t\t\t            </a>\r\n\t\t\t\t        </div>\r\n\t\t\t        </div>\r\n\t\t        </div>\r\n\t\t\t\t<div class="col col-tab-3 col-desk-2 viewing-options">\r\n\t\t\t\t<!-- \t<div class="head-col-wrap"> -->\r\n\t\t\t\t\t\t<div class="options-wrap">\r\n\t\t\t\t\t\t\t<div id="google_translate_element" tabindex="-1"></div>\r\n\t\t\t\t\t\t\t<script type="text/javascript">\r\n                            function fireEvent(element,event){\r\n                                if (document.createEventObject){\r\n                                    // dispatch for IE\r\n                                    var evt = document.createEventObject();\r\n                                    return element.fireEvent(\'on\'+event,evt)\r\n                                }\r\n                                else{\r\n                                    // dispatch for firefox + others\r\n                                    var evt = document.createEvent("HTMLEvents");\r\n                                    evt.initEvent(event, true, true ); // event type,bubbling,cancelable\r\n                                    return !element.dispatchEvent(evt);\r\n                                }\r\n                             }\r\n\r\n                            function googleTranslateElementInit() {\r\n                                new google.translate.TranslateElement({pageLanguage: \'en\', includedLanguages: \'af,ach,ak,am,ar,az,be,bem,bg,bh,bn,br,bs,ca,chr,ckb,co,crs,cs,cy,da,de,ee,el,en,eo,es,es-419,et,eu,fa,fi,fo,fr,fy,ga,gaa,gd,gl,gn,gu,ha,haw,hi,hr,ht,hu,hy,ia,id,ig,is,it,iw,ja,jw,ka,kg,kk,km,kn,ko,kri,ku,ky,la,lg,ln,lo,loz,lt,lua,lv,mfe,mg,mi,mk,ml,mn,mo,mr,ms,mt,ne,nl,nn,no,nso,ny,nyn,oc,om,or,pa,pcm,pl,ps,pt-BR,pt-PT,qu,rm,rn,ro,ru,rw,sd,sh,si,sk,sl,sn,so,sq,sr,sr-ME,st,su,sv,sw,ta,te,tg,th,ti,tk,tl,tn,to,tr,tt,tum,tw,ug,uk,ur,uz,vi,wo,xh,xx-bork,xx-elmer,xx-hacker,xx-klingon,xx-pirate,yi,yo,zh-CN,zh-TW,zu\', layout: google.translate.TranslateElement.InlineLayout}, \'google_translate_element\');\r\n                                setTimeout(function () {\r\n                                    $(\'select.goog-te-combo\').selectize({\r\n                                        onChange: function (val) {\r\n                                            $(\'select.goog-te-combo\').val(val);\r\n                                            fireEvent($(\'select.goog-te-combo\').get(0), \'change\');\r\n                                        }\r\n                                    });\r\n                                    // $(\'#google_translate_element .selectize-input > input\').attr(\'id\', \'google_translate_element\');\r\n                                    // $(\'#google_translate_element .selectize-input > input\').wrap("<label for=\'google_translate_element\'></label>");\r\n                                }, 1000);\r\n                            }\r\n\t\t\t\t\t\t\t</script>\r\n\t\t\t\t\t\t\t<script type="text/javascript" src="//translate.google.com/translate_a/element.js?cb=googleTranslateElementInit"></script>\r\n\t\t\t\t\t\t\t<span class=\'divide-line\'>|</span> <div class="open-wrap"><a class="modal-open" href="#">Text Size</a></div>\r\n\t\t\t\t\t\t</div>\r\n<!-- \t\t\t\t\t</div>\r\n -->\t\t\t\t</div>\r\n\t\t\t\t<div class="mob-hide col-desk-2 col-lg-3 rel">\r\n\t\t\t\t\t<div class="neat-search">\r\n                        <form action="/">\r\n                        \t<label for="search_2">Search the Site</label>\r\n    \t\t\t\t\t\t<input id="search_2" class="closed" type="text" name="s" placeholder="Search" />\r\n    \t\t\t\t\t\t<button type="submit" class="search-link t-indent">Search</button>\r\n                        </form>\r\n\t\t\t\t\t</div>\r\n\t\t\t\t</div>\r\n\t\t\t\t<div class="social col col-sm-desk-3 col-desk-2 rel mob-hide">\r\n\t\t\t\t\t<div class="share">\n\t<a href="https://www.facebook.com/MontgomeryParks" target="_blank">facebook<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1fb{fill:#547bbc}.cls-2{fill:#fff}</style></defs><title>social-fb</title><path class="cls-1fb" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M11.4 12.5H13V11a4 4 0 0 1 .5-2.3 2.8 2.8 0 0 1 2.4-1.1 9.7 9.7 0 0 1 2.8.3l-.4 2.3-1.3-.3a.9.9 0 0 0-1.1.8v1.8h2.5l-.2 2.2h-2.4v7.8h-2.9v-7.8h-1.5v-2.2z"/></svg></a>\n\t<a href="https://twitter.com/MontgomeryParks" target="_blank">twitter<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1tw{fill:#5ea8dd}.cls-2{fill:#fff}</style></defs><title>social-tw</title><path class="cls-1tw" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M22.5 11.1l-1.8.5a3.1 3.1 0 0 0 1.4-1.7 6.2 6.2 0 0 1-2 .7 3.1 3.1 0 0 0-5.3 2.8 8.8 8.8 0 0 1-6.4-3.2 3.1 3.1 0 0 0 1 4.1 3.1 3.1 0 0 1-1.4-.4 3.1 3.1 0 0 0 2.5 3H9.2a3.1 3.1 0 0 0 2.9 2.1 6.2 6.2 0 0 1-3.8 1.3h-.8A8.8 8.8 0 0 0 21 13.1v-.4a6.2 6.2 0 0 0 1.5-1.6"/></svg></a>\n\t<a href="https://www.instagram.com/montgomeryparks/" target="_blank">instagram<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 291.319 291.319"><path d="M145.659 0c80.44 0 145.66 65.219 145.66 145.66S226.1 291.319 145.66 291.319 0 226.1 0 145.66 65.21 0 145.659 0z" fill="#9b6954"/><path d="M195.93 63.708H95.38c-17.47 0-31.672 14.211-31.672 31.672v100.56c0 17.47 14.211 31.672 31.672 31.672h100.56c17.47 0 31.672-14.211 31.672-31.672V95.38c-.001-17.461-14.212-31.672-31.682-31.672zm9.978 18.326l3.587-.009v27.202l-27.402.091-.091-27.202 23.906-.082zm-60.248 36.205c22.732 0 27.42 21.339 27.42 27.429 0 15.103-12.308 27.411-27.42 27.411-15.121 0-27.42-12.308-27.42-27.411-.01-6.09 4.688-27.429 27.42-27.429zm63.99 75.716c0 8.658-7.037 15.704-15.713 15.704H97.073c-8.667 0-15.713-7.037-15.713-15.704v-66.539h22.759c-2.112 5.198-3.305 12.299-3.305 18.253 0 24.708 20.101 44.818 44.818 44.818s44.808-20.11 44.808-44.818c0-5.954-1.193-13.055-3.296-18.253h22.486v66.539h.02z" fill="#fff"/></svg></a>\n\t<a href="https://www.youtube.com/user/MontgomeryParks" target="_blank">youtube<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1yt{fill:#e9644a}.cls-2{fill:#fff}</style></defs><title>social-yt</title><path class="cls-1yt" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M17.9 15L13 17.9v-5.8zm4.6 2.7v-5.4a2.5 2.5 0 0 0-2.6-2.6h-9.8a2.5 2.5 0 0 0-2.6 2.6v5.4a2.5 2.5 0 0 0 2.6 2.6h9.8a2.5 2.5 0 0 0 2.6-2.6"/></svg></a>\n\t<a href="http://www.montgomeryplanningboard.org/rss/rss.shtm" target="_blank">rss<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1rss{fill:#f29100}.cls-2{fill:#fff}</style></defs><title>social-rss</title><path class="cls-1rss" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M10.9 19.1a2 2 0 0 1 0 2.8 2 2 0 1 1-2.8-2.8 2 2 0 0 1 2.8 0zm-3.4-6.5v2.9a7 7 0 0 1 7 7h2.9a9.9 9.9 0 0 0-9.9-9.9zm0-5.1v2.9a12.2 12.2 0 0 1 12.1 12.1h2.9a15 15 0 0 0-15-15z"/></svg></a>\n</div>\t\t\t\t</div>\r\n\r\n\t\t    </div>\r\n\t\t    <div class="nav-primary count-7">\r\n\t\t    \t<div class="icon-wrap rel">\r\n\t\t\t        <div class="ico-link">\r\n\t\t\t        \t<a href="/" class="ico ico-home" title="Back to Home">Back to Home<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 31.9 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-home</title><path class="cls-1" d="M26.7 14.7a.9.9 0 0 0-.9.9v12.6h-6.5v-5.1a.9.9 0 0 0-.9-.9h-4.7a.9.9 0 0 0-.9.9v5.1H6.2V16a.9.9 0 0 0-1.8 0v13.1a.9.9 0 0 0 .9.9h8.4a.9.9 0 0 0 .9-.9V24h2.9v5.1a.9.9 0 0 0 .9.9h8.3a.9.9 0 0 0 .9-.9V15.6a.9.9 0 0 0-.9-.9zm0 0l5 .7-7.3-7.3v-5a.9.9 0 0 0-.9-.9.9.9 0 0 0-.9.9v3.2l-6-6A.9.9 0 0 0 16 0a.9.9 0 0 0-.7.3l-15 15a.9.9 0 0 0 0 1.3.9.9 0 0 0 1.3 0L16 2.2l14.4 14.5a.9.9 0 0 0 .7.2l.7-.2a1 1 0 0 0 0-1.3zm0 0"/></svg></a>\r\n\t\t\t        </div>\r\n\t\t\t        <div class="ico-link">\r\n\t\t\t        \t<a href="/calendar/?v=1" class="ico ico-calendar" title="Calendar of Events">Calendar of Events<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-calendar</title><path class="cls-1" d="M27.5 2.8h-4.1V.9a.9.9 0 0 0-1.9 0v1.9h-5.6V.9a.9.9 0 0 0-1.9 0v1.9H8.4V.9a.9.9 0 1 0-1.8 0v1.9H2.5A2.5 2.5 0 0 0 0 5.3v22.2A2.5 2.5 0 0 0 2.5 30h25a2.5 2.5 0 0 0 2.5-2.5V5.3a2.5 2.5 0 0 0-2.5-2.5zm.6 24.7a.6.6 0 0 1-.6.6h-25a.6.6 0 0 1-.6-.6V5.3a.6.6 0 0 1 .6-.6h4.1v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h4.1a.6.6 0 0 1 .6.6v22.2zm0 0"/><path class="cls-1" d="M6.6 11.3h3.7v2.81H6.6zm0 4.6h3.7v2.81H6.6zm0 4.7h3.7v2.81H6.6zm6.5 0h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7zm6.6 9.3h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7z"/></svg></a>\r\n\t\t\t        </div>\r\n\t\t\t        <div class="ico-link">\r\n\t\t\t        \t<a href="/news/" class="ico ico-news" title="Latest News">Latest News<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30.7 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-news</title><path class="cls-1" d="M30 2.5a.8.8 0 0 0-.8.8v21.1a4 4 0 0 1-4 4H5.7a4 4 0 0 1-4-4V1.6h23.2v20.5a.8.8 0 1 0 1.6 0V.8a.8.8 0 0 0-.8-.8H.8a.8.8 0 0 0-.8.8v23.6A5.6 5.6 0 0 0 5.7 30h19.4a5.6 5.6 0 0 0 5.6-5.6V3.3a.8.8 0 0 0-.8-.8zm0 0"/><path class="cls-1" d="M21.9 6.2a.8.8 0 0 0 0-1.6H4.5a.8.8 0 1 0 0 1.6h17.4zm0 14.6H4.5a.8.8 0 1 0 0 1.6h17.4a.8.8 0 0 0 0-1.6zm0 4.1H7.3a.8.8 0 0 0 0 1.6h14.6a.8.8 0 0 0 0-1.6zM3.5 9v8.6a.8.8 0 0 0 .8.8h8.6a.8.8 0 0 0 .8-.8V9a.8.8 0 0 0-.8-.8H4.3a.8.8 0 0 0-.8.8zm1.6.8h7v7h-7v-7zM22.3 9h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 0-1.6zm0 3.6h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 1 0 0-1.6zm.8 4.6a.8.8 0 0 0-.8-.8h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 .8-.8zm0 0"/></svg></a>\r\n\t\t\t        </div>\r\n\t\t        </div>\r\n\t\t        <div class="menu-primary-container"><ul id="menu-primary" class="menu"><li id="menu-item-2177" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-2177"><a href="/activities/">Things To Do</a>\n<ul class="sub-menu">\n\t<li id="menu-item-664" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-664"><a href="/activities/">Activities</a></li>\n\t<li id="menu-item-8304" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-8304"><a href="/classes-camps/">Classes &#038; Camps</a></li>\n\t<li id="menu-item-950" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-950"><a href="/special-events/">Special Events &#038; Festivals</a></li>\n</ul>\n</li>\n<li id="menu-item-853" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-853"><a href="/parks-directory/">Parks &#038; Trails Directory</a>\n<ul class="sub-menu">\n\t<li id="menu-item-1089" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1089"><a href="/parks-trails/parks/">Parks</a></li>\n\t<li id="menu-item-39673" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-39673"><a href="https://www.montgomeryparks.org/activities/park-trails/">Trails</a></li>\n\t<li id="menu-item-1092" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1092"><a href="/parks-trails/gardens/">Gardens</a></li>\n\t<li id="menu-item-2941" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2941"><a href="/parks-trails/facilities/">Facilities</a></li>\n\t<li id="menu-item-1090" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1090"><a href="/parks-trails/nature-centers/">Nature Centers</a></li>\n\t<li id="menu-item-2360" class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2360"><a href="/parks-trails/event-centers/">Event Centers</a></li>\n</ul>\n</li>\n<li id="menu-item-837" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-837"><a href="/services/">Permits, Rentals &#038; Requests</a>\n<ul class="sub-menu">\n\t<li id="menu-item-2850" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2850"><a href="/services/permits-rentals/">Permits &#038; Rentals</a></li>\n\t<li id="menu-item-838" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-838"><a href="/services/report-a-problem/">Parks Information &#038; Customer Service</a></li>\n</ul>\n</li>\n<li id="menu-item-839" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-839"><a href="/caring-for-our-parks/">Caring for Our Parks</a>\n<ul class="sub-menu">\n\t<li id="menu-item-840" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-840"><a href="/caring-for-our-parks/culture-heritage/">Culture &#038; Heritage</a></li>\n\t<li id="menu-item-842" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-842"><a href="/caring-for-our-parks/natural-spaces/">Natural Spaces</a></li>\n\t<li id="menu-item-959" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-959"><a href="/caring-for-our-parks/parks-and-trails/">Park and Trail Planning</a></li>\n\t<li id="menu-item-31505" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-31505"><a href="/caring-for-our-parks/sustainability/">Sustainability</a></li>\n\t<li id="menu-item-844" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-844"><a href="/caring-for-our-parks/wildlife/">Living with Wildlife</a></li>\n</ul>\n</li>\n<li id="menu-item-122" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-122"><a href="/projects/">Parks &#038; Projects</a>\n<ul class="sub-menu">\n\t<li id="menu-item-26054" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-26054"><a href="/projects/">Projects</a></li>\n\t<li id="menu-item-2503" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2503"><a href="/projects/capital-improvements-program/">Funding</a></li>\n\t<li id="menu-item-16631" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-16631"><a href="/projects/public-input/">Open Town Hall</a></li>\n\t<li id="menu-item-912" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-912"><a href="/projects/bids-procurement/">Bids / Procurement</a></li>\n\t<li id="menu-item-897" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-897"><a href="/projects/resources/">Resources</a></li>\n</ul>\n</li>\n<li id="menu-item-123" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-123"><a href="/support/">Support the Parks</a>\n<ul class="sub-menu">\n\t<li id="menu-item-19531" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-19531"><a href="/support/advocacy/">Advocacy</a></li>\n\t<li id="menu-item-907" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-907"><a href="/support/donate/">Donate</a></li>\n\t<li id="menu-item-15740" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-15740"><a href="/support/partnerships/">Partnerships</a></li>\n\t<li id="menu-item-900" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-900"><a href="/support/sponsorships/">Sponsorships</a></li>\n\t<li id="menu-item-901" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-901"><a href="/support/volunteer/">Volunteer</a></li>\n</ul>\n</li>\n<li id="menu-item-1047" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-1047"><a href="/about/parks/">About</a>\n<ul class="sub-menu">\n\t<li id="menu-item-660" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-660"><a href="/about/parks/">About the Parks</a></li>\n\t<li id="menu-item-659" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-659"><a href="/about/divisions/">Divisions &#038; Offices</a></li>\n\t<li id="menu-item-30188" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-30188"><a href="/about/media-center/">Media Center</a></li>\n\t<li id="menu-item-2592" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2592"><a href="/about/park-police/">Police</a></li>\n\t<li id="menu-item-655" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-655"><a href="/about/employment/">Employment</a></li>\n\t<li id="menu-item-932" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-932"><a href="/about/contact-us/">Contact Us</a></li>\n</ul>\n</li>\n</ul></div>\t\t    </div>\r\n\t\t</div>\r\n\t\t<div id="modal">\r\n\t    \t<div class="close">\xc3\x97</div>\r\n\t\t\t<div class="modal-wrapper">\r\n\t\t\t\t<div class="close">\xc3\x97</div>\r\n\t\t\t\tTo change the text size you can use your web browser\'s settings. Most browsers include functionality to let you increase or decrease the text on a web page. For example, to increase text size using:<br><br>\r\n\r\n\t\t\t\t<strong>Chrome</strong><br>\r\n\r\n\t\t\t\tIn the menu to the right of the address bar, select and set Zoom level. Menu > Zoom > +<br><br>\r\n\r\n\t\t\t\t<strong>Firefox</strong><br>\r\n\r\n\t\t\t\tIn the View menu, select Zoom. View > Zoom > Zoom In<br><br>\r\n\r\n\t\t\t\t<strong>Internet Explorer</strong><br>\r\n\r\n\t\t\t\tIn the View menu, select Text Size. View > Text Size > Largest<br><br>\r\n\r\n\t\t\t\t<strong>Safari</strong><br>\r\n\r\n\t\t\t\tIn the View menu, select Zoom In. View > Zoom In\r\n\t\t\t\tMacintosh Shortcut: Command+\r\n\t\t\t</div>\r\n\t\t</div>\r\n\t\t<div class="mobile-nav count-7">\r\n\t\t\t<div class="just-click"></div>\r\n\t\t\t<div class="upper-half">\r\n\t\t\t\t<div class="menu-primary-container"><ul id="menu-primary-1" class="menu"><li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-2177"><a href="/activities/">Things To Do</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-664"><a href="/activities/">Activities</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-8304"><a href="/classes-camps/">Classes &#038; Camps</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-950"><a href="/special-events/">Special Events &#038; Festivals</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-853"><a href="/parks-directory/">Parks &#038; Trails Directory</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1089"><a href="/parks-trails/parks/">Parks</a></li>\n\t<li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-39673"><a href="https://www.montgomeryparks.org/activities/park-trails/">Trails</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1092"><a href="/parks-trails/gardens/">Gardens</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2941"><a href="/parks-trails/facilities/">Facilities</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-1090"><a href="/parks-trails/nature-centers/">Nature Centers</a></li>\n\t<li class="menu-item menu-item-type-taxonomy menu-item-object-park_category menu-item-2360"><a href="/parks-trails/event-centers/">Event Centers</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-837"><a href="/services/">Permits, Rentals &#038; Requests</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2850"><a href="/services/permits-rentals/">Permits &#038; Rentals</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-838"><a href="/services/report-a-problem/">Parks Information &#038; Customer Service</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-839"><a href="/caring-for-our-parks/">Caring for Our Parks</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-840"><a href="/caring-for-our-parks/culture-heritage/">Culture &#038; Heritage</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-842"><a href="/caring-for-our-parks/natural-spaces/">Natural Spaces</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-959"><a href="/caring-for-our-parks/parks-and-trails/">Park and Trail Planning</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-31505"><a href="/caring-for-our-parks/sustainability/">Sustainability</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-844"><a href="/caring-for-our-parks/wildlife/">Living with Wildlife</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-122"><a href="/projects/">Parks &#038; Projects</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-26054"><a href="/projects/">Projects</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2503"><a href="/projects/capital-improvements-program/">Funding</a></li>\n\t<li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-16631"><a href="/projects/public-input/">Open Town Hall</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-912"><a href="/projects/bids-procurement/">Bids / Procurement</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-897"><a href="/projects/resources/">Resources</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-has-children menu-item-123"><a href="/support/">Support the Parks</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-19531"><a href="/support/advocacy/">Advocacy</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-907"><a href="/support/donate/">Donate</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-15740"><a href="/support/partnerships/">Partnerships</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-900"><a href="/support/sponsorships/">Sponsorships</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-901"><a href="/support/volunteer/">Volunteer</a></li>\n</ul>\n</li>\n<li class="menu-item menu-item-type-custom menu-item-object-custom menu-item-has-children menu-item-1047"><a href="/about/parks/">About</a>\n<ul class="sub-menu">\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-660"><a href="/about/parks/">About the Parks</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-659"><a href="/about/divisions/">Divisions &#038; Offices</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-30188"><a href="/about/media-center/">Media Center</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-2592"><a href="/about/park-police/">Police</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-655"><a href="/about/employment/">Employment</a></li>\n\t<li class="menu-item menu-item-type-post_type menu-item-object-page menu-item-932"><a href="/about/contact-us/">Contact Us</a></li>\n</ul>\n</li>\n</ul></div>\t\t\t</div>\r\n\t\t\t<div class="lower-half rel">\r\n\t\t\t\t<div class="neat-search">\r\n\t\t\t\t\t<label for="m_search" />\r\n\t\t\t\t\t<input id="m_search" class="closed" type="text" placeholder="Search" />\r\n\t\t\t\t\t<a class="search-link t-indent" href="#m_search">Search</a>\r\n\t\t\t\t</div>\r\n\t\t\t\t<div class="page-nav">\r\n\t\t\t\t\t<a href="http://www.montgomeryparks.org/" target="_blank">Parks</a> &#8226;\r\n\t\t\t\t\t<a href="http://montgomeryplanning.org/" target="_blank">Planning</a> &#8226;\r\n\t\t\t\t\t<a href="http://montgomeryplanningboard.org/" target="_blank">Planning Board</a>\r\n\t\t\t\t</div>\r\n\t\t\t\t\t<div class="share">\n\t<a href="https://www.facebook.com/MontgomeryParks" target="_blank">facebook<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1fb{fill:#547bbc}.cls-2{fill:#fff}</style></defs><title>social-fb</title><path class="cls-1fb" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M11.4 12.5H13V11a4 4 0 0 1 .5-2.3 2.8 2.8 0 0 1 2.4-1.1 9.7 9.7 0 0 1 2.8.3l-.4 2.3-1.3-.3a.9.9 0 0 0-1.1.8v1.8h2.5l-.2 2.2h-2.4v7.8h-2.9v-7.8h-1.5v-2.2z"/></svg></a>\n\t<a href="https://twitter.com/MontgomeryParks" target="_blank">twitter<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1tw{fill:#5ea8dd}.cls-2{fill:#fff}</style></defs><title>social-tw</title><path class="cls-1tw" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M22.5 11.1l-1.8.5a3.1 3.1 0 0 0 1.4-1.7 6.2 6.2 0 0 1-2 .7 3.1 3.1 0 0 0-5.3 2.8 8.8 8.8 0 0 1-6.4-3.2 3.1 3.1 0 0 0 1 4.1 3.1 3.1 0 0 1-1.4-.4 3.1 3.1 0 0 0 2.5 3H9.2a3.1 3.1 0 0 0 2.9 2.1 6.2 6.2 0 0 1-3.8 1.3h-.8A8.8 8.8 0 0 0 21 13.1v-.4a6.2 6.2 0 0 0 1.5-1.6"/></svg></a>\n\t<a href="https://www.instagram.com/montgomeryparks/" target="_blank">instagram<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 291.319 291.319"><path d="M145.659 0c80.44 0 145.66 65.219 145.66 145.66S226.1 291.319 145.66 291.319 0 226.1 0 145.66 65.21 0 145.659 0z" fill="#9b6954"/><path d="M195.93 63.708H95.38c-17.47 0-31.672 14.211-31.672 31.672v100.56c0 17.47 14.211 31.672 31.672 31.672h100.56c17.47 0 31.672-14.211 31.672-31.672V95.38c-.001-17.461-14.212-31.672-31.682-31.672zm9.978 18.326l3.587-.009v27.202l-27.402.091-.091-27.202 23.906-.082zm-60.248 36.205c22.732 0 27.42 21.339 27.42 27.429 0 15.103-12.308 27.411-27.42 27.411-15.121 0-27.42-12.308-27.42-27.411-.01-6.09 4.688-27.429 27.42-27.429zm63.99 75.716c0 8.658-7.037 15.704-15.713 15.704H97.073c-8.667 0-15.713-7.037-15.713-15.704v-66.539h22.759c-2.112 5.198-3.305 12.299-3.305 18.253 0 24.708 20.101 44.818 44.818 44.818s44.808-20.11 44.808-44.818c0-5.954-1.193-13.055-3.296-18.253h22.486v66.539h.02z" fill="#fff"/></svg></a>\n\t<a href="https://www.youtube.com/user/MontgomeryParks" target="_blank">youtube<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1yt{fill:#e9644a}.cls-2{fill:#fff}</style></defs><title>social-yt</title><path class="cls-1yt" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M17.9 15L13 17.9v-5.8zm4.6 2.7v-5.4a2.5 2.5 0 0 0-2.6-2.6h-9.8a2.5 2.5 0 0 0-2.6 2.6v5.4a2.5 2.5 0 0 0 2.6 2.6h9.8a2.5 2.5 0 0 0 2.6-2.6"/></svg></a>\n\t<a href="http://www.montgomeryplanningboard.org/rss/rss.shtm" target="_blank">rss<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1rss{fill:#f29100}.cls-2{fill:#fff}</style></defs><title>social-rss</title><path class="cls-1rss" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M10.9 19.1a2 2 0 0 1 0 2.8 2 2 0 1 1-2.8-2.8 2 2 0 0 1 2.8 0zm-3.4-6.5v2.9a7 7 0 0 1 7 7h2.9a9.9 9.9 0 0 0-9.9-9.9zm0-5.1v2.9a12.2 12.2 0 0 1 12.1 12.1h2.9a15 15 0 0 0-15-15z"/></svg></a>\n</div>\t\t\t\t\t<a href="http://www.mncppc.org" title="Maryland-National Capital Parks and Commission" class="mncppc-link" target="_blank">Maryland-National Capital Park and Planning Commission</a>\r\n\t\t\t\t</div>\r\n\t\t\t</div>\r\n\t\t</div>\r\n\t\t<div id="mobile-footer">\r\n\t\t    <div class="footer-icon">\r\n\t\t    \t<a href="/" class="ico ico-home" title="Back to Home"><svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 31.9 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-home</title><path class="cls-1" d="M26.7 14.7a.9.9 0 0 0-.9.9v12.6h-6.5v-5.1a.9.9 0 0 0-.9-.9h-4.7a.9.9 0 0 0-.9.9v5.1H6.2V16a.9.9 0 0 0-1.8 0v13.1a.9.9 0 0 0 .9.9h8.4a.9.9 0 0 0 .9-.9V24h2.9v5.1a.9.9 0 0 0 .9.9h8.3a.9.9 0 0 0 .9-.9V15.6a.9.9 0 0 0-.9-.9zm0 0l5 .7-7.3-7.3v-5a.9.9 0 0 0-.9-.9.9.9 0 0 0-.9.9v3.2l-6-6A.9.9 0 0 0 16 0a.9.9 0 0 0-.7.3l-15 15a.9.9 0 0 0 0 1.3.9.9 0 0 0 1.3 0L16 2.2l14.4 14.5a.9.9 0 0 0 .7.2l.7-.2a1 1 0 0 0 0-1.3zm0 0"/></svg></a>\r\n\t\t    </div>\r\n\t\t    <div class="footer-icon">\r\n\t\t    <a href="/calendar" class="ico ico-calendar" title="Calendar"><svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-calendar</title><path class="cls-1" d="M27.5 2.8h-4.1V.9a.9.9 0 0 0-1.9 0v1.9h-5.6V.9a.9.9 0 0 0-1.9 0v1.9H8.4V.9a.9.9 0 1 0-1.8 0v1.9H2.5A2.5 2.5 0 0 0 0 5.3v22.2A2.5 2.5 0 0 0 2.5 30h25a2.5 2.5 0 0 0 2.5-2.5V5.3a2.5 2.5 0 0 0-2.5-2.5zm.6 24.7a.6.6 0 0 1-.6.6h-25a.6.6 0 0 1-.6-.6V5.3a.6.6 0 0 1 .6-.6h4.1v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h5.6v1.9a.9.9 0 0 0 1.9 0V4.7h4.1a.6.6 0 0 1 .6.6v22.2zm0 0"/><path class="cls-1" d="M6.6 11.3h3.7v2.81H6.6zm0 4.6h3.7v2.81H6.6zm0 4.7h3.7v2.81H6.6zm6.5 0h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7zm6.6 9.3h3.7v2.81h-3.7zm0-4.7h3.7v2.81h-3.7zm0-4.6h3.7v2.81h-3.7z"/></svg></a></div>\r\n\t\t    <div class="footer-icon"><a href="/news" class="ico ico-news" title="News"><svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30.7 30"><defs><style>.cls-1{fill:#79ba00}</style></defs><title>icon-news</title><path class="cls-1" d="M30 2.5a.8.8 0 0 0-.8.8v21.1a4 4 0 0 1-4 4H5.7a4 4 0 0 1-4-4V1.6h23.2v20.5a.8.8 0 1 0 1.6 0V.8a.8.8 0 0 0-.8-.8H.8a.8.8 0 0 0-.8.8v23.6A5.6 5.6 0 0 0 5.7 30h19.4a5.6 5.6 0 0 0 5.6-5.6V3.3a.8.8 0 0 0-.8-.8zm0 0"/><path class="cls-1" d="M21.9 6.2a.8.8 0 0 0 0-1.6H4.5a.8.8 0 1 0 0 1.6h17.4zm0 14.6H4.5a.8.8 0 1 0 0 1.6h17.4a.8.8 0 0 0 0-1.6zm0 4.1H7.3a.8.8 0 0 0 0 1.6h14.6a.8.8 0 0 0 0-1.6zM3.5 9v8.6a.8.8 0 0 0 .8.8h8.6a.8.8 0 0 0 .8-.8V9a.8.8 0 0 0-.8-.8H4.3a.8.8 0 0 0-.8.8zm1.6.8h7v7h-7v-7zM22.3 9h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 0-1.6zm0 3.6h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 1 0 0-1.6zm.8 4.6a.8.8 0 0 0-.8-.8h-5.7a.8.8 0 0 0 0 1.6h5.7a.8.8 0 0 0 .8-.8zm0 0"/></svg></a></div>\r\n\t\t    <div class="menu-open">\r\n\t\t    \t<div id="hamburger">\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t\t<span></span>\r\n\t\t\t\t</div>\r\n\t\t\t</div>\r\n\t\t</div>\r\n\t\t<div id="siteBody" class="site-body">\r\n<div class="calendar left-sb" data-date="2019-01-31" data-year="2019" data-month="01" data-view="month">\n    <div class="section-headline row">\n        <div class="p-lg-x-l">\n            <h1 class="section-head">Calendar of Events</h1>\n        </div>\n    </div>\n    <div class="sb-flex rel">\n        <div class="col col-sm-desk-12 col-lg-4 side-l col-sb">\n            <div class="sb-wrap">\n                <div class="sidebar">\n\n                    <div class="calendar-box col-sm-desk-6 col-lg-12">\n                        <div class="row controls rel">\n                            <div class="prev f-left">\n                                <a href="?cat=2901&v=0&date=2018-12-01"><span class="hv">Prev</span></a>\n                            </div>\n\n                            <div class="title f-center">\n                                <h2></h2>\n                            </div>\n\n                            <div class="next f-right">\n                                <a href="?cat=2901&v=0&date=2019-03-01"><span class="hv">Next</span></a>\n                            </div>\n                        </div>\n                        <div class="row days f-center m-y">\n                            <div class="col col-7th">Sun</div>\n                            <div class="col col-7th">Mon</div>\n                            <div class="col col-7th">Tue</div>\n                            <div class="col col-7th">Wed</div>\n                            <div class="col col-7th">Thu</div>\n                            <div class="col col-7th">Fri</div>\n                            <div class="col col-7th">Sat</div>\n                        </div>\n\n                        <div class="js-dates row">\n\n                        </div>\n                    </div>\n                    <div class="p-y">\n                        <h4>Search for An Event</h4>\n                        <div class="neat-search">\n                            <form action="" method="get" class="cal-search">\n                              <label for="search-2">Search Events</label>\n                              <input name="s" id="search-2" class="closed" type="text" placeholder="Search" value=""/>\n                              <button class="search-link" type="submit">Search Events</button>\n                            </form>\n                        </div>\n                    </div>\n                                            <div class="p-y-b inline-block">\n                            <h4>Search by Park</h4>\n                            <div class="neat-search">\n                                <form action="" method="get" class="cal-park-search">\n                                    <label for="search-5">Search Parks</label>\n                                    <input id="search-5" class="closed park-s" type="text" placeholder="Park Name" name="park" value=""/>\n                                    <button class="search-link" href="#search-5" type="submit">Search Parks</button>\n                                </form>\n                            </div>\n                        </div>\n                                            <div class="filter-list accordion col-sm-desk-6 col-lg-12">\n                        <div class="arrow"></div>\n                        <h4>Filter by category</h4>\n                        <a href="?cat=&v=0" id="clear" class="clear"><svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 47.5 45.1"><path d="M46.1 29.6c-.4-.1-.9.2-.9.6-.9 4.4-4.4 7.4-8.6 7.4-4 0-7.6-2.6-8.8-6.4h3.8v-1.6h-6.4V36h1.6v-3.1c1.7 3.7 5.5 6.3 9.8 6.3 5 0 9.1-3.5 10.2-8.6 0-.5-.3-.9-.7-1zM45.9 20v3.4c-1.8-3.6-5.4-5.8-9.4-5.8-4.9 0-9.2 3.2-10.6 7.8-.1.4.1.9.5 1 .4.1.9-.1 1-.5 1.1-3.8 4.9-6.6 9-6.6 3.7 0 6.8 2.2 8.3 5.6H41v1.6h6.4V20h-1.5zM30.8 10.5c0 1.7 1.4 3.1 3.1 3.1 1.7 0 3.2-1.4 3.2-3.1 0-1.7-1.4-3.1-3.2-3.1-1.7 0-3.1 1.4-3.1 3.1zm4.7 0c0 .9-.7 1.6-1.6 1.6-.9 0-1.5-.7-1.5-1.6 0-.9.7-1.5 1.5-1.5.9 0 1.6.7 1.6 1.5z"/><path d="M25.7 1.6h9.9l8 7.9v7.3h1.6V8.9l-9-8.9H25L4.8 20.2 0 25l20.1 20.2 6.2-5.9-19.2-19L25.7 1.6zM24 39.2l-3.9 3.7L2.3 25 6 21.3l18 17.9z"/></svg>clear</a>\n\n                        <ul class="filters accordion-wrap">\n                                                            <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1751&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Archaeology                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2230&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Butterfly Show                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2701&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Camp                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2235&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Clean Up                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1738&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Community Events                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1739&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Community Gardens                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2577&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Community Meeting                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=3105&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Earth Month                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2239&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Education                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=3331&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Event Center                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1740&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Events                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1741&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Family                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2328&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Featured                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2232&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Festival                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1746&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Free                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2243&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Gardens                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2242&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Golf                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1733&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Hikes                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1743&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Historical/Cultural                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2240&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                History                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=3215&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                History in the Parks                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2234&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                History Program                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1727&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Hockey                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2241&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Holiday                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1728&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Ice Skating                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1729&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Kids                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2705&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                MLK Day of Service                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1744&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Music                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=82&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Nature                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn selected"\n                                       href="?cat=2901&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                open house                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1753&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Openings/Dedications                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2233&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Other                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2902&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Park Planning                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2916&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Park Projects                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2751&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Planning Board Meeting                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2756&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Planning Board Worksession                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2569&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                pop up                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2754&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Public Hearing                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1750&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Public Meetings                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2231&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Run                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1730&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Special Events                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1731&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Sports                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1734&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Tennis                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2903&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Trail Planning                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=2706&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Trail Work                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1735&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Trails                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1736&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Trips                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1747&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Volunteering                                    </a>\n                                </li>\n                                                                <li>\n                                    <a class="btn filter-btn "\n                                       href="?cat=1748&v=0"\n                                       data-id=""\n                                       data-slug="">\n                                Weed Warrior                                    </a>\n                                </li>\n                                                        </ul>\n                    </div>\n                </div>\n            </div>\n        </div>\n        <div class="col col-sm-desk-12 col-lg-5 events">\n                <!--p>52 events meet your criteria of [filter item] & [filter item].</p-->\n            <section class="event-list" id="ajax-event-list">\n                <div class="p-y" style="text-align: right">\n                    <a href="#" class="toggle-view toggle-view-day">Switch to Day View</a>                                    </div>\n                <div class="event-item events-2019-01-31 view-m">\n                                        <ul>\n                                                <li>\n                                <a data-id="43700" href="/events/volunteer-fair-for-montgomery-parks-historic-sites/">\n                                    <span class="time">\n                                        Sat. March 23rd, 2019 9:00am to  3:00pm                                    </span>\n                                    <span class="event-name m-view">Join Parks Historic Sites Volunteer Program</span>\n                                </a>\n\n                                                            </li>\n                                                </ul>\n                    <div class="pag row">\n                        <div class="col col-6"></div>\n                        <div style="text-align: right;" class="col col-6 a-right"></div>\n                    </div>\n                                    </div>\n            </section>\n            <div class="loading"></div>\n        </div>\n        <div class="col col-sm-desk-12 col-lg-3 p-lg-y-t featured-events">\n                                <div class="popular-events p">\n                        <h3>Featured Events</h3>\n\n                                                    <div class="p-y single-pop-event">\n                                <a href="/events/we-speak-weddings-lodge-at-little-seneca-creek/">\n                                    We Speak Weddings &#8211; Lodge at Little Seneca Creek                                </a><br />\n                                Sun. February 3rd, 2019 10:00am to  1:00pm                            </div>\n                                                        <div class="p-y single-pop-event">\n                                <a href="/events/wedding-vendor-showcase-2/">\n                                    Wedding Vendor Showcase                                </a><br />\n                                Sun. March 3rd, 2019 11:00am to  2:00pm                            </div>\n                                                        <div class="p-y single-pop-event">\n                                <a href="/events/volunteer-fair-for-montgomery-parks-historic-sites/">\n                                    Join Parks Historic Sites Volunteer Program                                </a><br />\n                                Sat. March 23rd, 2019 9:00am to  3:00pm                            </div>\n                                                </div>\n                            </div>\n    </div>\n</div>\n<style>\n.m-no-time span.time{\n    display:none !important;\n}\n\n.m-no-time span.event-name{\n    padding-left: 0 !important;\n}\n</style>\n<div id="footer" class="main-footer">\r\n                        \r\n                    <section class="hot-topics row">\r\n                        <div class="cc-m">\r\n                            <h3>Topics</h3>\r\n                            <aside class="hot-topic">\r\n                                <div class="menu-hot-topics-container"><ul id="menu-hot-topics" class="menu"><li id="menu-item-14005" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-14005"><a href="/special-events/">Festivals &#038; Events</a></li>\n<li id="menu-item-14007" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-14007"><a href="/classes-camps/">Register Now for Classes &#038; Programs</a></li>\n<li id="menu-item-14008" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-14008"><a href="/support/volunteer/seasonal_volunteering/">Volunteer Events</a></li>\n<li id="menu-item-16517" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-16517"><a href="/calendar/">Calendar of Events</a></li>\n<li id="menu-item-14011" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-14011"><a href="/caring-for-our-parks/wildlife/ticks-lyme-disease/">Ticks and Lyme Disease</a></li>\n<li id="menu-item-16515" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-16515"><a href="/projects/public-input/">Open Town Hall</a></li>\n<li id="menu-item-18385" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-18385"><a href="/activities/history-in-the-parks/">History in the Parks</a></li>\n<li id="menu-item-33409" class="menu-item menu-item-type-custom menu-item-object-custom menu-item-33409"><a href="https://www.montgomeryparks.org/about/parks/pesticides/">Pesticide Reduction &#038; Integrated Pest Management</a></li>\n</ul></div>                            </aside>\r\n                                                    </div>\r\n                    </section>\r\n \r\n                    <section class="main-footer-bottom row">\r\n        <div class="col col-mob-12 col-desk-3">\r\n            <div class="address">\r\n                Montgomery County<br>Department of Parks<br>\r\n                9500 Brunett Ave.<br>\r\n                Silver Spring, MD 20901\r\n            </div>\r\n        </div>\r\n                        <div class="col col-mob-6 col-sm-desk-4 col-desk-2 footer-nav">\r\n                                            <a href="/about/contact-us/">Contact Us</a>\r\n                                                <a href="/about/media-center/press-contacts/">Press</a>\r\n                                                <a href="/about/parks/accessibility/">ADA Accessibility</a>\r\n                                                <a href="/terms/">Terms & Privacy</a>\r\n                                        </div>\r\n                                <div class="col col-mob-6 col-sm-desk-4 col-desk-2 footer-nav">\r\n                                            <a href="/about/park-police/">Park Police</a>\r\n                                                <a href="/about/rules-and-regulations/">Rules and Regulations</a>\r\n                                                <a href="/support/volunteer/">Volunteer</a>\r\n                                                <a href="/activities/">Activities</a>\r\n                                        </div>\r\n                                <div class="col col-mob-6 col-sm-desk-4 col-desk-2 footer-nav">\r\n                                            <a href="/permits/">Permits</a>\r\n                                                <a href="/calendar/">Calendar</a>\r\n                                                <a href="/news/">News</a>\r\n                                                <a href="/weather-alerts/">Weather Alerts</a>\r\n                                        </div>\r\n                        <div class="col col-sm-desk-3 pull-right mob-hide rel">\r\n            SOCIAL<br>\r\n            <div class="share">\n\t<a href="https://www.facebook.com/MontgomeryParks" target="_blank">facebook<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1fb{fill:#547bbc}.cls-2{fill:#fff}</style></defs><title>social-fb</title><path class="cls-1fb" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M11.4 12.5H13V11a4 4 0 0 1 .5-2.3 2.8 2.8 0 0 1 2.4-1.1 9.7 9.7 0 0 1 2.8.3l-.4 2.3-1.3-.3a.9.9 0 0 0-1.1.8v1.8h2.5l-.2 2.2h-2.4v7.8h-2.9v-7.8h-1.5v-2.2z"/></svg></a>\n\t<a href="https://twitter.com/MontgomeryParks" target="_blank">twitter<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1tw{fill:#5ea8dd}.cls-2{fill:#fff}</style></defs><title>social-tw</title><path class="cls-1tw" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M22.5 11.1l-1.8.5a3.1 3.1 0 0 0 1.4-1.7 6.2 6.2 0 0 1-2 .7 3.1 3.1 0 0 0-5.3 2.8 8.8 8.8 0 0 1-6.4-3.2 3.1 3.1 0 0 0 1 4.1 3.1 3.1 0 0 1-1.4-.4 3.1 3.1 0 0 0 2.5 3H9.2a3.1 3.1 0 0 0 2.9 2.1 6.2 6.2 0 0 1-3.8 1.3h-.8A8.8 8.8 0 0 0 21 13.1v-.4a6.2 6.2 0 0 0 1.5-1.6"/></svg></a>\n\t<a href="https://www.instagram.com/montgomeryparks/" target="_blank">instagram<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 291.319 291.319"><path d="M145.659 0c80.44 0 145.66 65.219 145.66 145.66S226.1 291.319 145.66 291.319 0 226.1 0 145.66 65.21 0 145.659 0z" fill="#9b6954"/><path d="M195.93 63.708H95.38c-17.47 0-31.672 14.211-31.672 31.672v100.56c0 17.47 14.211 31.672 31.672 31.672h100.56c17.47 0 31.672-14.211 31.672-31.672V95.38c-.001-17.461-14.212-31.672-31.682-31.672zm9.978 18.326l3.587-.009v27.202l-27.402.091-.091-27.202 23.906-.082zm-60.248 36.205c22.732 0 27.42 21.339 27.42 27.429 0 15.103-12.308 27.411-27.42 27.411-15.121 0-27.42-12.308-27.42-27.411-.01-6.09 4.688-27.429 27.42-27.429zm63.99 75.716c0 8.658-7.037 15.704-15.713 15.704H97.073c-8.667 0-15.713-7.037-15.713-15.704v-66.539h22.759c-2.112 5.198-3.305 12.299-3.305 18.253 0 24.708 20.101 44.818 44.818 44.818s44.808-20.11 44.808-44.818c0-5.954-1.193-13.055-3.296-18.253h22.486v66.539h.02z" fill="#fff"/></svg></a>\n\t<a href="https://www.youtube.com/user/MontgomeryParks" target="_blank">youtube<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1yt{fill:#e9644a}.cls-2{fill:#fff}</style></defs><title>social-yt</title><path class="cls-1yt" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M17.9 15L13 17.9v-5.8zm4.6 2.7v-5.4a2.5 2.5 0 0 0-2.6-2.6h-9.8a2.5 2.5 0 0 0-2.6 2.6v5.4a2.5 2.5 0 0 0 2.6 2.6h9.8a2.5 2.5 0 0 0 2.6-2.6"/></svg></a>\n\t<a href="http://www.montgomeryplanningboard.org/rss/rss.shtm" target="_blank">rss<svg data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30"><defs><style>.cls-1rss{fill:#f29100}.cls-2{fill:#fff}</style></defs><title>social-rss</title><path class="cls-1rss" d="M15 30A15 15 0 1 0 0 15a15 15 0 0 0 15 15"/><path class="cls-2" d="M10.9 19.1a2 2 0 0 1 0 2.8 2 2 0 1 1-2.8-2.8 2 2 0 0 1 2.8 0zm-3.4-6.5v2.9a7 7 0 0 1 7 7h2.9a9.9 9.9 0 0 0-9.9-9.9zm0-5.1v2.9a12.2 12.2 0 0 1 12.1 12.1h2.9a15 15 0 0 0-15-15z"/></svg></a>\n</div>        </div>\r\n    </section>\r\n</div>\r\n        <style>\n    .brink-gdpr-cookie-middle{\n      border-right:1px solid rgba(255,255,255,.5);\n    }\n\n    .brink-gdpr-cookie-left a{\n      color:#fff;\n    }\n\n    #brink-gdpr-cookie-accept{\n      background:rgb(0,0,0);\n      color:#fff;\n    }\n\n    #brink-gdpr-cookie-accept:hover, #brink-gdpr-cookie-accept:active{\n      background:rgb(255,255,255);\n      color:#000;\n    }\n\n    .brink-gdpr-cookie-right span:before{\n      border-top:1px solid #fff;\n      border-right:1px solid #fff;\n    }\n\n    .brink-gdpr-cookie-right span:after{\n      border-top:1px solid #fff;\n      border-left:1px solid #fff;\n    }\n    </style>\n    <div id="brink-gdpr-cookie-notice" class="brink-gdpr-cookie-notice bottom" style="background:rgba(102,102,102,.95); color:#fff">\n      <div class="brink-gdpr-container">\n        <div class="brink-gdpr-cookie-left">\n          <span class="h2">Cookies Notice</span>\n            <p>\n              This website uses cookies to provide you with an improved and personalized experience. By using this site you agree to our use of cookies. Please read our <a href="/terms/" target="_blank">cookies policy</a> for more information on the cookies we use and how to delete or block them.            </p>\n          </span>\n        </div>\n        <div class="brink-gdpr-cookie-middle">\n          <a href="#" id="brink-gdpr-cookie-accept">I Accept</a>\n        </div>\n        <div class="brink-gdpr-cookie-right">\n          <a href="#" id="brink-gdpr-cookie-close"></a>\n          <span></span>\n        </div>\n      </div>\n    </div>\n    <script type="text/javascript">\n      window.brink_gdpr_whitelist = [];\n    </script>\n    <script data-cfasync="false" type="text/javascript">if (window.addthis_product === undefined) { window.addthis_product = "wpp"; } if (window.wp_product_version === undefined) { window.wp_product_version = "wpp-6.1.8"; } if (window.wp_blog_version === undefined) { window.wp_blog_version = "5.0.3"; } if (window.addthis_share === undefined) { window.addthis_share = {}; } if (window.addthis_config === undefined) { window.addthis_config = {"data_track_clickback":true,"ui_atversion":"300"}; } if (window.addthis_plugin_info === undefined) { window.addthis_plugin_info = {"info_status":"enabled","cms_name":"WordPress","plugin_name":"Share Buttons by AddThis","plugin_version":"6.1.8","plugin_mode":"AddThis","anonymous_profile_id":"wp-f82eef1f055e09dcb6273ac5462e3333","page_info":{"template":"pages","post_type":""},"sharing_enabled_on_post_via_metabox":false}; } \n                    (function() {\n                      var first_load_interval_id = setInterval(function () {\n                        if (typeof window.addthis !== \'undefined\') {\n                          window.clearInterval(first_load_interval_id);\n                          if (typeof window.addthis_layers !== \'undefined\' && Object.getOwnPropertyNames(window.addthis_layers).length > 0) {\n                            window.addthis.layers(window.addthis_layers);\n                          }\n                          if (Array.isArray(window.addthis_layers_tools)) {\n                            for (i = 0; i < window.addthis_layers_tools.length; i++) {\n                              window.addthis.layers(window.addthis_layers_tools[i]);\n                            }\n                          }\n                        }\n                     },1000)\n                    }());\n                </script> <script data-cfasync="false" type="text/javascript" src="https://s7.addthis.com/js/300/addthis_widget.js#pubid=ra-5a0206a251dc7e59" async="async"></script><script type=\'text/javascript\'>\n/* <![CDATA[ */\nvar countVars = {"disqusShortname":"montgomeryparks"};\n/* ]]> */\n</script>\n<script type=\'text/javascript\' src=\'/wp-content/plugins/disqus-comment-system/public/js/comment_count.js?ver=3.0.16\'></script>\n<script type=\'text/javascript\' src=\'/wp-content/themes/mdparks/dist/js/main.js?ver=23\'></script>\n<script type=\'text/javascript\' src=\'https://js.arcgis.com/3.18/?ver=5.0.3\'></script>\n<script type=\'text/javascript\' src=\'/wp-includes/js/wp-embed.min.js?ver=5.0.3\'></script>\n<script>\r\n  window.addEventListener(\'load\',function(){\r\n    jQuery(\'[href*=".pdf"]\').click(function(){\r\n      __gaTracker(\'send\',\'event\',\'PDF\',\'click\',jQuery(this).attr(\'href\'));\r\n    })\r\n  });\r\n</script>\r\n<img height="1" width="1" style="border-style:none;" alt="" src="//googleads.g.doubleclick.net/pagead/viewthroughconversion/974628909/?guid=ON&amp;script=0"/>\r\n\r\n</body>\r\n\r\n</html>\r\n<!-- AWS -->\r\n'
